% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessFilter.r
\docType{data}
\name{AccessFilter}
\alias{AccessFilter}
\title{AccessFilter Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AccessFilter
}
\description{
AccessFilter Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{model}}{}

\item{\code{field}}{}

\item{\code{value}}{}

\item{\code{url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
