% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessToken.r
\docType{data}
\name{AccessToken}
\alias{AccessToken}
\title{AccessToken Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AccessToken
}
\description{
AccessToken Class
}
\section{Fields}{

\describe{
\item{\code{access_token}}{}

\item{\code{token_type}}{}

\item{\code{expires_in}}{}
}}

\keyword{datasets}
