% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi.R
\name{AddRelativeGazes}
\alias{AddRelativeGazes}
\title{Find gaze location relative to target image}
\usage{
AddRelativeGazes(x)
}
\arguments{
\item{x}{a Trial with \code{Task} or \code{Protocol} attributes or a
TrialList where each Trial has these attributes}
}
\value{
the Trial object(s) with the AOI data attached as new columns.
}
\description{
Creates new columns that measure gaze location relative to the location of
the target image, so that the x and y gaze positions increase the gaze gets
closer to the target. The new columns are \code{XLeftToTarget},
\code{XRightToTarget}, \code{XMeanToTarget}, \code{YLeftToTarget},
\code{YRightToTarget}, \code{YMeanToTarget}.
}
\details{
Normally, the x value (as in \code{trial$XMean}) increases as the gaze
  moves right on screen. When the target is left side of the screen, small x
  means the gaze approached the target, but when the target image is on the
  right, larger x means the gaze approached the target. This function flips
  the x value (sets x to 1-x) on trials where the target is on the left side
  of the screen, so that bigger x means the gaze approached the target
  regardless of which side of the screen the target was on. An analogous
  transformation is done for the y coordinate.
}

