% Generated by roxygen2 (4.0.1): do not edit by hand
\name{AdjustTimes}
\alias{AdjustTimes}
\title{Adjust trial times using a trial attribute}
\usage{
AdjustTimes(...)
}
\arguments{
\item{trial}{a Trial object}

\item{trials}{a list of Trial objects}

\item{alignment_event}{the name of the trial attribute to be used as time 0.
The default is \code{"TargetOnset"}.}
}
\value{
the Time column of each Trial is updated so that time 0 occurs at the
  time given by attribute. For example, if \code{attr(trial, "TargetOnset")}
  equals 1000 then the frame in the Time column that is closest to 1000ms
  gets the new value 0ms and all other frames and time-related attributes are
  updated relative to this value.
}
\description{
Adjust trial times using a trial attribute
}

