% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/time.R
\name{AdjustTimingMarks}
\alias{AdjustTimingMarks}
\title{Adjust the event-timing attributes of a trial}
\usage{
AdjustTimingMarks(trial, alignment_event)
}
\arguments{
\item{trial}{}

\item{alignment_event}{}
}
\value{
the inputted trial object with updated timing attributes
}
\description{
Adjust the event-timing attributes of a trial
}
\keyword{internal}

