% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alert.r
\name{Alert}
\alias{Alert}
\title{Alert Class}
\description{
Alert Class

Alert Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{applied_dashboard_filters}}{}

\item{\code{comparison_type}}{}

\item{\code{cron}}{}

\item{\code{custom_title}}{}

\item{\code{dashboard_element_id}}{}

\item{\code{description}}{}

\item{\code{destinations}}{}

\item{\code{field}}{}

\item{\code{followed}}{}

\item{\code{followable}}{}

\item{\code{id}}{}

\item{\code{is_disabled}}{}

\item{\code{disabled_reason}}{}

\item{\code{is_public}}{}

\item{\code{investigative_content_type}}{}

\item{\code{investigative_content_id}}{}

\item{\code{investigative_content_title}}{}

\item{\code{lookml_dashboard_id}}{}

\item{\code{lookml_link_id}}{}

\item{\code{owner_id}}{}

\item{\code{owner_display_name}}{}

\item{\code{threshold}}{}

\item{\code{time_series_condition_state}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Alert$new()}}
\item \href{#method-toJSON}{\code{Alert$toJSON()}}
\item \href{#method-fromJSONObject}{\code{Alert$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{Alert$fromJSON()}}
\item \href{#method-toJSONString}{\code{Alert$toJSONString()}}
\item \href{#method-fromJSONString}{\code{Alert$fromJSONString()}}
\item \href{#method-clone}{\code{Alert$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Alert$new(
  applied_dashboard_filters,
  comparison_type,
  cron,
  custom_title,
  dashboard_element_id,
  description,
  destinations,
  field,
  followed,
  followable,
  id,
  is_disabled,
  disabled_reason,
  is_public,
  investigative_content_type,
  investigative_content_id,
  investigative_content_title,
  lookml_dashboard_id,
  lookml_link_id,
  owner_id,
  owner_display_name,
  threshold,
  time_series_condition_state
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Alert$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Alert$fromJSONObject(AlertJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Alert$fromJSON(AlertJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Alert$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Alert$fromJSONString(AlertJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Alert$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
