% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlertApi.r
\name{AlertApi}
\alias{AlertApi}
\title{Alert operations}
\description{
looker.Alert
}
\section{Methods}{

\describe{

create_alert Create an alert


delete_alert Delete an alert


enqueue_alert Enqueue an alert


get_alert Get an alert


search_alerts Search Alerts


update_alert Update an alert


update_alert_field Update select fields on an alert

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AlertApi$new()}}
\item \href{#method-create_alert}{\code{AlertApi$create_alert()}}
\item \href{#method-delete_alert}{\code{AlertApi$delete_alert()}}
\item \href{#method-enqueue_alert}{\code{AlertApi$enqueue_alert()}}
\item \href{#method-get_alert}{\code{AlertApi$get_alert()}}
\item \href{#method-search_alerts}{\code{AlertApi$search_alerts()}}
\item \href{#method-update_alert}{\code{AlertApi$update_alert()}}
\item \href{#method-update_alert_field}{\code{AlertApi$update_alert_field()}}
\item \href{#method-clone}{\code{AlertApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlertApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_alert"></a>}}
\if{latex}{\out{\hypertarget{method-create_alert}{}}}
\subsection{Method \code{create_alert()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlertApi$create_alert(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_alert"></a>}}
\if{latex}{\out{\hypertarget{method-delete_alert}{}}}
\subsection{Method \code{delete_alert()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlertApi$delete_alert(alert_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-enqueue_alert"></a>}}
\if{latex}{\out{\hypertarget{method-enqueue_alert}{}}}
\subsection{Method \code{enqueue_alert()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlertApi$enqueue_alert(alert_id, force, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_alert"></a>}}
\if{latex}{\out{\hypertarget{method-get_alert}{}}}
\subsection{Method \code{get_alert()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlertApi$get_alert(alert_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_alerts"></a>}}
\if{latex}{\out{\hypertarget{method-search_alerts}{}}}
\subsection{Method \code{search_alerts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlertApi$search_alerts(
  limit,
  offset,
  group_by,
  fields,
  disabled,
  frequency,
  condition_met,
  last_run_start,
  last_run_end,
  all_owners,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_alert"></a>}}
\if{latex}{\out{\hypertarget{method-update_alert}{}}}
\subsection{Method \code{update_alert()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlertApi$update_alert(alert_id, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_alert_field"></a>}}
\if{latex}{\out{\hypertarget{method-update_alert_field}{}}}
\subsection{Method \code{update_alert_field()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlertApi$update_alert_field(alert_id, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlertApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
