% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlertAppliedDashboardFilter.r
\name{AlertAppliedDashboardFilter}
\alias{AlertAppliedDashboardFilter}
\title{AlertAppliedDashboardFilter Class}
\description{
AlertAppliedDashboardFilter Class

AlertAppliedDashboardFilter Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{filter_title}}{}

\item{\code{field_name}}{}

\item{\code{filter_value}}{}

\item{\code{filter_description}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AlertAppliedDashboardFilter$new()}}
\item \href{#method-toJSON}{\code{AlertAppliedDashboardFilter$toJSON()}}
\item \href{#method-fromJSONObject}{\code{AlertAppliedDashboardFilter$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{AlertAppliedDashboardFilter$fromJSON()}}
\item \href{#method-toJSONString}{\code{AlertAppliedDashboardFilter$toJSONString()}}
\item \href{#method-fromJSONString}{\code{AlertAppliedDashboardFilter$fromJSONString()}}
\item \href{#method-clone}{\code{AlertAppliedDashboardFilter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlertAppliedDashboardFilter$new(
  filter_title,
  field_name,
  filter_value,
  filter_description
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlertAppliedDashboardFilter$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlertAppliedDashboardFilter$fromJSONObject(
  AlertAppliedDashboardFilterJsonObject
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlertAppliedDashboardFilter$fromJSON(AlertAppliedDashboardFilterJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlertAppliedDashboardFilter$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlertAppliedDashboardFilter$fromJSONString(AlertAppliedDashboardFilterJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlertAppliedDashboardFilter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
