% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApiAuthApi.r
\docType{data}
\name{ApiAuthApi}
\alias{ApiAuthApi}
\title{ApiAuth operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApiAuthApi
}
\description{
looker.ApiAuth
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

login Login 


login_user Login user 


logout Logout 

}
}

\keyword{datasets}
