% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApiSession.r
\docType{data}
\name{ApiSession}
\alias{ApiSession}
\title{ApiSession Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApiSession
}
\description{
ApiSession Class
}
\section{Fields}{

\describe{
\item{\code{workspace_id}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
