% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApiSettings.r
\name{ApiSettings}
\alias{ApiSettings}
\title{ApiSettings}
\description{
Keeps track of the local settings for accessing the API (credentials, base path, etc).
This shouldn't generally be called directly.
A new \code{ApiSettings} will be generated when you created a new \code{LookerSDK}.
}
