% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApiVersion.r
\docType{data}
\name{ApiVersion}
\alias{ApiVersion}
\title{ApiVersion Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApiVersion
}
\description{
ApiVersion Class
}
\section{Fields}{

\describe{
\item{\code{looker_release_version}}{}

\item{\code{current_version}}{}

\item{\code{supported_versions}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
