% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApiVersionElement.r
\docType{data}
\name{ApiVersionElement}
\alias{ApiVersionElement}
\title{ApiVersionElement Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApiVersionElement
}
\description{
ApiVersionElement Class
}
\section{Fields}{

\describe{
\item{\code{version}}{}

\item{\code{full_version}}{}

\item{\code{status}}{}

\item{\code{swagger_url}}{}
}}

\keyword{datasets}
