% Generated by roxygen2 (4.0.1): do not edit by hand
\name{AssignNewTimes}
\alias{AssignNewTimes}
\title{Assign new times to a trial once it has been aligned at a reference frame}
\usage{
AssignNewTimes(trial, zero_frame)
}
\arguments{
\item{trial}{a Trial object}

\item{the}{index of the reference frame that will mark time = 0 in the
aligned trial}
}
\value{
the inputted trial object with updated values in its \code{Time}
  column
}
\description{
Assign new times to a trial once it has been aligned at a reference frame
}

