% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthApi.r
\name{AuthApi}
\alias{AuthApi}
\title{Auth operations}
\description{
looker.Auth
}
\section{Methods}{

\describe{

create_oidc_test_config Create OIDC Test Configuration 


create_saml_test_config Create SAML Test Configuration 


delete_oidc_test_config Delete OIDC Test Configuration 


delete_saml_test_config Delete SAML Test Configuration 


fetch_and_parse_saml_idp_metadata Parse SAML IdP Url 


ldap_config Get LDAP Configuration 


oidc_config Get OIDC Configuration 


oidc_test_config Get OIDC Test Configuration 


parse_saml_idp_metadata Parse SAML IdP XML 


saml_config Get SAML Configuration 


saml_test_config Get SAML Test Configuration 


test_ldap_config_auth Test LDAP Auth 


test_ldap_config_connection Test LDAP Connection 


test_ldap_config_user_auth Test LDAP User Auth 


test_ldap_config_user_info Test LDAP User Info 


update_ldap_config Update LDAP Configuration 


update_oidc_config Update OIDC Configuration 


update_saml_config Update SAML Configuration 

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AuthApi$new()}}
\item \href{#method-create_oidc_test_config}{\code{AuthApi$create_oidc_test_config()}}
\item \href{#method-create_saml_test_config}{\code{AuthApi$create_saml_test_config()}}
\item \href{#method-delete_oidc_test_config}{\code{AuthApi$delete_oidc_test_config()}}
\item \href{#method-delete_saml_test_config}{\code{AuthApi$delete_saml_test_config()}}
\item \href{#method-fetch_and_parse_saml_idp_metadata}{\code{AuthApi$fetch_and_parse_saml_idp_metadata()}}
\item \href{#method-ldap_config}{\code{AuthApi$ldap_config()}}
\item \href{#method-oidc_config}{\code{AuthApi$oidc_config()}}
\item \href{#method-oidc_test_config}{\code{AuthApi$oidc_test_config()}}
\item \href{#method-parse_saml_idp_metadata}{\code{AuthApi$parse_saml_idp_metadata()}}
\item \href{#method-saml_config}{\code{AuthApi$saml_config()}}
\item \href{#method-saml_test_config}{\code{AuthApi$saml_test_config()}}
\item \href{#method-test_ldap_config_auth}{\code{AuthApi$test_ldap_config_auth()}}
\item \href{#method-test_ldap_config_connection}{\code{AuthApi$test_ldap_config_connection()}}
\item \href{#method-test_ldap_config_user_auth}{\code{AuthApi$test_ldap_config_user_auth()}}
\item \href{#method-test_ldap_config_user_info}{\code{AuthApi$test_ldap_config_user_info()}}
\item \href{#method-update_ldap_config}{\code{AuthApi$update_ldap_config()}}
\item \href{#method-update_oidc_config}{\code{AuthApi$update_oidc_config()}}
\item \href{#method-update_saml_config}{\code{AuthApi$update_saml_config()}}
\item \href{#method-clone}{\code{AuthApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_oidc_test_config"></a>}}
\if{latex}{\out{\hypertarget{method-create_oidc_test_config}{}}}
\subsection{Method \code{create_oidc_test_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$create_oidc_test_config(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_saml_test_config"></a>}}
\if{latex}{\out{\hypertarget{method-create_saml_test_config}{}}}
\subsection{Method \code{create_saml_test_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$create_saml_test_config(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_oidc_test_config"></a>}}
\if{latex}{\out{\hypertarget{method-delete_oidc_test_config}{}}}
\subsection{Method \code{delete_oidc_test_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$delete_oidc_test_config(test_slug, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_saml_test_config"></a>}}
\if{latex}{\out{\hypertarget{method-delete_saml_test_config}{}}}
\subsection{Method \code{delete_saml_test_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$delete_saml_test_config(test_slug, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fetch_and_parse_saml_idp_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-fetch_and_parse_saml_idp_metadata}{}}}
\subsection{Method \code{fetch_and_parse_saml_idp_metadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$fetch_and_parse_saml_idp_metadata(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ldap_config"></a>}}
\if{latex}{\out{\hypertarget{method-ldap_config}{}}}
\subsection{Method \code{ldap_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$ldap_config(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-oidc_config"></a>}}
\if{latex}{\out{\hypertarget{method-oidc_config}{}}}
\subsection{Method \code{oidc_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$oidc_config(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-oidc_test_config"></a>}}
\if{latex}{\out{\hypertarget{method-oidc_test_config}{}}}
\subsection{Method \code{oidc_test_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$oidc_test_config(test_slug, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-parse_saml_idp_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-parse_saml_idp_metadata}{}}}
\subsection{Method \code{parse_saml_idp_metadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$parse_saml_idp_metadata(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saml_config"></a>}}
\if{latex}{\out{\hypertarget{method-saml_config}{}}}
\subsection{Method \code{saml_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$saml_config(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saml_test_config"></a>}}
\if{latex}{\out{\hypertarget{method-saml_test_config}{}}}
\subsection{Method \code{saml_test_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$saml_test_config(test_slug, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-test_ldap_config_auth"></a>}}
\if{latex}{\out{\hypertarget{method-test_ldap_config_auth}{}}}
\subsection{Method \code{test_ldap_config_auth()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$test_ldap_config_auth(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-test_ldap_config_connection"></a>}}
\if{latex}{\out{\hypertarget{method-test_ldap_config_connection}{}}}
\subsection{Method \code{test_ldap_config_connection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$test_ldap_config_connection(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-test_ldap_config_user_auth"></a>}}
\if{latex}{\out{\hypertarget{method-test_ldap_config_user_auth}{}}}
\subsection{Method \code{test_ldap_config_user_auth()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$test_ldap_config_user_auth(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-test_ldap_config_user_info"></a>}}
\if{latex}{\out{\hypertarget{method-test_ldap_config_user_info}{}}}
\subsection{Method \code{test_ldap_config_user_info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$test_ldap_config_user_info(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_ldap_config"></a>}}
\if{latex}{\out{\hypertarget{method-update_ldap_config}{}}}
\subsection{Method \code{update_ldap_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$update_ldap_config(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_oidc_config"></a>}}
\if{latex}{\out{\hypertarget{method-update_oidc_config}{}}}
\subsection{Method \code{update_oidc_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$update_oidc_config(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_saml_config"></a>}}
\if{latex}{\out{\hypertarget{method-update_saml_config}{}}}
\subsection{Method \code{update_saml_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$update_saml_config(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
