% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoardApi.r
\name{BoardApi}
\alias{BoardApi}
\title{Board operations}
\description{
looker.Board
}
\section{Methods}{

\describe{

all_board_items Get All Board Items


all_board_sections Get All Board sections


all_boards Get All Boards


board Get Board


board_item Get Board Item


board_section Get Board section


create_board Create Board


create_board_item Create Board Item


create_board_section Create Board section


delete_board Delete Board


delete_board_item Delete Board Item


delete_board_section Delete Board section


search_boards Search Boards


update_board Update Board


update_board_item Update Board Item


update_board_section Update Board section

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BoardApi$new()}}
\item \href{#method-all_board_items}{\code{BoardApi$all_board_items()}}
\item \href{#method-all_board_sections}{\code{BoardApi$all_board_sections()}}
\item \href{#method-all_boards}{\code{BoardApi$all_boards()}}
\item \href{#method-board}{\code{BoardApi$board()}}
\item \href{#method-board_item}{\code{BoardApi$board_item()}}
\item \href{#method-board_section}{\code{BoardApi$board_section()}}
\item \href{#method-create_board}{\code{BoardApi$create_board()}}
\item \href{#method-create_board_item}{\code{BoardApi$create_board_item()}}
\item \href{#method-create_board_section}{\code{BoardApi$create_board_section()}}
\item \href{#method-delete_board}{\code{BoardApi$delete_board()}}
\item \href{#method-delete_board_item}{\code{BoardApi$delete_board_item()}}
\item \href{#method-delete_board_section}{\code{BoardApi$delete_board_section()}}
\item \href{#method-search_boards}{\code{BoardApi$search_boards()}}
\item \href{#method-update_board}{\code{BoardApi$update_board()}}
\item \href{#method-update_board_item}{\code{BoardApi$update_board_item()}}
\item \href{#method-update_board_section}{\code{BoardApi$update_board_section()}}
\item \href{#method-clone}{\code{BoardApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoardApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_board_items"></a>}}
\if{latex}{\out{\hypertarget{method-all_board_items}{}}}
\subsection{Method \code{all_board_items()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoardApi$all_board_items(fields, sorts, board_section_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_board_sections"></a>}}
\if{latex}{\out{\hypertarget{method-all_board_sections}{}}}
\subsection{Method \code{all_board_sections()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoardApi$all_board_sections(fields, sorts, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_boards"></a>}}
\if{latex}{\out{\hypertarget{method-all_boards}{}}}
\subsection{Method \code{all_boards()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoardApi$all_boards(fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-board"></a>}}
\if{latex}{\out{\hypertarget{method-board}{}}}
\subsection{Method \code{board()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoardApi$board(board_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-board_item"></a>}}
\if{latex}{\out{\hypertarget{method-board_item}{}}}
\subsection{Method \code{board_item()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoardApi$board_item(board_item_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-board_section"></a>}}
\if{latex}{\out{\hypertarget{method-board_section}{}}}
\subsection{Method \code{board_section()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoardApi$board_section(board_section_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_board"></a>}}
\if{latex}{\out{\hypertarget{method-create_board}{}}}
\subsection{Method \code{create_board()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoardApi$create_board(body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_board_item"></a>}}
\if{latex}{\out{\hypertarget{method-create_board_item}{}}}
\subsection{Method \code{create_board_item()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoardApi$create_board_item(body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_board_section"></a>}}
\if{latex}{\out{\hypertarget{method-create_board_section}{}}}
\subsection{Method \code{create_board_section()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoardApi$create_board_section(body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_board"></a>}}
\if{latex}{\out{\hypertarget{method-delete_board}{}}}
\subsection{Method \code{delete_board()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoardApi$delete_board(board_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_board_item"></a>}}
\if{latex}{\out{\hypertarget{method-delete_board_item}{}}}
\subsection{Method \code{delete_board_item()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoardApi$delete_board_item(board_item_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_board_section"></a>}}
\if{latex}{\out{\hypertarget{method-delete_board_section}{}}}
\subsection{Method \code{delete_board_section()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoardApi$delete_board_section(board_section_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_boards"></a>}}
\if{latex}{\out{\hypertarget{method-search_boards}{}}}
\subsection{Method \code{search_boards()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoardApi$search_boards(
  title,
  created_at,
  first_name,
  last_name,
  fields,
  favorited,
  creator_id,
  sorts,
  page,
  per_page,
  offset,
  limit,
  filter_or,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_board"></a>}}
\if{latex}{\out{\hypertarget{method-update_board}{}}}
\subsection{Method \code{update_board()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoardApi$update_board(board_id, body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_board_item"></a>}}
\if{latex}{\out{\hypertarget{method-update_board_item}{}}}
\subsection{Method \code{update_board_item()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoardApi$update_board_item(board_item_id, body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_board_section"></a>}}
\if{latex}{\out{\hypertarget{method-update_board_section}{}}}
\subsection{Method \code{update_board_section()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoardApi$update_board_section(board_section_id, body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoardApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
