% Generated by roxygen2 (4.0.1): do not edit by hand
\name{CalculateMistrackings}
\alias{CalculateMistrackings}
\title{Calculate the amount of missing data in a list of Trials}
\usage{
CalculateMistrackings(...)
}
\arguments{
\item{trial}{a Trial object with a \code{GazeByImageAOI} column}

\item{trials}{a list of Trial objects, each with a \code{GazeByImageAOI}
column}

\item{column}{A character string that names one of the columns of Gazedata in
the trial object. Default is `"GazeByImageAOI"`.}
}
\value{
A list of Trial objects that has been updated to include attributes
  containing percentage of mistracked data.
}
\description{
Calculate the amount of missing data in a list of Trials
}

