% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorCollectionApi.r
\name{ColorCollectionApi}
\alias{ColorCollectionApi}
\title{ColorCollection operations}
\description{
looker.ColorCollection
}
\section{Methods}{

\describe{

all_color_collections Get all Color Collections


color_collection Get Color Collection by ID


color_collections_custom Get all Custom Color Collections


color_collections_standard Get all Standard Color Collections


create_color_collection Create ColorCollection


default_color_collection Get Default Color Collection


delete_color_collection Delete ColorCollection


set_default_color_collection Set Default Color Collection


update_color_collection Update Custom Color collection

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ColorCollectionApi$new()}}
\item \href{#method-all_color_collections}{\code{ColorCollectionApi$all_color_collections()}}
\item \href{#method-color_collection}{\code{ColorCollectionApi$color_collection()}}
\item \href{#method-color_collections_custom}{\code{ColorCollectionApi$color_collections_custom()}}
\item \href{#method-color_collections_standard}{\code{ColorCollectionApi$color_collections_standard()}}
\item \href{#method-create_color_collection}{\code{ColorCollectionApi$create_color_collection()}}
\item \href{#method-default_color_collection}{\code{ColorCollectionApi$default_color_collection()}}
\item \href{#method-delete_color_collection}{\code{ColorCollectionApi$delete_color_collection()}}
\item \href{#method-set_default_color_collection}{\code{ColorCollectionApi$set_default_color_collection()}}
\item \href{#method-update_color_collection}{\code{ColorCollectionApi$update_color_collection()}}
\item \href{#method-clone}{\code{ColorCollectionApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorCollectionApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_color_collections"></a>}}
\if{latex}{\out{\hypertarget{method-all_color_collections}{}}}
\subsection{Method \code{all_color_collections()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorCollectionApi$all_color_collections(fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-color_collection"></a>}}
\if{latex}{\out{\hypertarget{method-color_collection}{}}}
\subsection{Method \code{color_collection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorCollectionApi$color_collection(collection_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-color_collections_custom"></a>}}
\if{latex}{\out{\hypertarget{method-color_collections_custom}{}}}
\subsection{Method \code{color_collections_custom()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorCollectionApi$color_collections_custom(fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-color_collections_standard"></a>}}
\if{latex}{\out{\hypertarget{method-color_collections_standard}{}}}
\subsection{Method \code{color_collections_standard()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorCollectionApi$color_collections_standard(fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_color_collection"></a>}}
\if{latex}{\out{\hypertarget{method-create_color_collection}{}}}
\subsection{Method \code{create_color_collection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorCollectionApi$create_color_collection(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-default_color_collection"></a>}}
\if{latex}{\out{\hypertarget{method-default_color_collection}{}}}
\subsection{Method \code{default_color_collection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorCollectionApi$default_color_collection(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_color_collection"></a>}}
\if{latex}{\out{\hypertarget{method-delete_color_collection}{}}}
\subsection{Method \code{delete_color_collection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorCollectionApi$delete_color_collection(collection_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_default_color_collection"></a>}}
\if{latex}{\out{\hypertarget{method-set_default_color_collection}{}}}
\subsection{Method \code{set_default_color_collection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorCollectionApi$set_default_color_collection(collection_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_color_collection"></a>}}
\if{latex}{\out{\hypertarget{method-update_color_collection}{}}}
\subsection{Method \code{update_color_collection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorCollectionApi$update_color_collection(collection_id, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorCollectionApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
