% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{CombineGazedataStimdata}
\alias{CombineGazedataStimdata}
\title{Combine Gazedata and Stimdata}
\usage{
CombineGazedataStimdata(gazedata, stimdata)
}
\arguments{
\item{gazedata}{A Gazedata object.}

\item{stimdata}{A Stimdata object.}
}
\value{
A data.frame of the gazedata, augmented with attributes for the
  stimdata.
}
\description{
CombineGazedataStimdata is a utility function for combining the gazedata and
stimdata of a single trial from a Gazedata and Stimdata object. This function
is curried.
}
\details{
The function returns a higher order function with the
  \code{gazedata} and \code{stimdata} fixed. Details about the returned
  function:
\describe{
  \item{Usage}{
  \code{CombineGazedataStimdata(gazedata, stimdata)(trial_number)}}
  \item{\code{trial_number}}{The trial number of the trial whose gazedata
  and stimdata will be combined.}
}
}
\keyword{internal}
