% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommandApi.r
\name{CommandApi}
\alias{CommandApi}
\title{Command operations}
\description{
looker.Command
}
\section{Methods}{

\describe{

create_command Create a custom command


delete_command Delete a custom command


get_all_commands Get All Commands


update_command Update a custom command

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CommandApi$new()}}
\item \href{#method-create_command}{\code{CommandApi$create_command()}}
\item \href{#method-delete_command}{\code{CommandApi$delete_command()}}
\item \href{#method-get_all_commands}{\code{CommandApi$get_all_commands()}}
\item \href{#method-update_command}{\code{CommandApi$update_command()}}
\item \href{#method-clone}{\code{CommandApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommandApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_command"></a>}}
\if{latex}{\out{\hypertarget{method-create_command}{}}}
\subsection{Method \code{create_command()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommandApi$create_command(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_command"></a>}}
\if{latex}{\out{\hypertarget{method-delete_command}{}}}
\subsection{Method \code{delete_command()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommandApi$delete_command(command_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_all_commands"></a>}}
\if{latex}{\out{\hypertarget{method-get_all_commands}{}}}
\subsection{Method \code{get_all_commands()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommandApi$get_all_commands(content_id, content_type, limit, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_command"></a>}}
\if{latex}{\out{\hypertarget{method-update_command}{}}}
\subsection{Method \code{update_command()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommandApi$update_command(command_id, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommandApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
