% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfigApi.r
\docType{data}
\name{ConfigApi}
\alias{ConfigApi}
\title{Config operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ConfigApi
}
\description{
looker.Config
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

all_legacy_features Get All Legacy Features 


all_timezones Get All Timezones 


backup_configuration Get Backup Configuration 


legacy_feature Get Legacy Feature 


update_backup_configuration Update Backup Configuration 


update_legacy_feature Update Legacy Feature 


update_whitelabel_configuration Update Whitelabel configuration 


versions Get ApiVersion 


whitelabel_configuration Get Whitelabel configuration 

}
}

\keyword{datasets}
