% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectionApi.r
\docType{data}
\name{ConnectionApi}
\alias{ConnectionApi}
\title{Connection operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ConnectionApi
}
\description{
looker.Connection
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

all_connections Get All Connections 


all_dialect_infos Get All Dialect Infos 


connection Get Connection 


create_connection Create Connection 


delete_connection Delete Connection 


test_connection Test Connection 


test_connection_config Test Connection Configuration 


update_connection Update Connection 

}
}

\keyword{datasets}
