% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectionFeatures.r
\name{ConnectionFeatures}
\alias{ConnectionFeatures}
\title{ConnectionFeatures Class}
\description{
ConnectionFeatures Class

ConnectionFeatures Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dialect_name}}{}

\item{\code{cost_estimate}}{}

\item{\code{multiple_databases}}{}

\item{\code{column_search}}{}

\item{\code{persistent_table_indexes}}{}

\item{\code{persistent_derived_tables}}{}

\item{\code{turtles}}{}

\item{\code{percentile}}{}

\item{\code{distinct_percentile}}{}

\item{\code{stable_views}}{}

\item{\code{milliseconds}}{}

\item{\code{microseconds}}{}

\item{\code{subtotals}}{}

\item{\code{location}}{}

\item{\code{timezone}}{}

\item{\code{connection_pooling}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ConnectionFeatures$new()}}
\item \href{#method-toJSON}{\code{ConnectionFeatures$toJSON()}}
\item \href{#method-fromJSONObject}{\code{ConnectionFeatures$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{ConnectionFeatures$fromJSON()}}
\item \href{#method-toJSONString}{\code{ConnectionFeatures$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ConnectionFeatures$fromJSONString()}}
\item \href{#method-clone}{\code{ConnectionFeatures$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionFeatures$new(
  dialect_name,
  cost_estimate,
  multiple_databases,
  column_search,
  persistent_table_indexes,
  persistent_derived_tables,
  turtles,
  percentile,
  distinct_percentile,
  stable_views,
  milliseconds,
  microseconds,
  subtotals,
  location,
  timezone,
  connection_pooling
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionFeatures$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionFeatures$fromJSONObject(ConnectionFeaturesJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionFeatures$fromJSON(ConnectionFeaturesJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionFeatures$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionFeatures$fromJSONString(ConnectionFeaturesJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionFeatures$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
