% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContentApi.r
\name{ContentApi}
\alias{ContentApi}
\title{Content operations}
\description{
looker.Content
}
\section{Methods}{

\describe{

all_content_metadata_accesss Get All Content Metadata Accesss 


all_content_metadatas Get All Content Metadatas 


content_favorite Get Favorite Content 


content_metadata Get Content Metadata 


create_content_favorite Create Favorite Content 


create_content_metadata_access Create Content Metadata Access 


delete_content_favorite Delete Favorite Content 


delete_content_metadata_access Delete Content Metadata Access 


search_content_favorites Search Favorite Contents 


search_content_views Search Content Views 


update_content_metadata Update Content Metadata 


update_content_metadata_access Update Content Metadata Access 

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ContentApi$new()}}
\item \href{#method-all_content_metadata_accesss}{\code{ContentApi$all_content_metadata_accesss()}}
\item \href{#method-all_content_metadatas}{\code{ContentApi$all_content_metadatas()}}
\item \href{#method-content_favorite}{\code{ContentApi$content_favorite()}}
\item \href{#method-content_metadata}{\code{ContentApi$content_metadata()}}
\item \href{#method-create_content_favorite}{\code{ContentApi$create_content_favorite()}}
\item \href{#method-create_content_metadata_access}{\code{ContentApi$create_content_metadata_access()}}
\item \href{#method-delete_content_favorite}{\code{ContentApi$delete_content_favorite()}}
\item \href{#method-delete_content_metadata_access}{\code{ContentApi$delete_content_metadata_access()}}
\item \href{#method-search_content_favorites}{\code{ContentApi$search_content_favorites()}}
\item \href{#method-search_content_views}{\code{ContentApi$search_content_views()}}
\item \href{#method-update_content_metadata}{\code{ContentApi$update_content_metadata()}}
\item \href{#method-update_content_metadata_access}{\code{ContentApi$update_content_metadata_access()}}
\item \href{#method-clone}{\code{ContentApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_content_metadata_accesss"></a>}}
\if{latex}{\out{\hypertarget{method-all_content_metadata_accesss}{}}}
\subsection{Method \code{all_content_metadata_accesss()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentApi$all_content_metadata_accesss(content_metadata_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_content_metadatas"></a>}}
\if{latex}{\out{\hypertarget{method-all_content_metadatas}{}}}
\subsection{Method \code{all_content_metadatas()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentApi$all_content_metadatas(parent_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-content_favorite"></a>}}
\if{latex}{\out{\hypertarget{method-content_favorite}{}}}
\subsection{Method \code{content_favorite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentApi$content_favorite(content_favorite_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-content_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-content_metadata}{}}}
\subsection{Method \code{content_metadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentApi$content_metadata(content_metadata_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_content_favorite"></a>}}
\if{latex}{\out{\hypertarget{method-create_content_favorite}{}}}
\subsection{Method \code{create_content_favorite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentApi$create_content_favorite(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_content_metadata_access"></a>}}
\if{latex}{\out{\hypertarget{method-create_content_metadata_access}{}}}
\subsection{Method \code{create_content_metadata_access()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentApi$create_content_metadata_access(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_content_favorite"></a>}}
\if{latex}{\out{\hypertarget{method-delete_content_favorite}{}}}
\subsection{Method \code{delete_content_favorite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentApi$delete_content_favorite(content_favorite_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_content_metadata_access"></a>}}
\if{latex}{\out{\hypertarget{method-delete_content_metadata_access}{}}}
\subsection{Method \code{delete_content_metadata_access()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentApi$delete_content_metadata_access(content_metadata_access_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_content_favorites"></a>}}
\if{latex}{\out{\hypertarget{method-search_content_favorites}{}}}
\subsection{Method \code{search_content_favorites()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentApi$search_content_favorites(user_id, limit, offset, sorts, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_content_views"></a>}}
\if{latex}{\out{\hypertarget{method-search_content_views}{}}}
\subsection{Method \code{search_content_views()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentApi$search_content_views(
  view_count,
  group_id,
  look_id,
  dashboard_id,
  content_metadata_id,
  start_of_week_date,
  all_time,
  user_id,
  limit,
  offset,
  sorts,
  fields,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_content_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-update_content_metadata}{}}}
\subsection{Method \code{update_content_metadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentApi$update_content_metadata(content_metadata_id, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_content_metadata_access"></a>}}
\if{latex}{\out{\hypertarget{method-update_content_metadata_access}{}}}
\subsection{Method \code{update_content_metadata_access()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentApi$update_content_metadata_access(
  content_metadata_access_id,
  body,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
