% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContentFavorite.r
\docType{data}
\name{ContentFavorite}
\alias{ContentFavorite}
\title{ContentFavorite Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ContentFavorite
}
\description{
ContentFavorite Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{user_id}}{}

\item{\code{content_metadata_id}}{}

\item{\code{look_id}}{}

\item{\code{dashboard_id}}{}

\item{\code{look}}{}

\item{\code{dashboard}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
