% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContentMeta.r
\docType{data}
\name{ContentMeta}
\alias{ContentMeta}
\title{ContentMeta Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ContentMeta
}
\description{
ContentMeta Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{name}}{}

\item{\code{parent_id}}{}

\item{\code{space_id}}{}

\item{\code{dashboard_id}}{}

\item{\code{look_id}}{}

\item{\code{content_type}}{}

\item{\code{inherits}}{}

\item{\code{inheriting_id}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
