% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContentMetaGroupUser.r
\docType{data}
\name{ContentMetaGroupUser}
\alias{ContentMetaGroupUser}
\title{ContentMetaGroupUser Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ContentMetaGroupUser
}
\description{
ContentMetaGroupUser Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{content_metadata_id}}{}

\item{\code{permission_type}}{}

\item{\code{group_id}}{}

\item{\code{user_id}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
