% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContentValidation.r
\name{ContentValidation}
\alias{ContentValidation}
\title{ContentValidation Class}
\description{
ContentValidation Class

ContentValidation Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{content_with_errors}}{}

\item{\code{computation_time}}{}

\item{\code{total_looks_validated}}{}

\item{\code{total_dashboard_elements_validated}}{}

\item{\code{total_dashboard_filters_validated}}{}

\item{\code{total_scheduled_plans_validated}}{}

\item{\code{total_alerts_validated}}{}

\item{\code{total_explores_validated}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ContentValidation$new()}}
\item \href{#method-toJSON}{\code{ContentValidation$toJSON()}}
\item \href{#method-fromJSONObject}{\code{ContentValidation$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{ContentValidation$fromJSON()}}
\item \href{#method-toJSONString}{\code{ContentValidation$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ContentValidation$fromJSONString()}}
\item \href{#method-clone}{\code{ContentValidation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentValidation$new(
  content_with_errors,
  computation_time,
  total_looks_validated,
  total_dashboard_elements_validated,
  total_dashboard_filters_validated,
  total_scheduled_plans_validated,
  total_alerts_validated,
  total_explores_validated
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentValidation$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentValidation$fromJSONObject(ContentValidationJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentValidation$fromJSON(ContentValidationJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentValidation$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentValidation$fromJSONString(ContentValidationJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentValidation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
