% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContentValidationDashboardElement.r
\name{ContentValidationDashboardElement}
\alias{ContentValidationDashboardElement}
\title{ContentValidationDashboardElement Class}
\description{
ContentValidationDashboardElement Class

ContentValidationDashboardElement Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{body_text}}{}

\item{\code{dashboard_id}}{}

\item{\code{id}}{}

\item{\code{look_id}}{}

\item{\code{note_display}}{}

\item{\code{note_state}}{}

\item{\code{note_text}}{}

\item{\code{note_text_as_html}}{}

\item{\code{query_id}}{}

\item{\code{subtitle_text}}{}

\item{\code{title}}{}

\item{\code{title_hidden}}{}

\item{\code{title_text}}{}

\item{\code{type}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ContentValidationDashboardElement$new()}}
\item \href{#method-toJSON}{\code{ContentValidationDashboardElement$toJSON()}}
\item \href{#method-fromJSONObject}{\code{ContentValidationDashboardElement$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{ContentValidationDashboardElement$fromJSON()}}
\item \href{#method-toJSONString}{\code{ContentValidationDashboardElement$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ContentValidationDashboardElement$fromJSONString()}}
\item \href{#method-clone}{\code{ContentValidationDashboardElement$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentValidationDashboardElement$new(
  body_text,
  dashboard_id,
  id,
  look_id,
  note_display,
  note_state,
  note_text,
  note_text_as_html,
  query_id,
  subtitle_text,
  title,
  title_hidden,
  title_text,
  type
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentValidationDashboardElement$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentValidationDashboardElement$fromJSONObject(
  ContentValidationDashboardElementJsonObject
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentValidationDashboardElement$fromJSON(
  ContentValidationDashboardElementJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentValidationDashboardElement$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentValidationDashboardElement$fromJSONString(
  ContentValidationDashboardElementJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentValidationDashboardElement$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
