% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContentValidationDashboardFilter.r
\name{ContentValidationDashboardFilter}
\alias{ContentValidationDashboardFilter}
\title{ContentValidationDashboardFilter Class}
\description{
ContentValidationDashboardFilter Class

ContentValidationDashboardFilter Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{}

\item{\code{dashboard_id}}{}

\item{\code{name}}{}

\item{\code{title}}{}

\item{\code{type}}{}

\item{\code{default_value}}{}

\item{\code{model}}{}

\item{\code{explore}}{}

\item{\code{dimension}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ContentValidationDashboardFilter$new()}}
\item \href{#method-toJSON}{\code{ContentValidationDashboardFilter$toJSON()}}
\item \href{#method-fromJSONObject}{\code{ContentValidationDashboardFilter$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{ContentValidationDashboardFilter$fromJSON()}}
\item \href{#method-toJSONString}{\code{ContentValidationDashboardFilter$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ContentValidationDashboardFilter$fromJSONString()}}
\item \href{#method-clone}{\code{ContentValidationDashboardFilter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentValidationDashboardFilter$new(
  id,
  dashboard_id,
  name,
  title,
  type,
  default_value,
  model,
  explore,
  dimension
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentValidationDashboardFilter$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentValidationDashboardFilter$fromJSONObject(
  ContentValidationDashboardFilterJsonObject
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentValidationDashboardFilter$fromJSON(ContentValidationDashboardFilterJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentValidationDashboardFilter$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentValidationDashboardFilter$fromJSONString(
  ContentValidationDashboardFilterJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentValidationDashboardFilter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
