% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContentValidationLookMLDashboardElement.r
\name{ContentValidationLookMLDashboardElement}
\alias{ContentValidationLookMLDashboardElement}
\title{ContentValidationLookMLDashboardElement Class}
\description{
ContentValidationLookMLDashboardElement Class

ContentValidationLookMLDashboardElement Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lookml_link_id}}{}

\item{\code{title}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ContentValidationLookMLDashboardElement$new()}}
\item \href{#method-toJSON}{\code{ContentValidationLookMLDashboardElement$toJSON()}}
\item \href{#method-fromJSONObject}{\code{ContentValidationLookMLDashboardElement$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{ContentValidationLookMLDashboardElement$fromJSON()}}
\item \href{#method-toJSONString}{\code{ContentValidationLookMLDashboardElement$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ContentValidationLookMLDashboardElement$fromJSONString()}}
\item \href{#method-clone}{\code{ContentValidationLookMLDashboardElement$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentValidationLookMLDashboardElement$new(lookml_link_id, title)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentValidationLookMLDashboardElement$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentValidationLookMLDashboardElement$fromJSONObject(
  ContentValidationLookMLDashboardElementJsonObject
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentValidationLookMLDashboardElement$fromJSON(
  ContentValidationLookMLDashboardElementJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentValidationLookMLDashboardElement$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentValidationLookMLDashboardElement$fromJSONString(
  ContentValidationLookMLDashboardElementJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentValidationLookMLDashboardElement$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
