% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateDashboardRenderTask.r
\docType{data}
\name{CreateDashboardRenderTask}
\alias{CreateDashboardRenderTask}
\title{CreateDashboardRenderTask Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CreateDashboardRenderTask
}
\description{
CreateDashboardRenderTask Class
}
\section{Fields}{

\describe{
\item{\code{dashboard_style}}{}

\item{\code{dashboard_filters}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
