% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateOAuthApplicationUserStateRequest.r
\name{CreateOAuthApplicationUserStateRequest}
\alias{CreateOAuthApplicationUserStateRequest}
\title{CreateOAuthApplicationUserStateRequest Class}
\description{
CreateOAuthApplicationUserStateRequest Class

CreateOAuthApplicationUserStateRequest Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{user_id}}{}

\item{\code{oauth_application_id}}{}

\item{\code{access_token}}{}

\item{\code{access_token_expires_at}}{}

\item{\code{refresh_token}}{}

\item{\code{refresh_token_expires_at}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CreateOAuthApplicationUserStateRequest$new()}}
\item \href{#method-toJSON}{\code{CreateOAuthApplicationUserStateRequest$toJSON()}}
\item \href{#method-fromJSONObject}{\code{CreateOAuthApplicationUserStateRequest$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{CreateOAuthApplicationUserStateRequest$fromJSON()}}
\item \href{#method-toJSONString}{\code{CreateOAuthApplicationUserStateRequest$toJSONString()}}
\item \href{#method-fromJSONString}{\code{CreateOAuthApplicationUserStateRequest$fromJSONString()}}
\item \href{#method-clone}{\code{CreateOAuthApplicationUserStateRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateOAuthApplicationUserStateRequest$new(
  user_id,
  oauth_application_id,
  access_token,
  access_token_expires_at,
  refresh_token,
  refresh_token_expires_at
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateOAuthApplicationUserStateRequest$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateOAuthApplicationUserStateRequest$fromJSONObject(
  CreateOAuthApplicationUserStateRequestJsonObject
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateOAuthApplicationUserStateRequest$fromJSON(
  CreateOAuthApplicationUserStateRequestJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateOAuthApplicationUserStateRequest$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateOAuthApplicationUserStateRequest$fromJSONString(
  CreateOAuthApplicationUserStateRequestJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateOAuthApplicationUserStateRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
