% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateOAuthApplicationUserStateResponse.r
\name{CreateOAuthApplicationUserStateResponse}
\alias{CreateOAuthApplicationUserStateResponse}
\title{CreateOAuthApplicationUserStateResponse Class}
\description{
CreateOAuthApplicationUserStateResponse Class

CreateOAuthApplicationUserStateResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{user_id}}{}

\item{\code{oauth_application_id}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CreateOAuthApplicationUserStateResponse$new()}}
\item \href{#method-toJSON}{\code{CreateOAuthApplicationUserStateResponse$toJSON()}}
\item \href{#method-fromJSONObject}{\code{CreateOAuthApplicationUserStateResponse$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{CreateOAuthApplicationUserStateResponse$fromJSON()}}
\item \href{#method-toJSONString}{\code{CreateOAuthApplicationUserStateResponse$toJSONString()}}
\item \href{#method-fromJSONString}{\code{CreateOAuthApplicationUserStateResponse$fromJSONString()}}
\item \href{#method-clone}{\code{CreateOAuthApplicationUserStateResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateOAuthApplicationUserStateResponse$new(user_id, oauth_application_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateOAuthApplicationUserStateResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateOAuthApplicationUserStateResponse$fromJSONObject(
  CreateOAuthApplicationUserStateResponseJsonObject
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateOAuthApplicationUserStateResponse$fromJSON(
  CreateOAuthApplicationUserStateResponseJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateOAuthApplicationUserStateResponse$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateOAuthApplicationUserStateResponse$fromJSONString(
  CreateOAuthApplicationUserStateResponseJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateOAuthApplicationUserStateResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
