% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateQueryTask.r
\docType{data}
\name{CreateQueryTask}
\alias{CreateQueryTask}
\title{CreateQueryTask Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CreateQueryTask
}
\description{
CreateQueryTask Class
}
\section{Fields}{

\describe{
\item{\code{query_id}}{}

\item{\code{result_format}}{}

\item{\code{source}}{}

\item{\code{deferred}}{}

\item{\code{look_id}}{}

\item{\code{dashboard_id}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
