% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CredentialsApi.r
\docType{data}
\name{CredentialsApi}
\alias{CredentialsApi}
\title{CredentialsApi Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CredentialsApi
}
\description{
CredentialsApi Class
}
\section{Fields}{

\describe{
\item{\code{token}}{}

\item{\code{created_at}}{}

\item{\code{is_disabled}}{}

\item{\code{type}}{}

\item{\code{url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
