% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CredentialsEmail.r
\docType{data}
\name{CredentialsEmail}
\alias{CredentialsEmail}
\title{CredentialsEmail Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CredentialsEmail
}
\description{
CredentialsEmail Class
}
\section{Fields}{

\describe{
\item{\code{email}}{}

\item{\code{created_at}}{}

\item{\code{logged_in_at}}{}

\item{\code{is_disabled}}{}

\item{\code{type}}{}

\item{\code{password_reset_url}}{}

\item{\code{url}}{}

\item{\code{user_url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
