% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CredentialsEmailSearch.r
\name{CredentialsEmailSearch}
\alias{CredentialsEmailSearch}
\title{CredentialsEmailSearch Class}
\description{
CredentialsEmailSearch Class

CredentialsEmailSearch Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{can}}{}

\item{\code{created_at}}{}

\item{\code{email}}{}

\item{\code{forced_password_reset_at_next_login}}{}

\item{\code{is_disabled}}{}

\item{\code{logged_in_at}}{}

\item{\code{password_reset_url}}{}

\item{\code{type}}{}

\item{\code{url}}{}

\item{\code{user_url}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CredentialsEmailSearch$new()}}
\item \href{#method-toJSON}{\code{CredentialsEmailSearch$toJSON()}}
\item \href{#method-fromJSONObject}{\code{CredentialsEmailSearch$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{CredentialsEmailSearch$fromJSON()}}
\item \href{#method-toJSONString}{\code{CredentialsEmailSearch$toJSONString()}}
\item \href{#method-fromJSONString}{\code{CredentialsEmailSearch$fromJSONString()}}
\item \href{#method-clone}{\code{CredentialsEmailSearch$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsEmailSearch$new(
  can,
  created_at,
  email,
  forced_password_reset_at_next_login,
  is_disabled,
  logged_in_at,
  password_reset_url,
  type,
  url,
  user_url
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsEmailSearch$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsEmailSearch$fromJSONObject(CredentialsEmailSearchJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsEmailSearch$fromJSON(CredentialsEmailSearchJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsEmailSearch$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsEmailSearch$fromJSONString(CredentialsEmailSearchJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsEmailSearch$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
