% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CredentialsEmbed.r
\docType{data}
\name{CredentialsEmbed}
\alias{CredentialsEmbed}
\title{CredentialsEmbed Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CredentialsEmbed
}
\description{
CredentialsEmbed Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{external_user_id}}{}

\item{\code{external_group_id}}{}

\item{\code{created_at}}{}

\item{\code{logged_in_at}}{}

\item{\code{is_disabled}}{}

\item{\code{type}}{}

\item{\code{url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
