% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CredentialsGoogle.r
\docType{data}
\name{CredentialsGoogle}
\alias{CredentialsGoogle}
\title{CredentialsGoogle Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CredentialsGoogle
}
\description{
CredentialsGoogle Class
}
\section{Fields}{

\describe{
\item{\code{email}}{}

\item{\code{google_user_id}}{}

\item{\code{domain}}{}

\item{\code{created_at}}{}

\item{\code{logged_in_at}}{}

\item{\code{is_disabled}}{}

\item{\code{type}}{}

\item{\code{url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
