% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CredentialsLDAP.r
\docType{data}
\name{CredentialsLDAP}
\alias{CredentialsLDAP}
\title{CredentialsLDAP Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CredentialsLDAP
}
\description{
CredentialsLDAP Class
}
\section{Fields}{

\describe{
\item{\code{email}}{}

\item{\code{ldap_id}}{}

\item{\code{ldap_dn}}{}

\item{\code{created_at}}{}

\item{\code{logged_in_at}}{}

\item{\code{is_disabled}}{}

\item{\code{type}}{}

\item{\code{url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
