% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CredentialsOIDC.r
\docType{data}
\name{CredentialsOIDC}
\alias{CredentialsOIDC}
\title{CredentialsOIDC Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CredentialsOIDC
}
\description{
CredentialsOIDC Class
}
\section{Fields}{

\describe{
\item{\code{email}}{}

\item{\code{oidc_user_id}}{}

\item{\code{created_at}}{}

\item{\code{logged_in_at}}{}

\item{\code{is_disabled}}{}

\item{\code{type}}{}

\item{\code{url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
