% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CredentialsTotp.r
\docType{data}
\name{CredentialsTotp}
\alias{CredentialsTotp}
\title{CredentialsTotp Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CredentialsTotp
}
\description{
CredentialsTotp Class
}
\section{Fields}{

\describe{
\item{\code{verified}}{}

\item{\code{created_at}}{}

\item{\code{is_disabled}}{}

\item{\code{type}}{}

\item{\code{url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
