% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dashboard.r
\docType{data}
\name{Dashboard}
\alias{Dashboard}
\title{Dashboard Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Dashboard
}
\description{
Dashboard Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{content_metadata_id}}{}

\item{\code{content_favorite_id}}{}

\item{\code{view_count}}{}

\item{\code{last_accessed_at}}{}

\item{\code{favorite_count}}{}

\item{\code{user_id}}{}

\item{\code{title}}{}

\item{\code{description}}{}

\item{\code{readonly}}{}

\item{\code{hidden}}{}

\item{\code{refresh_interval}}{}

\item{\code{refresh_interval_to_i}}{}

\item{\code{space}}{}

\item{\code{load_configuration}}{}

\item{\code{model}}{}

\item{\code{space_id}}{}

\item{\code{dashboard_elements}}{}

\item{\code{dashboard_layouts}}{}

\item{\code{dashboard_filters}}{}

\item{\code{last_viewed_at}}{}

\item{\code{background_color}}{}

\item{\code{show_title}}{}

\item{\code{title_color}}{}

\item{\code{show_filters_bar}}{}

\item{\code{tile_background_color}}{}

\item{\code{tile_text_color}}{}

\item{\code{text_tile_text_color}}{}

\item{\code{last_updater_id}}{}

\item{\code{deleter_id}}{}

\item{\code{deleted}}{}

\item{\code{created_at}}{}

\item{\code{deleted_at}}{}

\item{\code{query_timezone}}{}

\item{\code{edit_uri}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
