% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DashboardAggregateTableLookml.r
\name{DashboardAggregateTableLookml}
\alias{DashboardAggregateTableLookml}
\title{DashboardAggregateTableLookml Class}
\description{
DashboardAggregateTableLookml Class

DashboardAggregateTableLookml Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dashboard_id}}{}

\item{\code{aggregate_table_lookml}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DashboardAggregateTableLookml$new()}}
\item \href{#method-toJSON}{\code{DashboardAggregateTableLookml$toJSON()}}
\item \href{#method-fromJSONObject}{\code{DashboardAggregateTableLookml$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{DashboardAggregateTableLookml$fromJSON()}}
\item \href{#method-toJSONString}{\code{DashboardAggregateTableLookml$toJSONString()}}
\item \href{#method-fromJSONString}{\code{DashboardAggregateTableLookml$fromJSONString()}}
\item \href{#method-clone}{\code{DashboardAggregateTableLookml$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardAggregateTableLookml$new(dashboard_id, aggregate_table_lookml)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardAggregateTableLookml$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardAggregateTableLookml$fromJSONObject(
  DashboardAggregateTableLookmlJsonObject
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardAggregateTableLookml$fromJSON(DashboardAggregateTableLookmlJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardAggregateTableLookml$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardAggregateTableLookml$fromJSONString(DashboardAggregateTableLookmlJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardAggregateTableLookml$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
