% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DashboardApi.r
\docType{data}
\name{DashboardApi}
\alias{DashboardApi}
\title{Dashboard operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DashboardApi
}
\description{
looker.Dashboard
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

all_dashboards Get All Dashboards 


create_dashboard Create Dashboard 


create_dashboard_prefetch Create Dashboard Prefetch 


dashboard Get Dashboard 


dashboard_prefetch Get Dashboard Prefetch 


delete_dashboard Delete Dashboard 


search_dashboards Search Dashboards 


update_dashboard Update Dashboard 

}
}

\keyword{datasets}
