% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DashboardAppearance.r
\name{DashboardAppearance}
\alias{DashboardAppearance}
\title{DashboardAppearance Class}
\description{
DashboardAppearance Class

DashboardAppearance Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{page_side_margins}}{}

\item{\code{page_background_color}}{}

\item{\code{tile_title_alignment}}{}

\item{\code{tile_space_between}}{}

\item{\code{tile_background_color}}{}

\item{\code{tile_shadow}}{}

\item{\code{key_color}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DashboardAppearance$new()}}
\item \href{#method-toJSON}{\code{DashboardAppearance$toJSON()}}
\item \href{#method-fromJSONObject}{\code{DashboardAppearance$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{DashboardAppearance$fromJSON()}}
\item \href{#method-toJSONString}{\code{DashboardAppearance$toJSONString()}}
\item \href{#method-fromJSONString}{\code{DashboardAppearance$fromJSONString()}}
\item \href{#method-clone}{\code{DashboardAppearance$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardAppearance$new(
  page_side_margins,
  page_background_color,
  tile_title_alignment,
  tile_space_between,
  tile_background_color,
  tile_shadow,
  key_color
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardAppearance$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardAppearance$fromJSONObject(DashboardAppearanceJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardAppearance$fromJSON(DashboardAppearanceJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardAppearance$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardAppearance$fromJSONString(DashboardAppearanceJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardAppearance$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
