% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DashboardBase.r
\docType{data}
\name{DashboardBase}
\alias{DashboardBase}
\title{DashboardBase Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DashboardBase
}
\description{
DashboardBase Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{user_id}}{}

\item{\code{title}}{}

\item{\code{description}}{}

\item{\code{readonly}}{}

\item{\code{hidden}}{}

\item{\code{refresh_interval}}{}

\item{\code{refresh_interval_to_i}}{}

\item{\code{space}}{}

\item{\code{model}}{}

\item{\code{content_favorite_id}}{}

\item{\code{scheduled_plan}}{}

\item{\code{content_metadata_id}}{}

\item{\code{query_timezone}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
