% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DashboardElement.r
\docType{data}
\name{DashboardElement}
\alias{DashboardElement}
\title{DashboardElement Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DashboardElement
}
\description{
DashboardElement Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{dashboard_id}}{}

\item{\code{look_id}}{}

\item{\code{query_id}}{}

\item{\code{type}}{}

\item{\code{listen}}{}

\item{\code{refresh_interval}}{}

\item{\code{refresh_interval_to_i}}{}

\item{\code{note_text}}{}

\item{\code{note_text_as_html}}{}

\item{\code{note_display}}{}

\item{\code{note_state}}{}

\item{\code{title_hidden}}{}

\item{\code{title_text}}{}

\item{\code{title}}{}

\item{\code{subtitle_text}}{}

\item{\code{body_text}}{}

\item{\code{body_text_as_html}}{}

\item{\code{look}}{}

\item{\code{query}}{}

\item{\code{edit_uri}}{}

\item{\code{merge_result_id}}{}

\item{\code{result_maker_id}}{}

\item{\code{result_maker}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
