% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DashboardFilter.r
\name{DashboardFilter}
\alias{DashboardFilter}
\title{DashboardFilter Class}
\description{
DashboardFilter Class

DashboardFilter Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{}

\item{\code{dashboard_id}}{}

\item{\code{name}}{}

\item{\code{title}}{}

\item{\code{type}}{}

\item{\code{default_value}}{}

\item{\code{model}}{}

\item{\code{explore}}{}

\item{\code{dimension}}{}

\item{\code{field}}{}

\item{\code{row}}{}

\item{\code{listens_to_filters}}{}

\item{\code{allow_multiple_values}}{}

\item{\code{required}}{}

\item{\code{can}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DashboardFilter$new()}}
\item \href{#method-toJSON}{\code{DashboardFilter$toJSON()}}
\item \href{#method-fromJSON}{\code{DashboardFilter$fromJSON()}}
\item \href{#method-toJSONString}{\code{DashboardFilter$toJSONString()}}
\item \href{#method-fromJSONString}{\code{DashboardFilter$fromJSONString()}}
\item \href{#method-clone}{\code{DashboardFilter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardFilter$new(
  id,
  dashboard_id,
  name,
  title,
  type,
  default_value,
  model,
  explore,
  dimension,
  field,
  row,
  listens_to_filters,
  allow_multiple_values,
  required,
  can
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardFilter$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardFilter$fromJSON(DashboardFilterJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardFilter$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardFilter$fromJSONString(DashboardFilterJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardFilter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
