% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DashboardLayout.r
\docType{data}
\name{DashboardLayout}
\alias{DashboardLayout}
\title{DashboardLayout Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DashboardLayout
}
\description{
DashboardLayout Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{dashboard_id}}{}

\item{\code{type}}{}

\item{\code{active}}{}

\item{\code{column_width}}{}

\item{\code{width}}{}

\item{\code{deleted}}{}

\item{\code{dashboard_layout_components}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
