% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DashboardLayoutComponent.r
\docType{data}
\name{DashboardLayoutComponent}
\alias{DashboardLayoutComponent}
\title{DashboardLayoutComponent Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DashboardLayoutComponent
}
\description{
DashboardLayoutComponent Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{dashboard_layout_id}}{}

\item{\code{dashboard_element_id}}{}

\item{\code{row}}{}

\item{\code{column}}{}

\item{\code{width}}{}

\item{\code{height}}{}

\item{\code{deleted}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
