% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataActionApi.r
\name{DataActionApi}
\alias{DataActionApi}
\title{DataAction operations}
\description{
looker.DataAction
}
\section{Methods}{

\describe{

fetch_remote_data_action_form Fetch Remote Data Action Form 


perform_data_action Send a Data Action 

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataActionApi$new()}}
\item \href{#method-fetch_remote_data_action_form}{\code{DataActionApi$fetch_remote_data_action_form()}}
\item \href{#method-perform_data_action}{\code{DataActionApi$perform_data_action()}}
\item \href{#method-clone}{\code{DataActionApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataActionApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fetch_remote_data_action_form"></a>}}
\if{latex}{\out{\hypertarget{method-fetch_remote_data_action_form}{}}}
\subsection{Method \code{fetch_remote_data_action_form()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataActionApi$fetch_remote_data_action_form(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-perform_data_action"></a>}}
\if{latex}{\out{\hypertarget{method-perform_data_action}{}}}
\subsection{Method \code{perform_data_action()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataActionApi$perform_data_action(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataActionApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
