% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataActionForm.r
\docType{data}
\name{DataActionForm}
\alias{DataActionForm}
\title{DataActionForm Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataActionForm
}
\description{
DataActionForm Class
}
\section{Fields}{

\describe{
\item{\code{fields}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
