% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataActionFormField.r
\docType{data}
\name{DataActionFormField}
\alias{DataActionFormField}
\title{DataActionFormField Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataActionFormField
}
\description{
DataActionFormField Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{label}}{}

\item{\code{description}}{}

\item{\code{type}}{}

\item{\code{default}}{}

\item{\code{required}}{}

\item{\code{options}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
