% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataActionFormSelectOption.r
\docType{data}
\name{DataActionFormSelectOption}
\alias{DataActionFormSelectOption}
\title{DataActionFormSelectOption Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataActionFormSelectOption
}
\description{
DataActionFormSelectOption Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{label}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
