% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataActionRequest.r
\docType{data}
\name{DataActionRequest}
\alias{DataActionRequest}
\title{DataActionRequest Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataActionRequest
}
\description{
DataActionRequest Class
}
\section{Fields}{

\describe{
\item{\code{action}}{}

\item{\code{form_values}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
