% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datagroup.r
\docType{data}
\name{Datagroup}
\alias{Datagroup}
\title{Datagroup Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Datagroup
}
\description{
Datagroup Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{name}}{}

\item{\code{model_name}}{}

\item{\code{trigger_value}}{}

\item{\code{trigger_error}}{}

\item{\code{stale_before}}{}

\item{\code{triggered_at}}{}

\item{\code{trigger_check_at}}{}

\item{\code{created_at}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
