% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatagroupApi.r
\name{DatagroupApi}
\alias{DatagroupApi}
\title{Datagroup operations}
\description{
looker.Datagroup
}
\section{Methods}{

\describe{

all_datagroups Get All Datagroups


datagroup Get Datagroup


update_datagroup Update Datagroup

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DatagroupApi$new()}}
\item \href{#method-all_datagroups}{\code{DatagroupApi$all_datagroups()}}
\item \href{#method-datagroup}{\code{DatagroupApi$datagroup()}}
\item \href{#method-update_datagroup}{\code{DatagroupApi$update_datagroup()}}
\item \href{#method-clone}{\code{DatagroupApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatagroupApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_datagroups"></a>}}
\if{latex}{\out{\hypertarget{method-all_datagroups}{}}}
\subsection{Method \code{all_datagroups()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatagroupApi$all_datagroups(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-datagroup"></a>}}
\if{latex}{\out{\hypertarget{method-datagroup}{}}}
\subsection{Method \code{datagroup()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatagroupApi$datagroup(datagroup_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_datagroup"></a>}}
\if{latex}{\out{\hypertarget{method-update_datagroup}{}}}
\subsection{Method \code{update_datagroup()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatagroupApi$update_datagroup(datagroup_id, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatagroupApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
