% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBConnectionBase.r
\docType{data}
\name{DBConnectionBase}
\alias{DBConnectionBase}
\title{DBConnectionBase Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DBConnectionBase
}
\description{
DBConnectionBase Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{dialect}}{}

\item{\code{snippets}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
