% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBConnectionOverride.r
\name{DBConnectionOverride}
\alias{DBConnectionOverride}
\title{DBConnectionOverride Class}
\description{
DBConnectionOverride Class

DBConnectionOverride Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{context}}{}

\item{\code{host}}{}

\item{\code{port}}{}

\item{\code{username}}{}

\item{\code{password}}{}

\item{\code{has_password}}{}

\item{\code{certificate}}{}

\item{\code{file_type}}{}

\item{\code{database}}{}

\item{\code{schema}}{}

\item{\code{jdbc_additional_params}}{}

\item{\code{after_connect_statements}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DBConnectionOverride$new()}}
\item \href{#method-toJSON}{\code{DBConnectionOverride$toJSON()}}
\item \href{#method-fromJSONObject}{\code{DBConnectionOverride$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{DBConnectionOverride$fromJSON()}}
\item \href{#method-toJSONString}{\code{DBConnectionOverride$toJSONString()}}
\item \href{#method-fromJSONString}{\code{DBConnectionOverride$fromJSONString()}}
\item \href{#method-clone}{\code{DBConnectionOverride$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBConnectionOverride$new(
  context,
  host,
  port,
  username,
  password,
  has_password,
  certificate,
  file_type,
  database,
  schema,
  jdbc_additional_params,
  after_connect_statements
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBConnectionOverride$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBConnectionOverride$fromJSONObject(DBConnectionOverrideJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBConnectionOverride$fromJSON(DBConnectionOverrideJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBConnectionOverride$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBConnectionOverride$fromJSONString(DBConnectionOverrideJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBConnectionOverride$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
