% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBConnectionTestResult.r
\docType{data}
\name{DBConnectionTestResult}
\alias{DBConnectionTestResult}
\title{DBConnectionTestResult Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DBConnectionTestResult
}
\description{
DBConnectionTestResult Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{status}}{}

\item{\code{message}}{}

\item{\code{connection_string}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
