% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dialect.r
\docType{data}
\name{Dialect}
\alias{Dialect}
\title{Dialect Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Dialect
}
\description{
Dialect Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{label}}{}

\item{\code{supports_cost_estimate}}{}

\item{\code{supports_upload_tables}}{}

\item{\code{persistent_table_indexes}}{}

\item{\code{persistent_table_sortkeys}}{}

\item{\code{persistent_table_distkey}}{}

\item{\code{supports_streaming}}{}

\item{\code{automatically_run_sql_runner_snippets}}{}

\item{\code{connection_tests}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
