% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/firstclass.R
\name{DropTrials}
\alias{DropTrials}
\title{Drop trials from a Session}
\usage{
DropTrials(trials, drop)
}
\arguments{
\item{trials}{a `list` of `Trial` objects}

\item{drop}{an integer vector specifying which trials should be dropped}
}
\value{
the `list` of `Trial` objects minus the dropped trials. The classes
  of the input list are preserved in the outputted list.
}
\description{
Drop trials from a Session
}

