% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmbedSsoParams.r
\name{EmbedSsoParams}
\alias{EmbedSsoParams}
\title{EmbedSsoParams Class}
\description{
EmbedSsoParams Class

EmbedSsoParams Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{target_url}}{}

\item{\code{session_length}}{}

\item{\code{force_logout_login}}{}

\item{\code{external_user_id}}{}

\item{\code{first_name}}{}

\item{\code{last_name}}{}

\item{\code{user_timezone}}{}

\item{\code{permissions}}{}

\item{\code{models}}{}

\item{\code{group_ids}}{}

\item{\code{external_group_id}}{}

\item{\code{user_attributes}}{}

\item{\code{secret_id}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EmbedSsoParams$new()}}
\item \href{#method-toJSON}{\code{EmbedSsoParams$toJSON()}}
\item \href{#method-fromJSONObject}{\code{EmbedSsoParams$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{EmbedSsoParams$fromJSON()}}
\item \href{#method-toJSONString}{\code{EmbedSsoParams$toJSONString()}}
\item \href{#method-fromJSONString}{\code{EmbedSsoParams$fromJSONString()}}
\item \href{#method-clone}{\code{EmbedSsoParams$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmbedSsoParams$new(
  target_url,
  session_length,
  force_logout_login,
  external_user_id,
  first_name,
  last_name,
  user_timezone,
  permissions,
  models,
  group_ids,
  external_group_id,
  user_attributes,
  secret_id
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmbedSsoParams$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmbedSsoParams$fromJSONObject(EmbedSsoParamsJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmbedSsoParams$fromJSON(EmbedSsoParamsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmbedSsoParams$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmbedSsoParams$fromJSONString(EmbedSsoParamsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmbedSsoParams$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
