% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gca.R
\name{empirical_logit}
\alias{empirical_logit}
\alias{empirical_logit_weight}
\title{Compute empirical logit}
\usage{
empirical_logit(x, y)

empirical_logit_weight(x, y)
}
\arguments{
\item{x}{vector containing number of looks to target}

\item{y}{vector containing number of looks to distractors}
}
\value{
\code{empirical_logit} returns the empirical logit of looking to
  target. \code{empirical_logit_weight} returns weights for these values.
}
\description{
Compute empirical logit
}
\references{
Dale Barr's
\href{http://talklab.psy.gla.ac.uk/tvw/elogit-wt.html}{Walkthrough of an
"empirical logit" analysis in R}
}

