% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Error.r
\docType{data}
\name{Error}
\alias{Error}
\title{Error Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Error
}
\description{
Error Class
}
\section{Fields}{

\describe{
\item{\code{message}}{}

\item{\code{documentation_url}}{}
}}

\keyword{datasets}
