% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalOauthApplication.r
\name{ExternalOauthApplication}
\alias{ExternalOauthApplication}
\title{ExternalOauthApplication Class}
\description{
ExternalOauthApplication Class

ExternalOauthApplication Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{can}}{}

\item{\code{id}}{}

\item{\code{name}}{}

\item{\code{client_id}}{}

\item{\code{client_secret}}{}

\item{\code{dialect_name}}{}

\item{\code{created_at}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ExternalOauthApplication$new()}}
\item \href{#method-toJSON}{\code{ExternalOauthApplication$toJSON()}}
\item \href{#method-fromJSONObject}{\code{ExternalOauthApplication$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{ExternalOauthApplication$fromJSON()}}
\item \href{#method-toJSONString}{\code{ExternalOauthApplication$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ExternalOauthApplication$fromJSONString()}}
\item \href{#method-clone}{\code{ExternalOauthApplication$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalOauthApplication$new(
  can,
  id,
  name,
  client_id,
  client_secret,
  dialect_name,
  created_at
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalOauthApplication$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalOauthApplication$fromJSONObject(ExternalOauthApplicationJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalOauthApplication$fromJSON(ExternalOauthApplicationJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalOauthApplication$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalOauthApplication$fromJSONString(ExternalOauthApplicationJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalOauthApplication$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
