% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/time.R
\name{FindClosestFrame}
\alias{FindClosestFrame}
\title{Find frame of a trial that is closest to a given time}
\usage{
FindClosestFrame(trial, time)
}
\arguments{
\item{trial}{a \code{Trial} object}

\item{time}{a \code{numeric} value (in ms.)}
}
\value{
The number of the frame of the trial whose time is closest to the
  given time.
}
\description{
If two frames are temporally equidistant from the given time, choose the
earlier frame.
}

