% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FolderApi.r
\name{FolderApi}
\alias{FolderApi}
\title{Folder operations}
\description{
looker.Folder
}
\section{Methods}{

\describe{

all_folders Get All Folders


create_folder Create Folder


delete_folder Delete Folder


folder Get Folder


folder_ancestors Get Folder Ancestors


folder_children Get Folder Children


folder_children_search Search Folder Children


folder_dashboards Get Folder Dashboards


folder_looks Get Folder Looks


folder_parent Get Folder Parent


search_folders Search Folders


update_folder Update Folder

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FolderApi$new()}}
\item \href{#method-all_folders}{\code{FolderApi$all_folders()}}
\item \href{#method-create_folder}{\code{FolderApi$create_folder()}}
\item \href{#method-delete_folder}{\code{FolderApi$delete_folder()}}
\item \href{#method-folder}{\code{FolderApi$folder()}}
\item \href{#method-folder_ancestors}{\code{FolderApi$folder_ancestors()}}
\item \href{#method-folder_children}{\code{FolderApi$folder_children()}}
\item \href{#method-folder_children_search}{\code{FolderApi$folder_children_search()}}
\item \href{#method-folder_dashboards}{\code{FolderApi$folder_dashboards()}}
\item \href{#method-folder_looks}{\code{FolderApi$folder_looks()}}
\item \href{#method-folder_parent}{\code{FolderApi$folder_parent()}}
\item \href{#method-search_folders}{\code{FolderApi$search_folders()}}
\item \href{#method-update_folder}{\code{FolderApi$update_folder()}}
\item \href{#method-clone}{\code{FolderApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FolderApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_folders"></a>}}
\if{latex}{\out{\hypertarget{method-all_folders}{}}}
\subsection{Method \code{all_folders()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FolderApi$all_folders(fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_folder"></a>}}
\if{latex}{\out{\hypertarget{method-create_folder}{}}}
\subsection{Method \code{create_folder()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FolderApi$create_folder(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_folder"></a>}}
\if{latex}{\out{\hypertarget{method-delete_folder}{}}}
\subsection{Method \code{delete_folder()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FolderApi$delete_folder(folder_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-folder"></a>}}
\if{latex}{\out{\hypertarget{method-folder}{}}}
\subsection{Method \code{folder()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FolderApi$folder(folder_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-folder_ancestors"></a>}}
\if{latex}{\out{\hypertarget{method-folder_ancestors}{}}}
\subsection{Method \code{folder_ancestors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FolderApi$folder_ancestors(folder_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-folder_children"></a>}}
\if{latex}{\out{\hypertarget{method-folder_children}{}}}
\subsection{Method \code{folder_children()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FolderApi$folder_children(folder_id, fields, page, per_page, sorts, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-folder_children_search"></a>}}
\if{latex}{\out{\hypertarget{method-folder_children_search}{}}}
\subsection{Method \code{folder_children_search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FolderApi$folder_children_search(folder_id, fields, sorts, name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-folder_dashboards"></a>}}
\if{latex}{\out{\hypertarget{method-folder_dashboards}{}}}
\subsection{Method \code{folder_dashboards()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FolderApi$folder_dashboards(folder_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-folder_looks"></a>}}
\if{latex}{\out{\hypertarget{method-folder_looks}{}}}
\subsection{Method \code{folder_looks()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FolderApi$folder_looks(folder_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-folder_parent"></a>}}
\if{latex}{\out{\hypertarget{method-folder_parent}{}}}
\subsection{Method \code{folder_parent()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FolderApi$folder_parent(folder_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_folders"></a>}}
\if{latex}{\out{\hypertarget{method-search_folders}{}}}
\subsection{Method \code{search_folders()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FolderApi$search_folders(
  fields,
  page,
  per_page,
  limit,
  offset,
  sorts,
  name,
  id,
  parent_id,
  creator_id,
  filter_or,
  is_shared_root,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_folder"></a>}}
\if{latex}{\out{\hypertarget{method-update_folder}{}}}
\subsection{Method \code{update_folder()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FolderApi$update_folder(folder_id, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FolderApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
