% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FolderBase.r
\name{FolderBase}
\alias{FolderBase}
\title{FolderBase Class}
\description{
FolderBase Class

FolderBase Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{}

\item{\code{parent_id}}{}

\item{\code{id}}{}

\item{\code{content_metadata_id}}{}

\item{\code{created_at}}{}

\item{\code{creator_id}}{}

\item{\code{child_count}}{}

\item{\code{external_id}}{}

\item{\code{is_embed}}{}

\item{\code{is_embed_shared_root}}{}

\item{\code{is_embed_users_root}}{}

\item{\code{is_personal}}{}

\item{\code{is_personal_descendant}}{}

\item{\code{is_shared_root}}{}

\item{\code{is_users_root}}{}

\item{\code{can}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FolderBase$new()}}
\item \href{#method-toJSON}{\code{FolderBase$toJSON()}}
\item \href{#method-fromJSONObject}{\code{FolderBase$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{FolderBase$fromJSON()}}
\item \href{#method-toJSONString}{\code{FolderBase$toJSONString()}}
\item \href{#method-fromJSONString}{\code{FolderBase$fromJSONString()}}
\item \href{#method-clone}{\code{FolderBase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FolderBase$new(
  name,
  parent_id,
  id,
  content_metadata_id,
  created_at,
  creator_id,
  child_count,
  external_id,
  is_embed,
  is_embed_shared_root,
  is_embed_users_root,
  is_personal,
  is_personal_descendant,
  is_shared_root,
  is_users_root,
  can
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FolderBase$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FolderBase$fromJSONObject(FolderBaseJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FolderBase$fromJSON(FolderBaseJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FolderBase$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FolderBase$fromJSONString(FolderBaseJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FolderBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
