% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/attributes.R
\name{gather_attributes}
\alias{gather_attributes}
\title{Create a data-frame of attributes}
\usage{
gather_attributes(x, attrs, omit_na = FALSE)
}
\arguments{
\item{x}{a list of objects with attributes}

\item{attrs}{a character vector. If an element in the vector is named, the
name is used as the new (column) name for that attribute's column in the
resulting data-frame.}

\item{omit_na}{whether to drop columns with all \code{NA} values}
}
\value{
a data-frame with a column for each of the attributes in
  \code{attrs}. By default, non-existent attributes will be stored with NA
  values and strings are not factors.
}
\description{
This function is intended to abstract over the pattern of writing code like
\code{data.frame(x = trials \%@\% "x", ZName = trials \%@\% "z")} by
instead allowing us to write \code{gather_attributes(trials, c("x", ZName =
z))}.
}

