% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gazedata.R
\name{Gazedata}
\alias{Gazedata}
\title{Load a \code{.gazedata} file for an experiment}
\usage{
Gazedata(gazedata_path)
}
\arguments{
\item{gazedata_path}{path to the \code{.gazedata} file that is to be parsed.}
}
\value{
A dataframe containing the parsed gazedata. Each row of the dataframe
  contains the eye-tracking data for a single frame of time recorded during
  the experiment. The dataframe has the additional class of \code{Gazedata}.
}
\description{
Load a \code{.gazedata} file for an experiment
}
\details{
The following columns are captured in the gazedata file:
\tabular{ll}{
   Subject              \tab Session               \cr
   ID                   \tab TrialID               \cr
   TETTime              \tab RTTime                \cr
   CursorX              \tab CursorY               \cr
   TimestampSec         \tab TimestampMicrosec     \cr
   XGazePosLeftEye      \tab XGazePosRightEye      \cr
   YGazePosLeftEye      \tab YGazePosRightEye      \cr
   XCameraPosLeftEye    \tab XCameraPosRightEye    \cr
   YCameraPosLeftEye    \tab YCameraPosRightEye    \cr
   DiameterPupilLeftEye \tab DiameterPupilRightEye \cr
   DistanceLeftEye      \tab DistanceRightEye      \cr
   ValidityLeftEye      \tab ValidityRightEye      \cr
   Target               \tab Stimulus
}

We extract the columns the following columns: \code{TrialId}, \code{RTTime},
\code{XGazePosLeftEye}, \code{XGazePosRightEye}, \code{YGazePosLeftEye},
\code{YGazePosRightEye}, \code{DistanceLeftEye}, \code{DistanceRightEye},
\code{DiameterPupilLeftEye} and \code{DiameterPupilRightEye}.

Once these column values are loaded, we make three modifications to the
gazedata.

\enumerate{
  \item Gaze measurements with \code{Validity} codes greater than or equal to
  1 are replaced with NA values.

  \item X,Y gaze values are defined in screen proportions. Values that fall
  outside [0,1] are outside of the boundaries of the screen and therefore are
  nonsensical. Replace them with \code{NA}. We perform a similar correction
  on pupil diameters and eye-distances by replacing negative
  values with \code{NA}.

  \item The origin of the screen is the upper-left-hand corner of the screen.
  Flip the y-values so that the origin is in a more familiar position in the
  lower-left-hand corner of the screen. This way, low y values are closer to
  the bottom of the screen.

  \item Compute the mean x, y, distance and diameter values for the left and
  right eyes. \code{NA} values are ignored when computing the mean, so the
  pair \code{(XLeft = NA, XRight = .5)} yields \code{XMean = .5}.
}

Information about the task, block number, and subject id are extracted from
the basename of the gazedata file. This function expects the gazedata file
to have the format \code{[Task]_[BlockNo]_[SubjectID].gazedata}.

Information about the stimuli for each trial is extracted from the associated
\code{.txt} file that is output by E-prime. These values are handled by the
\code{Stimdata} function.
}
\references{
\href{http://bit.ly/1AtKyhR}{Tobii Toolbox for Matlab: Product
  Description & User Guide}
}

