% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/missing.R
\name{GetDwellTimes}
\alias{GetDwellTimes}
\title{Compute dwell times}
\usage{
GetDwellTimes(x)
}
\arguments{
\item{x}{a Trial with a \code{GazeByImageAOI} column or a TrialList where
each Trial has a \code{GazeByImageAOI} column}
}
\value{
the Trial object(s) updated to include a \code{DwellSummary}
  attribute which contains a dataframe describing the maximum and total dwell
  times for each image location and stimulus.
}
\description{
A "dwell" is an uninterrupted gaze. This function calculates the longest
dwell time per image and the total time dwelling in each trial.
}

