% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/time.R
\name{GetFrameAtTime}
\alias{GetFrameAtTime}
\title{Get the frame number for a certain Time value in a list of Trials}
\usage{
GetFrameAtTime(trials, time_point = 0)
}
\arguments{
\item{trials}{a list of \code{Trial} objects}

\item{time_point}{a numeric value indicating the time to find in the Trial}
}
\value{
a numeric vector listing the frames at which the desired time occurs
  in each \code{Trial}
}
\description{
Get the frame number for a certain Time value in a list of Trials
}
\keyword{internal}

