% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetTrialLengths}
\alias{GetTrialLengths}
\title{Get the lengths (in frames) of a list of trials}
\usage{
GetTrialLengths(trials)
}
\arguments{
\item{trials}{a list of \code{Trial} objects.}
}
\value{
a numeric vector with the number of frames in each \code{Trial}.
}
\description{
Get the lengths (in frames) of a list of trials
}
\keyword{internal}

