% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitBranch.r
\docType{data}
\name{GitBranch}
\alias{GitBranch}
\title{GitBranch Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GitBranch
}
\description{
GitBranch Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{remote}}{}

\item{\code{remote_name}}{}

\item{\code{error}}{}

\item{\code{message}}{}

\item{\code{owner_name}}{}

\item{\code{readonly}}{}

\item{\code{personal}}{}

\item{\code{is_local}}{}

\item{\code{is_remote}}{}

\item{\code{is_production}}{}

\item{\code{ahead_count}}{}

\item{\code{behind_count}}{}

\item{\code{commit_at}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
