% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitConnectionTestResult.r
\docType{data}
\name{GitConnectionTestResult}
\alias{GitConnectionTestResult}
\title{GitConnectionTestResult Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GitConnectionTestResult
}
\description{
GitConnectionTestResult Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{status}}{}

\item{\code{message}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
