% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitStatus.r
\docType{data}
\name{GitStatus}
\alias{GitStatus}
\title{GitStatus Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GitStatus
}
\description{
GitStatus Class
}
\section{Fields}{

\describe{
\item{\code{action}}{}

\item{\code{text}}{}

\item{\code{revertable}}{}

\item{\code{conflict}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
