% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group.r
\docType{data}
\name{Group}
\alias{Group}
\title{Group Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Group
}
\description{
Group Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{name}}{}

\item{\code{user_count}}{}

\item{\code{contains_current_user}}{}

\item{\code{externally_managed}}{}

\item{\code{include_by_default}}{}

\item{\code{external_group_id}}{}

\item{\code{can_add_to_content_metadata}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
