% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupApi.r
\docType{data}
\name{GroupApi}
\alias{GroupApi}
\title{Group operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GroupApi
}
\description{
looker.Group
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

add_group_group Add a Group to Group 


add_group_user Add a User to Group 


all_group_groups Get All Groups in Group 


all_group_users Get All Users in Group 


all_groups Get All Groups 


create_group Create Group 


delete_group Delete Group 


delete_group_from_group Deletes a Group from Group 


delete_group_user Remove a User from Group 


delete_user_attribute_group_value Delete User Attribute Group Value 


group Get Group 


update_group Update Group 


update_user_attribute_group_value Set User Attribute Group Value 

}
}

\keyword{datasets}
