% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupHierarchy.r
\name{GroupHierarchy}
\alias{GroupHierarchy}
\title{GroupHierarchy Class}
\description{
GroupHierarchy Class

GroupHierarchy Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{can}}{}

\item{\code{can_add_to_content_metadata}}{}

\item{\code{contains_current_user}}{}

\item{\code{external_group_id}}{}

\item{\code{externally_managed}}{}

\item{\code{id}}{}

\item{\code{include_by_default}}{}

\item{\code{name}}{}

\item{\code{user_count}}{}

\item{\code{parent_group_ids}}{}

\item{\code{role_ids}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GroupHierarchy$new()}}
\item \href{#method-toJSON}{\code{GroupHierarchy$toJSON()}}
\item \href{#method-fromJSONObject}{\code{GroupHierarchy$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{GroupHierarchy$fromJSON()}}
\item \href{#method-toJSONString}{\code{GroupHierarchy$toJSONString()}}
\item \href{#method-fromJSONString}{\code{GroupHierarchy$fromJSONString()}}
\item \href{#method-clone}{\code{GroupHierarchy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupHierarchy$new(
  can,
  can_add_to_content_metadata,
  contains_current_user,
  external_group_id,
  externally_managed,
  id,
  include_by_default,
  name,
  user_count,
  parent_group_ids,
  role_ids
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupHierarchy$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupHierarchy$fromJSONObject(GroupHierarchyJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupHierarchy$fromJSON(GroupHierarchyJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupHierarchy$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupHierarchy$fromJSONString(GroupHierarchyJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupHierarchy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
