% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupIdForGroupInclusion.r
\docType{data}
\name{GroupIdForGroupInclusion}
\alias{GroupIdForGroupInclusion}
\title{GroupIdForGroupInclusion Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GroupIdForGroupInclusion
}
\description{
GroupIdForGroupInclusion Class
}
\section{Fields}{

\describe{
\item{\code{group_id}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
