% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupIdForGroupUserInclusion.r
\docType{data}
\name{GroupIdForGroupUserInclusion}
\alias{GroupIdForGroupUserInclusion}
\title{GroupIdForGroupUserInclusion Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GroupIdForGroupUserInclusion
}
\description{
GroupIdForGroupUserInclusion Class
}
\section{Fields}{

\describe{
\item{\code{user_id}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
