% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomepageApi.r
\docType{data}
\name{HomepageApi}
\alias{HomepageApi}
\title{Homepage operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
HomepageApi
}
\description{
looker.Homepage
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

all_homepage_items Get All Homepage Items 


all_homepage_sections Get All Homepage sections 


create_homepage_item Create Homepage Item 


create_homepage_section Create Homepage section 


delete_homepage_item Delete Homepage Item 


delete_homepage_section Delete Homepage section 


homepage_item Get Homepage Item 


homepage_section Get Homepage section 


update_homepage_item Update Homepage Item 


update_homepage_section Update Homepage section 

}
}

\keyword{datasets}
