% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomepageItem.r
\docType{data}
\name{HomepageItem}
\alias{HomepageItem}
\title{HomepageItem Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
HomepageItem
}
\description{
HomepageItem Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{homepage_section_id}}{}

\item{\code{order}}{}

\item{\code{title}}{}

\item{\code{custom_title}}{}

\item{\code{use_custom_title}}{}

\item{\code{description}}{}

\item{\code{custom_description}}{}

\item{\code{use_custom_description}}{}

\item{\code{url}}{}

\item{\code{custom_url}}{}

\item{\code{use_custom_url}}{}

\item{\code{image_url}}{}

\item{\code{custom_image_url}}{}

\item{\code{custom_image_data_base64}}{}

\item{\code{use_custom_image}}{}

\item{\code{look_id}}{}

\item{\code{dashboard_id}}{}

\item{\code{lookml_dashboard_id}}{}

\item{\code{view_count}}{}

\item{\code{favorite_count}}{}

\item{\code{content_created_by}}{}

\item{\code{content_updated_at}}{}

\item{\code{content_favorite_id}}{}

\item{\code{content_metadata_id}}{}

\item{\code{section_fetch_time}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
