% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomepageSection.r
\name{HomepageSection}
\alias{HomepageSection}
\title{HomepageSection Class}
\description{
HomepageSection Class

HomepageSection Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{can}}{}

\item{\code{created_at}}{}

\item{\code{deleted_at}}{}

\item{\code{detail_url}}{}

\item{\code{homepage_id}}{}

\item{\code{homepage_items}}{}

\item{\code{id}}{}

\item{\code{is_header}}{}

\item{\code{item_order}}{}

\item{\code{title}}{}

\item{\code{updated_at}}{}

\item{\code{description}}{}

\item{\code{visible_item_order}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{HomepageSection$new()}}
\item \href{#method-toJSON}{\code{HomepageSection$toJSON()}}
\item \href{#method-fromJSONObject}{\code{HomepageSection$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{HomepageSection$fromJSON()}}
\item \href{#method-toJSONString}{\code{HomepageSection$toJSONString()}}
\item \href{#method-fromJSONString}{\code{HomepageSection$fromJSONString()}}
\item \href{#method-clone}{\code{HomepageSection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HomepageSection$new(
  can,
  created_at,
  deleted_at,
  detail_url,
  homepage_id,
  homepage_items,
  id,
  is_header,
  item_order,
  title,
  updated_at,
  description,
  visible_item_order
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HomepageSection$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HomepageSection$fromJSONObject(HomepageSectionJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HomepageSection$fromJSON(HomepageSectionJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HomepageSection$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HomepageSection$fromJSONString(HomepageSectionJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HomepageSection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
