% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Integration.r
\docType{data}
\name{Integration}
\alias{Integration}
\title{Integration Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Integration
}
\description{
Integration Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{integration_hub_id}}{}

\item{\code{label}}{}

\item{\code{description}}{}

\item{\code{enabled}}{}

\item{\code{params}}{}

\item{\code{supported_formats}}{}

\item{\code{supported_action_types}}{}

\item{\code{supported_formattings}}{}

\item{\code{supported_visualization_formattings}}{}

\item{\code{supported_download_settings}}{}

\item{\code{icon_url}}{}

\item{\code{required_fields}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
