% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrationApi.r
\name{IntegrationApi}
\alias{IntegrationApi}
\title{Integration operations}
\description{
looker.Integration
}
\section{Methods}{

\describe{

accept_integration_hub_legal_agreement Accept Integration Hub Legal Agreement


all_integration_hubs Get All Integration Hubs


all_integrations Get All Integrations


create_integration_hub Create Integration Hub


delete_integration_hub Delete Integration Hub


fetch_integration_form Fetch Remote Integration Form


integration Get Integration


integration_hub Get Integration Hub


test_integration Test integration


update_integration Update Integration


update_integration_hub Update Integration Hub

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{IntegrationApi$new()}}
\item \href{#method-accept_integration_hub_legal_agreement}{\code{IntegrationApi$accept_integration_hub_legal_agreement()}}
\item \href{#method-all_integration_hubs}{\code{IntegrationApi$all_integration_hubs()}}
\item \href{#method-all_integrations}{\code{IntegrationApi$all_integrations()}}
\item \href{#method-create_integration_hub}{\code{IntegrationApi$create_integration_hub()}}
\item \href{#method-delete_integration_hub}{\code{IntegrationApi$delete_integration_hub()}}
\item \href{#method-fetch_integration_form}{\code{IntegrationApi$fetch_integration_form()}}
\item \href{#method-integration}{\code{IntegrationApi$integration()}}
\item \href{#method-integration_hub}{\code{IntegrationApi$integration_hub()}}
\item \href{#method-test_integration}{\code{IntegrationApi$test_integration()}}
\item \href{#method-update_integration}{\code{IntegrationApi$update_integration()}}
\item \href{#method-update_integration_hub}{\code{IntegrationApi$update_integration_hub()}}
\item \href{#method-clone}{\code{IntegrationApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntegrationApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-accept_integration_hub_legal_agreement"></a>}}
\if{latex}{\out{\hypertarget{method-accept_integration_hub_legal_agreement}{}}}
\subsection{Method \code{accept_integration_hub_legal_agreement()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntegrationApi$accept_integration_hub_legal_agreement(integration_hub_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_integration_hubs"></a>}}
\if{latex}{\out{\hypertarget{method-all_integration_hubs}{}}}
\subsection{Method \code{all_integration_hubs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntegrationApi$all_integration_hubs(fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_integrations"></a>}}
\if{latex}{\out{\hypertarget{method-all_integrations}{}}}
\subsection{Method \code{all_integrations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntegrationApi$all_integrations(fields, integration_hub_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_integration_hub"></a>}}
\if{latex}{\out{\hypertarget{method-create_integration_hub}{}}}
\subsection{Method \code{create_integration_hub()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntegrationApi$create_integration_hub(body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_integration_hub"></a>}}
\if{latex}{\out{\hypertarget{method-delete_integration_hub}{}}}
\subsection{Method \code{delete_integration_hub()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntegrationApi$delete_integration_hub(integration_hub_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fetch_integration_form"></a>}}
\if{latex}{\out{\hypertarget{method-fetch_integration_form}{}}}
\subsection{Method \code{fetch_integration_form()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntegrationApi$fetch_integration_form(integration_id, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-integration"></a>}}
\if{latex}{\out{\hypertarget{method-integration}{}}}
\subsection{Method \code{integration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntegrationApi$integration(integration_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-integration_hub"></a>}}
\if{latex}{\out{\hypertarget{method-integration_hub}{}}}
\subsection{Method \code{integration_hub()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntegrationApi$integration_hub(integration_hub_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-test_integration"></a>}}
\if{latex}{\out{\hypertarget{method-test_integration}{}}}
\subsection{Method \code{test_integration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntegrationApi$test_integration(integration_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_integration"></a>}}
\if{latex}{\out{\hypertarget{method-update_integration}{}}}
\subsection{Method \code{update_integration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntegrationApi$update_integration(integration_id, body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_integration_hub"></a>}}
\if{latex}{\out{\hypertarget{method-update_integration_hub}{}}}
\subsection{Method \code{update_integration_hub()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntegrationApi$update_integration_hub(integration_hub_id, body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntegrationApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
