% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrationHub.r
\docType{data}
\name{IntegrationHub}
\alias{IntegrationHub}
\title{IntegrationHub Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
IntegrationHub
}
\description{
IntegrationHub Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{url}}{}

\item{\code{label}}{}

\item{\code{official}}{}

\item{\code{fetch_error_message}}{}

\item{\code{authorization_token}}{}

\item{\code{has_authorization_token}}{}

\item{\code{legal_agreement_signed}}{}

\item{\code{legal_agreement_required}}{}

\item{\code{legal_agreement_text}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
