% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrationParam.r
\docType{data}
\name{IntegrationParam}
\alias{IntegrationParam}
\title{IntegrationParam Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
IntegrationParam
}
\description{
IntegrationParam Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{label}}{}

\item{\code{description}}{}

\item{\code{required}}{}

\item{\code{has_value}}{}

\item{\code{value}}{}

\item{\code{user_attribute_name}}{}

\item{\code{sensitive}}{}
}}

\keyword{datasets}
