% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrationRequiredField.r
\docType{data}
\name{IntegrationRequiredField}
\alias{IntegrationRequiredField}
\title{IntegrationRequiredField Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
IntegrationRequiredField
}
\description{
IntegrationRequiredField Class
}
\section{Fields}{

\describe{
\item{\code{tag}}{}

\item{\code{any_tag}}{}

\item{\code{all_tags}}{}
}}

\keyword{datasets}
