% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalHelpResourcesContent.r
\name{InternalHelpResourcesContent}
\alias{InternalHelpResourcesContent}
\title{InternalHelpResourcesContent Class}
\description{
InternalHelpResourcesContent Class

InternalHelpResourcesContent Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{can}}{}

\item{\code{organization_name}}{}

\item{\code{markdown_content}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InternalHelpResourcesContent$new()}}
\item \href{#method-toJSON}{\code{InternalHelpResourcesContent$toJSON()}}
\item \href{#method-fromJSONObject}{\code{InternalHelpResourcesContent$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{InternalHelpResourcesContent$fromJSON()}}
\item \href{#method-toJSONString}{\code{InternalHelpResourcesContent$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InternalHelpResourcesContent$fromJSONString()}}
\item \href{#method-clone}{\code{InternalHelpResourcesContent$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternalHelpResourcesContent$new(can, organization_name, markdown_content)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternalHelpResourcesContent$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternalHelpResourcesContent$fromJSONObject(
  InternalHelpResourcesContentJsonObject
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternalHelpResourcesContent$fromJSON(InternalHelpResourcesContentJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternalHelpResourcesContent$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternalHelpResourcesContent$fromJSONString(InternalHelpResourcesContentJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternalHelpResourcesContent$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
