% Generated by roxygen2 (4.0.1): do not edit by hand
\name{InterpolateMissingFrames}
\alias{InterpolateMissingFrames}
\title{Interpolate missing gaze data}
\usage{
InterpolateMissingFrames(...)
}
\arguments{
\item{window}{the time window (in ms.) to interpolate over}
}
\value{
the list of Trials with missing data interpolated
}
\description{
Interpolate missing gaze data
}

