% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDAPConfig.r
\docType{data}
\name{LDAPConfig}
\alias{LDAPConfig}
\title{LDAPConfig Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LDAPConfig
}
\description{
LDAPConfig Class
}
\section{Fields}{

\describe{
\item{\code{enabled}}{}

\item{\code{connection_host}}{}

\item{\code{connection_port}}{}

\item{\code{connection_tls}}{}

\item{\code{auth_username}}{}

\item{\code{auth_password}}{}

\item{\code{has_auth_password}}{}

\item{\code{user_bind_base_dn}}{}

\item{\code{user_id_attribute_names}}{}

\item{\code{user_objectclass}}{}

\item{\code{user_custom_filter}}{}

\item{\code{user_attribute_map_email}}{}

\item{\code{user_attribute_map_first_name}}{}

\item{\code{user_attribute_map_last_name}}{}

\item{\code{user_attribute_map_ldap_id}}{}

\item{\code{merge_new_users_by_email}}{}

\item{\code{alternate_email_login_allowed}}{}

\item{\code{modified_at}}{}

\item{\code{modified_by}}{}

\item{\code{default_new_user_groups}}{}

\item{\code{default_new_user_roles}}{}

\item{\code{default_new_user_role_ids}}{}

\item{\code{default_new_user_group_ids}}{}

\item{\code{set_roles_from_groups}}{}

\item{\code{groups}}{}

\item{\code{groups_with_role_ids}}{}

\item{\code{auth_requires_role}}{}

\item{\code{groups_finder_type}}{}

\item{\code{groups_base_dn}}{}

\item{\code{groups_member_attribute}}{}

\item{\code{groups_user_attribute}}{}

\item{\code{groups_objectclasses}}{}

\item{\code{force_no_page}}{}

\item{\code{test_ldap_user}}{}

\item{\code{test_ldap_password}}{}

\item{\code{user_attributes}}{}

\item{\code{user_attributes_with_ids}}{}

\item{\code{url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
