% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDAPConfigTestResult.r
\docType{data}
\name{LDAPConfigTestResult}
\alias{LDAPConfigTestResult}
\title{LDAPConfigTestResult Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LDAPConfigTestResult
}
\description{
LDAPConfigTestResult Class
}
\section{Fields}{

\describe{
\item{\code{status}}{}

\item{\code{message}}{}

\item{\code{details}}{}

\item{\code{user}}{}

\item{\code{trace}}{}

\item{\code{url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
