% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDAPGroupRead.r
\docType{data}
\name{LDAPGroupRead}
\alias{LDAPGroupRead}
\title{LDAPGroupRead Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LDAPGroupRead
}
\description{
LDAPGroupRead Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{roles}}{}

\item{\code{url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
