% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDAPGroupWrite.r
\docType{data}
\name{LDAPGroupWrite}
\alias{LDAPGroupWrite}
\title{LDAPGroupWrite Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LDAPGroupWrite
}
\description{
LDAPGroupWrite Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{role_ids}}{}

\item{\code{url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
