% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDAPUser.r
\docType{data}
\name{LDAPUser}
\alias{LDAPUser}
\title{LDAPUser Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LDAPUser
}
\description{
LDAPUser Class
}
\section{Fields}{

\describe{
\item{\code{email}}{}

\item{\code{first_name}}{}

\item{\code{last_name}}{}

\item{\code{ldap_id}}{}

\item{\code{all_emails}}{}

\item{\code{ldap_dn}}{}

\item{\code{roles}}{}

\item{\code{groups}}{}

\item{\code{attributes}}{}

\item{\code{url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
