% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDAPUserAttributeRead.r
\docType{data}
\name{LDAPUserAttributeRead}
\alias{LDAPUserAttributeRead}
\title{LDAPUserAttributeRead Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LDAPUserAttributeRead
}
\description{
LDAPUserAttributeRead Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{required}}{}

\item{\code{user_attributes}}{}

\item{\code{url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
