% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDAPUserAttributeWrite.r
\docType{data}
\name{LDAPUserAttributeWrite}
\alias{LDAPUserAttributeWrite}
\title{LDAPUserAttributeWrite Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LDAPUserAttributeWrite
}
\description{
LDAPUserAttributeWrite Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{required}}{}

\item{\code{user_attribute_ids}}{}

\item{\code{url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
