% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LegacyFeature.r
\docType{data}
\name{LegacyFeature}
\alias{LegacyFeature}
\title{LegacyFeature Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LegacyFeature
}
\description{
LegacyFeature Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{name}}{}

\item{\code{description}}{}

\item{\code{enabled_locally}}{}

\item{\code{enabled}}{}

\item{\code{disallowed_as_of_version}}{}

\item{\code{disable_on_upgrade_to_version}}{}

\item{\code{end_of_life_version}}{}

\item{\code{documentation_url}}{}

\item{\code{approximate_disable_date}}{}

\item{\code{approximate_end_of_life_date}}{}

\item{\code{has_disabled_on_upgrade}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
