% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookApi.r
\docType{data}
\name{LookApi}
\alias{LookApi}
\title{Look operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LookApi
}
\description{
looker.Look
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

all_looks Get All Looks 


create_look Create Look 


delete_look Delete Look 


look Get Look 


run_look Run Look 


search_looks Search Looks 


update_look Update Look 

}
}

\keyword{datasets}
