% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookBasic.r
\docType{data}
\name{LookBasic}
\alias{LookBasic}
\title{LookBasic Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LookBasic
}
\description{
LookBasic Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{title}}{}

\item{\code{content_metadata_id}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
