% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookerSDK.R
\name{LookerSDK}
\alias{LookerSDK}
\title{LookerSDK}
\description{
Top-level container for current user session and API settings.
Uses API settings contained in a config file, \code{"looker.ini"} by default.
All of the endpoints that should be called by end-users are methods of this object.
Instantiate a new instance of this class with the \code{$new} method shown below.
}
\section{Usage}{

\preformatted{
sdk <- LookerSDK$new(configFile = "looker.ini", login = TRUE,
                     settings, userSession)

sdk$login()
sdk$isAuthenticated()
sdk$logout()
}
}

\section{Arguments}{

\describe{
  \item{configFile}{Path to configuration file containing API settings and credentials.}
  \item{login}{If \code{TRUE}, get an access token upon object creation.}
  \item{settings}{Optional, directly supply an \code{ApiSettings} object.}
  \item{userSession}{Optional, directly supply a \code{UserSession} object.}
}
}

\section{Authentication Methods}{

\describe{
  \item{\code{login()}}{Acquire an access token and authenticate the client.}
  \item{\code{logout()}}{Revoke current access token.}
  \item{\code{isAuthenticated()}}{Returns \code{TRUE} when an unexpired token is being used.}
  \item{\code{activeToken()}}{Returns a list representing the current active token.}
  \item{\code{refresh()}}{If current active token is missing or expired, acquires a new token.}
  \item{\code{me()}}{Returns information about the current authenticated user.}
}
}

\section{Look Endpoints}{

\describe{
  \item{\code{runLook(lookId)}}{Returns the query result for the Look specified in \code{lookId}.}
}
}

\section{Dashboard Endpoints}{

\describe{
  \item{\code{allDashboards()}}{Returns a list of all dashboards in the instance.}
}
}

\section{Query Endpoints}{

\describe{
  \item{\code{runInlineQuery(model, view, fields, filters, sorts, limit, queryTimezone)}}{Provided with mandatory query parameters (\code{model, view, fields}) plus optional parameters (\code{filters, sorts, limit, queryTimezone}), returns the results of the defined query.}
}
}

\section{LookML Project Endpoints}{

\describe{
  \item{\code{allProjects()}}{Returns a list of all projects in the instance.}
}
}

\section{LookML Model Endpoints}{

\describe{
  \item{\code{allModels()}}{Return a list of all models in the instance.}
}
}

\examples{
\dontrun{
# You can specify the configFile path directly
sdk <- lookr::LookerSDK$new(configFile = "path/to/file")

# Or use the looker.ini file in your working directory
sdk <- lookr::LookerSDK$new()
}
}
