% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookmlModel.r
\docType{data}
\name{LookmlModel}
\alias{LookmlModel}
\title{LookmlModel Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LookmlModel
}
\description{
LookmlModel Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{project_name}}{}

\item{\code{allowed_db_connection_names}}{}

\item{\code{unlimited_db_connections}}{}

\item{\code{has_content}}{}

\item{\code{label}}{}

\item{\code{explores}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
