% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookmlModelApi.r
\name{LookmlModelApi}
\alias{LookmlModelApi}
\title{LookmlModel operations}
\description{
looker.LookmlModel
}
\section{Methods}{

\describe{

all_lookml_models Get All LookML Models 


create_lookml_model Create LookML Model 


delete_lookml_model Delete LookML Model 


lookml_model Get LookML Model 


lookml_model_explore Get LookML Model Explore 


update_lookml_model Update LookML Model 

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LookmlModelApi$new()}}
\item \href{#method-all_lookml_models}{\code{LookmlModelApi$all_lookml_models()}}
\item \href{#method-create_lookml_model}{\code{LookmlModelApi$create_lookml_model()}}
\item \href{#method-delete_lookml_model}{\code{LookmlModelApi$delete_lookml_model()}}
\item \href{#method-lookml_model}{\code{LookmlModelApi$lookml_model()}}
\item \href{#method-lookml_model_explore}{\code{LookmlModelApi$lookml_model_explore()}}
\item \href{#method-update_lookml_model}{\code{LookmlModelApi$update_lookml_model()}}
\item \href{#method-clone}{\code{LookmlModelApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_lookml_models"></a>}}
\if{latex}{\out{\hypertarget{method-all_lookml_models}{}}}
\subsection{Method \code{all_lookml_models()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelApi$all_lookml_models(fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_lookml_model"></a>}}
\if{latex}{\out{\hypertarget{method-create_lookml_model}{}}}
\subsection{Method \code{create_lookml_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelApi$create_lookml_model(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_lookml_model"></a>}}
\if{latex}{\out{\hypertarget{method-delete_lookml_model}{}}}
\subsection{Method \code{delete_lookml_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelApi$delete_lookml_model(lookml_model_name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lookml_model"></a>}}
\if{latex}{\out{\hypertarget{method-lookml_model}{}}}
\subsection{Method \code{lookml_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelApi$lookml_model(lookml_model_name, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lookml_model_explore"></a>}}
\if{latex}{\out{\hypertarget{method-lookml_model_explore}{}}}
\subsection{Method \code{lookml_model_explore()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelApi$lookml_model_explore(
  lookml_model_name,
  explore_name,
  fields,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_lookml_model"></a>}}
\if{latex}{\out{\hypertarget{method-update_lookml_model}{}}}
\subsection{Method \code{update_lookml_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelApi$update_lookml_model(lookml_model_name, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
