% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookmlModelExplore.r
\docType{data}
\name{LookmlModelExplore}
\alias{LookmlModelExplore}
\title{LookmlModelExplore Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LookmlModelExplore
}
\description{
LookmlModelExplore Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{name}}{}

\item{\code{description}}{}

\item{\code{label}}{}

\item{\code{scopes}}{}

\item{\code{can_total}}{}

\item{\code{can_save}}{}

\item{\code{can_explain}}{}

\item{\code{can_pivot_in_db}}{}

\item{\code{has_timezone_support}}{}

\item{\code{supports_cost_estimate}}{}

\item{\code{connection_name}}{}

\item{\code{null_sort_treatment}}{}

\item{\code{files}}{}

\item{\code{source_file}}{}

\item{\code{project_name}}{}

\item{\code{model_name}}{}

\item{\code{view_name}}{}

\item{\code{hidden}}{}

\item{\code{sql_table_name}}{}

\item{\code{access_filter_fields}}{}

\item{\code{access_filters}}{}

\item{\code{aliases}}{}

\item{\code{always_filter}}{}

\item{\code{conditionally_filter}}{}

\item{\code{index_fields}}{}

\item{\code{sets}}{}

\item{\code{errors}}{}

\item{\code{fields}}{}

\item{\code{joins}}{}

\item{\code{group_label}}{}
}}

\keyword{datasets}
