% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookmlModelExploreAccessFilter.r
\docType{data}
\name{LookmlModelExploreAccessFilter}
\alias{LookmlModelExploreAccessFilter}
\title{LookmlModelExploreAccessFilter Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LookmlModelExploreAccessFilter
}
\description{
LookmlModelExploreAccessFilter Class
}
\section{Fields}{

\describe{
\item{\code{field}}{}

\item{\code{user_attribute}}{}
}}

\keyword{datasets}
