% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookmlModelExploreAlias.r
\docType{data}
\name{LookmlModelExploreAlias}
\alias{LookmlModelExploreAlias}
\title{LookmlModelExploreAlias Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LookmlModelExploreAlias
}
\description{
LookmlModelExploreAlias Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{value}}{}
}}

\keyword{datasets}
