% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookmlModelExploreAlwaysFilter.r
\docType{data}
\name{LookmlModelExploreAlwaysFilter}
\alias{LookmlModelExploreAlwaysFilter}
\title{LookmlModelExploreAlwaysFilter Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LookmlModelExploreAlwaysFilter
}
\description{
LookmlModelExploreAlwaysFilter Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{value}}{}
}}

\keyword{datasets}
