% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookmlModelExploreConditionallyFilter.r
\docType{data}
\name{LookmlModelExploreConditionallyFilter}
\alias{LookmlModelExploreConditionallyFilter}
\title{LookmlModelExploreConditionallyFilter Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LookmlModelExploreConditionallyFilter
}
\description{
LookmlModelExploreConditionallyFilter Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{value}}{}
}}

\keyword{datasets}
