% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookmlModelExploreError.r
\docType{data}
\name{LookmlModelExploreError}
\alias{LookmlModelExploreError}
\title{LookmlModelExploreError Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LookmlModelExploreError
}
\description{
LookmlModelExploreError Class
}
\section{Fields}{

\describe{
\item{\code{message}}{}

\item{\code{details}}{}

\item{\code{error_pos}}{}

\item{\code{field_error}}{}
}}

\keyword{datasets}
