% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookmlModelExploreField.r
\name{LookmlModelExploreField}
\alias{LookmlModelExploreField}
\title{LookmlModelExploreField Class}
\description{
LookmlModelExploreField Class

LookmlModelExploreField Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{align}}{}

\item{\code{can_filter}}{}

\item{\code{category}}{}

\item{\code{default_filter_value}}{}

\item{\code{description}}{}

\item{\code{dimension_group}}{}

\item{\code{enumerations}}{}

\item{\code{error}}{}

\item{\code{field_group_label}}{}

\item{\code{field_group_variant}}{}

\item{\code{fill_style}}{}

\item{\code{fiscal_month_offset}}{}

\item{\code{has_allowed_values}}{}

\item{\code{hidden}}{}

\item{\code{is_filter}}{}

\item{\code{is_fiscal}}{}

\item{\code{is_numeric}}{}

\item{\code{is_timeframe}}{}

\item{\code{can_time_filter}}{}

\item{\code{time_interval}}{}

\item{\code{label}}{}

\item{\code{label_from_parameter}}{}

\item{\code{label_short}}{}

\item{\code{lookml_link}}{}

\item{\code{map_layer}}{}

\item{\code{measure}}{}

\item{\code{name}}{}

\item{\code{strict_value_format}}{}

\item{\code{parameter}}{}

\item{\code{permanent}}{}

\item{\code{primary_key}}{}

\item{\code{project_name}}{}

\item{\code{requires_refresh_on_sort}}{}

\item{\code{scope}}{}

\item{\code{sortable}}{}

\item{\code{source_file}}{}

\item{\code{source_file_path}}{}

\item{\code{sql}}{}

\item{\code{sql_case}}{}

\item{\code{filters}}{}

\item{\code{suggest_dimension}}{}

\item{\code{suggest_explore}}{}

\item{\code{suggestable}}{}

\item{\code{suggestions}}{}

\item{\code{tags}}{}

\item{\code{type}}{}

\item{\code{user_attribute_filter_types}}{}

\item{\code{value_format}}{}

\item{\code{view}}{}

\item{\code{view_label}}{}

\item{\code{dynamic}}{}

\item{\code{week_start_day}}{}

\item{\code{times_used}}{}

\item{\code{original_view}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LookmlModelExploreField$new()}}
\item \href{#method-toJSON}{\code{LookmlModelExploreField$toJSON()}}
\item \href{#method-fromJSONObject}{\code{LookmlModelExploreField$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{LookmlModelExploreField$fromJSON()}}
\item \href{#method-toJSONString}{\code{LookmlModelExploreField$toJSONString()}}
\item \href{#method-fromJSONString}{\code{LookmlModelExploreField$fromJSONString()}}
\item \href{#method-clone}{\code{LookmlModelExploreField$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreField$new(
  align,
  can_filter,
  category,
  default_filter_value,
  description,
  dimension_group,
  enumerations,
  error,
  field_group_label,
  field_group_variant,
  fill_style,
  fiscal_month_offset,
  has_allowed_values,
  hidden,
  is_filter,
  is_fiscal,
  is_numeric,
  is_timeframe,
  can_time_filter,
  time_interval,
  label,
  label_from_parameter,
  label_short,
  lookml_link,
  map_layer,
  measure,
  name,
  strict_value_format,
  parameter,
  permanent,
  primary_key,
  project_name,
  requires_refresh_on_sort,
  scope,
  sortable,
  source_file,
  source_file_path,
  sql,
  sql_case,
  filters,
  suggest_dimension,
  suggest_explore,
  suggestable,
  suggestions,
  tags,
  type,
  user_attribute_filter_types,
  value_format,
  view,
  view_label,
  dynamic,
  week_start_day,
  times_used,
  original_view
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreField$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreField$fromJSONObject(LookmlModelExploreFieldJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreField$fromJSON(LookmlModelExploreFieldJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreField$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreField$fromJSONString(LookmlModelExploreFieldJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreField$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
