% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookmlModelExploreFieldEnumeration.r
\docType{data}
\name{LookmlModelExploreFieldEnumeration}
\alias{LookmlModelExploreFieldEnumeration}
\title{LookmlModelExploreFieldEnumeration Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LookmlModelExploreFieldEnumeration
}
\description{
LookmlModelExploreFieldEnumeration Class
}
\section{Fields}{

\describe{
\item{\code{label}}{}

\item{\code{value}}{}
}}

\keyword{datasets}
