% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookmlModelExploreFieldMapLayer.r
\docType{data}
\name{LookmlModelExploreFieldMapLayer}
\alias{LookmlModelExploreFieldMapLayer}
\title{LookmlModelExploreFieldMapLayer Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LookmlModelExploreFieldMapLayer
}
\description{
LookmlModelExploreFieldMapLayer Class
}
\section{Fields}{

\describe{
\item{\code{url}}{}

\item{\code{feature_key}}{}

\item{\code{property_key}}{}

\item{\code{property_label_key}}{}

\item{\code{projection}}{}

\item{\code{format}}{}

\item{\code{extents_json_url}}{}

\item{\code{max_zoom_level}}{}

\item{\code{min_zoom_level}}{}
}}

\keyword{datasets}
