% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookmlModelExploreFieldMeasureFilters.r
\name{LookmlModelExploreFieldMeasureFilters}
\alias{LookmlModelExploreFieldMeasureFilters}
\title{LookmlModelExploreFieldMeasureFilters Class}
\description{
LookmlModelExploreFieldMeasureFilters Class

LookmlModelExploreFieldMeasureFilters Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{field}}{}

\item{\code{condition}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LookmlModelExploreFieldMeasureFilters$new()}}
\item \href{#method-toJSON}{\code{LookmlModelExploreFieldMeasureFilters$toJSON()}}
\item \href{#method-fromJSONObject}{\code{LookmlModelExploreFieldMeasureFilters$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{LookmlModelExploreFieldMeasureFilters$fromJSON()}}
\item \href{#method-toJSONString}{\code{LookmlModelExploreFieldMeasureFilters$toJSONString()}}
\item \href{#method-fromJSONString}{\code{LookmlModelExploreFieldMeasureFilters$fromJSONString()}}
\item \href{#method-clone}{\code{LookmlModelExploreFieldMeasureFilters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreFieldMeasureFilters$new(field, condition)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreFieldMeasureFilters$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreFieldMeasureFilters$fromJSONObject(
  LookmlModelExploreFieldMeasureFiltersJsonObject
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreFieldMeasureFilters$fromJSON(
  LookmlModelExploreFieldMeasureFiltersJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreFieldMeasureFilters$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreFieldMeasureFilters$fromJSONString(
  LookmlModelExploreFieldMeasureFiltersJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreFieldMeasureFilters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
