% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookmlModelExploreFieldset.r
\docType{data}
\name{LookmlModelExploreFieldset}
\alias{LookmlModelExploreFieldset}
\title{LookmlModelExploreFieldset Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LookmlModelExploreFieldset
}
\description{
LookmlModelExploreFieldset Class
}
\section{Fields}{

\describe{
\item{\code{dimensions}}{}

\item{\code{measures}}{}

\item{\code{filters}}{}

\item{\code{parameters}}{}
}}

\keyword{datasets}
