% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookmlModelExploreFieldSqlCase.r
\docType{data}
\name{LookmlModelExploreFieldSqlCase}
\alias{LookmlModelExploreFieldSqlCase}
\title{LookmlModelExploreFieldSqlCase Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LookmlModelExploreFieldSqlCase
}
\description{
LookmlModelExploreFieldSqlCase Class
}
\section{Fields}{

\describe{
\item{\code{value}}{}

\item{\code{condition}}{}
}}

\keyword{datasets}
