% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookmlModelExploreFieldTimeInterval.r
\name{LookmlModelExploreFieldTimeInterval}
\alias{LookmlModelExploreFieldTimeInterval}
\title{LookmlModelExploreFieldTimeInterval Class}
\description{
LookmlModelExploreFieldTimeInterval Class

LookmlModelExploreFieldTimeInterval Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{}

\item{\code{count}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LookmlModelExploreFieldTimeInterval$new()}}
\item \href{#method-toJSON}{\code{LookmlModelExploreFieldTimeInterval$toJSON()}}
\item \href{#method-fromJSONObject}{\code{LookmlModelExploreFieldTimeInterval$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{LookmlModelExploreFieldTimeInterval$fromJSON()}}
\item \href{#method-toJSONString}{\code{LookmlModelExploreFieldTimeInterval$toJSONString()}}
\item \href{#method-fromJSONString}{\code{LookmlModelExploreFieldTimeInterval$fromJSONString()}}
\item \href{#method-clone}{\code{LookmlModelExploreFieldTimeInterval$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreFieldTimeInterval$new(name, count)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreFieldTimeInterval$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreFieldTimeInterval$fromJSONObject(
  LookmlModelExploreFieldTimeIntervalJsonObject
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreFieldTimeInterval$fromJSON(
  LookmlModelExploreFieldTimeIntervalJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreFieldTimeInterval$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreFieldTimeInterval$fromJSONString(
  LookmlModelExploreFieldTimeIntervalJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreFieldTimeInterval$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
