% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookmlModelExploreJoins.r
\docType{data}
\name{LookmlModelExploreJoins}
\alias{LookmlModelExploreJoins}
\title{LookmlModelExploreJoins Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LookmlModelExploreJoins
}
\description{
LookmlModelExploreJoins Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{dependent_fields}}{}

\item{\code{fields}}{}

\item{\code{foreign_key}}{}

\item{\code{from}}{}

\item{\code{outer_only}}{}

\item{\code{relationship}}{}

\item{\code{required_joins}}{}

\item{\code{sql_foreign_key}}{}

\item{\code{sql_on}}{}

\item{\code{sql_table_name}}{}

\item{\code{type}}{}

\item{\code{view_label}}{}
}}

\keyword{datasets}
