% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookmlModelExploreSet.r
\docType{data}
\name{LookmlModelExploreSet}
\alias{LookmlModelExploreSet}
\title{LookmlModelExploreSet Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LookmlModelExploreSet
}
\description{
LookmlModelExploreSet Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{value}}{}
}}

\keyword{datasets}
