% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookmlModelExploreSupportedMeasureType.r
\name{LookmlModelExploreSupportedMeasureType}
\alias{LookmlModelExploreSupportedMeasureType}
\title{LookmlModelExploreSupportedMeasureType Class}
\description{
LookmlModelExploreSupportedMeasureType Class

LookmlModelExploreSupportedMeasureType Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dimension_type}}{}

\item{\code{measure_types}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LookmlModelExploreSupportedMeasureType$new()}}
\item \href{#method-toJSON}{\code{LookmlModelExploreSupportedMeasureType$toJSON()}}
\item \href{#method-fromJSONObject}{\code{LookmlModelExploreSupportedMeasureType$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{LookmlModelExploreSupportedMeasureType$fromJSON()}}
\item \href{#method-toJSONString}{\code{LookmlModelExploreSupportedMeasureType$toJSONString()}}
\item \href{#method-fromJSONString}{\code{LookmlModelExploreSupportedMeasureType$fromJSONString()}}
\item \href{#method-clone}{\code{LookmlModelExploreSupportedMeasureType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreSupportedMeasureType$new(dimension_type, measure_types)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreSupportedMeasureType$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreSupportedMeasureType$fromJSONObject(
  LookmlModelExploreSupportedMeasureTypeJsonObject
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreSupportedMeasureType$fromJSON(
  LookmlModelExploreSupportedMeasureTypeJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreSupportedMeasureType$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreSupportedMeasureType$fromJSONString(
  LookmlModelExploreSupportedMeasureTypeJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreSupportedMeasureType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
