% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookModel.r
\docType{data}
\name{LookModel}
\alias{LookModel}
\title{LookModel Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LookModel
}
\description{
LookModel Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{label}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
