% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{lookr_classes}
\alias{lookr_classes}
\alias{as.Task}
\alias{as.Session}
\alias{as.Block}
\alias{as.TrialList}
\alias{as.Trial}
\alias{as.Gazedata}
\alias{is.Task}
\alias{is.Session}
\alias{is.Block}
\alias{is.TrialList}
\alias{is.Trial}
\title{lookr_classes}
\usage{
as.Task(x)

as.Session(x)

as.Block(x)

as.TrialList(x)

as.Trial(x)

as.Gazedata(x)

is.Task(x)

is.Session(x)

is.Block(x)

is.TrialList(x)

is.Trial(x)
}
\arguments{
\item{x}{an object to coerce to another class (as.Class function) or an
object whose class we are checking (is.Class) function}
}
\description{
lookr_classes
}
