% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{lwl_constants}
\alias{lwl_constants}
\title{Package-wide constants}
\format{An object of class \code{list} of length 6.}
\usage{
lwl_constants
}
\description{
Package-wide constants
}
\keyword{datasets}

