% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\docType{data}
\name{lwl_opts}
\alias{lwl_opts}
\title{Default analysis options}
\format{An OptionList object. \Sexpr[results=verbatim]{str(lookr::lwl_opts)}}
\usage{
lwl_opts
}
\description{
Default analysis options
}
\details{
Below is a list of default package options, retrieved via
\code{lwl_opts$get()}: \Sexpr[results=verbatim]{str(lookr::lwl_opts$get())}
}
\keyword{datasets}
