% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{MakeAttributeFilter}
\alias{MakeAttributeFilter}
\title{Make a function to filter out trials based on attribute values}
\usage{
MakeAttributeFilter(attr_name)
}
\arguments{
\item{attr_name}{the name of attribute to filter by}
}
\value{
a function \code{f(trials, attr_values)} that filters out trials with
  the specified \code{attr_values} for the given \code{attr_name}.
}
\description{
This is a higher-order function.
}
\details{
The function returns a higher order function with the \code{attr_name} fixed.
Details about the returned function:
\describe{
  \item{Usage}{\code{MakeAttributeFilter(attr_name)(trials, attr_value)}}
  \item{\code{trials}}{trials a list of Trial objects}
  \item{\code{attr_value}}{a vector of values to filter out}
}
}
