% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MakeTrialGrabber}
\alias{MakeTrialGrabber}
\title{Create a function that grabs all trials with a certain attribute}
\usage{
MakeTrialGrabber(attr_name, reversed = FALSE)
}
\arguments{
\item{attr_name}{a string naming the attribute by which to select trials}

\item{reversed}{a boolean that specifies whether the parameters in the
returns function should be reversed. Default value is `FALSE`, so the
function will have the argument order `function(trials, attr_value)`.
This is useful for `apply`-like functions.}
}
\value{
A two-argument function that will grab all `Trial` objects (`trials`)
  with a certain attribute value (`attr_value`)
}
\description{
Create a function that grabs all trials with a certain attribute
}
\examples{
# Define a function that selects trials based on the StimType attribute
GetTrialsByStimType <- MakeTrialGrabber("StimType")
# Grab all trials with a StimType value of "neutral"
neutrals <- GetTrialsByStimType(trials, "neutral")

# Make function to select trials based on TargetWord. Grab all "duck" trials.
GetTrialsByTarget <- MakeTrialGrabber("TargetWord")
duck_trials <- GetTrialsByTarget(trials, "duck")
}

