% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melting.R
\name{MeltLooks}
\alias{MeltLooks}
\title{Convert Trial and list of Trials into a long data-frame}
\usage{
MeltLooks(x, other_cols, other_attrs)
}
\arguments{
\item{x}{a Trial or a TrialList object}

\item{other_cols}{additional data-frame columns to extract from each Trial}

\item{other_attrs}{additional attributes to pull from each Trial}

\item{...}{Additional arguments passed onto S3 methods. Currently ignored.}
}
\value{
a single data-frame containing the Time and GazeByImageAOI data from
  the Trial(s) with columns for other attribute values of the Trial(s)
}
\description{
Convert Trial and list of Trials into a long data-frame
}
