% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataApi.r
\name{MetadataApi}
\alias{MetadataApi}
\title{Metadata operations}
\description{
looker.Metadata
}
\section{Methods}{

\describe{

connection_columns Get columns for a connection


connection_cost_estimate Estimate costs for a connection


connection_databases List accessible databases to this connection


connection_features Metadata features supported by this connection


connection_schemas Get schemas for a connection


connection_search_columns Search a connection for columns


connection_tables Get tables for a connection


get_model Get a single model


model_fieldname_suggestions Model field name suggestions

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MetadataApi$new()}}
\item \href{#method-connection_columns}{\code{MetadataApi$connection_columns()}}
\item \href{#method-connection_cost_estimate}{\code{MetadataApi$connection_cost_estimate()}}
\item \href{#method-connection_databases}{\code{MetadataApi$connection_databases()}}
\item \href{#method-connection_features}{\code{MetadataApi$connection_features()}}
\item \href{#method-connection_schemas}{\code{MetadataApi$connection_schemas()}}
\item \href{#method-connection_search_columns}{\code{MetadataApi$connection_search_columns()}}
\item \href{#method-connection_tables}{\code{MetadataApi$connection_tables()}}
\item \href{#method-get_model}{\code{MetadataApi$get_model()}}
\item \href{#method-model_fieldname_suggestions}{\code{MetadataApi$model_fieldname_suggestions()}}
\item \href{#method-clone}{\code{MetadataApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connection_columns"></a>}}
\if{latex}{\out{\hypertarget{method-connection_columns}{}}}
\subsection{Method \code{connection_columns()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataApi$connection_columns(
  connection_name,
  database,
  schema_name,
  cache,
  table_limit,
  table_names,
  fields,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connection_cost_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-connection_cost_estimate}{}}}
\subsection{Method \code{connection_cost_estimate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataApi$connection_cost_estimate(connection_name, body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connection_databases"></a>}}
\if{latex}{\out{\hypertarget{method-connection_databases}{}}}
\subsection{Method \code{connection_databases()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataApi$connection_databases(connection_name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connection_features"></a>}}
\if{latex}{\out{\hypertarget{method-connection_features}{}}}
\subsection{Method \code{connection_features()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataApi$connection_features(connection_name, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connection_schemas"></a>}}
\if{latex}{\out{\hypertarget{method-connection_schemas}{}}}
\subsection{Method \code{connection_schemas()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataApi$connection_schemas(connection_name, database, cache, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connection_search_columns"></a>}}
\if{latex}{\out{\hypertarget{method-connection_search_columns}{}}}
\subsection{Method \code{connection_search_columns()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataApi$connection_search_columns(
  connection_name,
  column_name,
  fields,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connection_tables"></a>}}
\if{latex}{\out{\hypertarget{method-connection_tables}{}}}
\subsection{Method \code{connection_tables()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataApi$connection_tables(
  connection_name,
  database,
  schema_name,
  cache,
  fields,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_model"></a>}}
\if{latex}{\out{\hypertarget{method-get_model}{}}}
\subsection{Method \code{get_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataApi$get_model(model_name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model_fieldname_suggestions"></a>}}
\if{latex}{\out{\hypertarget{method-model_fieldname_suggestions}{}}}
\subsection{Method \code{model_fieldname_suggestions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataApi$model_fieldname_suggestions(
  model_name,
  view_name,
  field_name,
  term,
  filters,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
