% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFieldSuggestions.r
\name{ModelFieldSuggestions}
\alias{ModelFieldSuggestions}
\title{ModelFieldSuggestions Class}
\description{
ModelFieldSuggestions Class

ModelFieldSuggestions Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{suggestions}}{}

\item{\code{error}}{}

\item{\code{from_cache}}{}

\item{\code{hit_limit}}{}

\item{\code{used_calcite_materialization}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ModelFieldSuggestions$new()}}
\item \href{#method-toJSON}{\code{ModelFieldSuggestions$toJSON()}}
\item \href{#method-fromJSONObject}{\code{ModelFieldSuggestions$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{ModelFieldSuggestions$fromJSON()}}
\item \href{#method-toJSONString}{\code{ModelFieldSuggestions$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ModelFieldSuggestions$fromJSONString()}}
\item \href{#method-clone}{\code{ModelFieldSuggestions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelFieldSuggestions$new(
  suggestions,
  error,
  from_cache,
  hit_limit,
  used_calcite_materialization
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelFieldSuggestions$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelFieldSuggestions$fromJSONObject(ModelFieldSuggestionsJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelFieldSuggestions$fromJSON(ModelFieldSuggestionsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelFieldSuggestions$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelFieldSuggestions$fromJSONString(ModelFieldSuggestionsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelFieldSuggestions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
