% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelsNotValidated.r
\docType{data}
\name{ModelsNotValidated}
\alias{ModelsNotValidated}
\title{ModelsNotValidated Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ModelsNotValidated
}
\description{
ModelsNotValidated Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{project_file_id}}{}
}}

\keyword{datasets}
