% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OauthClientApp.r
\name{OauthClientApp}
\alias{OauthClientApp}
\title{OauthClientApp Class}
\description{
OauthClientApp Class

OauthClientApp Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{can}}{}

\item{\code{client_guid}}{}

\item{\code{redirect_uri}}{}

\item{\code{display_name}}{}

\item{\code{description}}{}

\item{\code{enabled}}{}

\item{\code{group_id}}{}

\item{\code{tokens_invalid_before}}{}

\item{\code{activated_users}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OauthClientApp$new()}}
\item \href{#method-toJSON}{\code{OauthClientApp$toJSON()}}
\item \href{#method-fromJSONObject}{\code{OauthClientApp$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{OauthClientApp$fromJSON()}}
\item \href{#method-toJSONString}{\code{OauthClientApp$toJSONString()}}
\item \href{#method-fromJSONString}{\code{OauthClientApp$fromJSONString()}}
\item \href{#method-clone}{\code{OauthClientApp$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OauthClientApp$new(
  can,
  client_guid,
  redirect_uri,
  display_name,
  description,
  enabled,
  group_id,
  tokens_invalid_before,
  activated_users
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OauthClientApp$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OauthClientApp$fromJSONObject(OauthClientAppJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OauthClientApp$fromJSON(OauthClientAppJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OauthClientApp$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OauthClientApp$fromJSONString(OauthClientAppJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OauthClientApp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
