% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OIDCConfig.r
\name{OIDCConfig}
\alias{OIDCConfig}
\title{OIDCConfig Class}
\description{
OIDCConfig Class

OIDCConfig Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{can}}{}

\item{\code{alternate_email_login_allowed}}{}

\item{\code{audience}}{}

\item{\code{auth_requires_role}}{}

\item{\code{authorization_endpoint}}{}

\item{\code{default_new_user_group_ids}}{}

\item{\code{default_new_user_groups}}{}

\item{\code{default_new_user_role_ids}}{}

\item{\code{default_new_user_roles}}{}

\item{\code{enabled}}{}

\item{\code{groups}}{}

\item{\code{groups_attribute}}{}

\item{\code{groups_with_role_ids}}{}

\item{\code{identifier}}{}

\item{\code{issuer}}{}

\item{\code{modified_at}}{}

\item{\code{modified_by}}{}

\item{\code{new_user_migration_types}}{}

\item{\code{scopes}}{}

\item{\code{secret}}{}

\item{\code{set_roles_from_groups}}{}

\item{\code{test_slug}}{}

\item{\code{token_endpoint}}{}

\item{\code{user_attribute_map_email}}{}

\item{\code{user_attribute_map_first_name}}{}

\item{\code{user_attribute_map_last_name}}{}

\item{\code{user_attributes}}{}

\item{\code{user_attributes_with_ids}}{}

\item{\code{userinfo_endpoint}}{}

\item{\code{allow_normal_group_membership}}{}

\item{\code{allow_roles_from_normal_groups}}{}

\item{\code{allow_direct_roles}}{}

\item{\code{url}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OIDCConfig$new()}}
\item \href{#method-toJSON}{\code{OIDCConfig$toJSON()}}
\item \href{#method-fromJSONObject}{\code{OIDCConfig$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{OIDCConfig$fromJSON()}}
\item \href{#method-toJSONString}{\code{OIDCConfig$toJSONString()}}
\item \href{#method-fromJSONString}{\code{OIDCConfig$fromJSONString()}}
\item \href{#method-clone}{\code{OIDCConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OIDCConfig$new(
  can,
  alternate_email_login_allowed,
  audience,
  auth_requires_role,
  authorization_endpoint,
  default_new_user_group_ids,
  default_new_user_groups,
  default_new_user_role_ids,
  default_new_user_roles,
  enabled,
  groups,
  groups_attribute,
  groups_with_role_ids,
  identifier,
  issuer,
  modified_at,
  modified_by,
  new_user_migration_types,
  scopes,
  secret,
  set_roles_from_groups,
  test_slug,
  token_endpoint,
  user_attribute_map_email,
  user_attribute_map_first_name,
  user_attribute_map_last_name,
  user_attributes,
  user_attributes_with_ids,
  userinfo_endpoint,
  allow_normal_group_membership,
  allow_roles_from_normal_groups,
  allow_direct_roles,
  url
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OIDCConfig$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OIDCConfig$fromJSONObject(OIDCConfigJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OIDCConfig$fromJSON(OIDCConfigJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OIDCConfig$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OIDCConfig$fromJSONString(OIDCConfigJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OIDCConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
