% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OIDCGroupRead.r
\docType{data}
\name{OIDCGroupRead}
\alias{OIDCGroupRead}
\title{OIDCGroupRead Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
OIDCGroupRead
}
\description{
OIDCGroupRead Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{roles}}{}

\item{\code{url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
