% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OIDCGroupWrite.r
\docType{data}
\name{OIDCGroupWrite}
\alias{OIDCGroupWrite}
\title{OIDCGroupWrite Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
OIDCGroupWrite
}
\description{
OIDCGroupWrite Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{role_ids}}{}

\item{\code{url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
