% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OIDCUserAttributeRead.r
\docType{data}
\name{OIDCUserAttributeRead}
\alias{OIDCUserAttributeRead}
\title{OIDCUserAttributeRead Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
OIDCUserAttributeRead
}
\description{
OIDCUserAttributeRead Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{required}}{}

\item{\code{user_attributes}}{}

\item{\code{url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
