% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OIDCUserAttributeWrite.r
\docType{data}
\name{OIDCUserAttributeWrite}
\alias{OIDCUserAttributeWrite}
\title{OIDCUserAttributeWrite Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
OIDCUserAttributeWrite
}
\description{
OIDCUserAttributeWrite Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{required}}{}

\item{\code{user_attribute_ids}}{}

\item{\code{url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
