% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gca.R
\name{orthogonal_time}
\alias{orthogonal_time}
\title{Compute orthogonal times}
\usage{
orthogonal_time(times, degree)
}
\arguments{
\item{times}{a vector of time values}

\item{degree}{of the desired polynomial}
}
\value{
a data-frame with original time values and an \code{ot} column for
  each polynomial degree
}
\description{
Compute orthogonal times
}
