% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Permission.r
\docType{data}
\name{Permission}
\alias{Permission}
\title{Permission Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Permission
}
\description{
Permission Class
}
\section{Fields}{

\describe{
\item{\code{permission}}{}

\item{\code{parent}}{}

\item{\code{description}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
