% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermissionSet.r
\docType{data}
\name{PermissionSet}
\alias{PermissionSet}
\title{PermissionSet Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PermissionSet
}
\description{
PermissionSet Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{name}}{}

\item{\code{permissions}}{}

\item{\code{built_in}}{}

\item{\code{all_access}}{}

\item{\code{url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
