% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrefetchAccessFilterValue.r
\docType{data}
\name{PrefetchAccessFilterValue}
\alias{PrefetchAccessFilterValue}
\title{PrefetchAccessFilterValue Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PrefetchAccessFilterValue
}
\description{
PrefetchAccessFilterValue Class
}
\section{Fields}{

\describe{
\item{\code{model}}{}

\item{\code{field}}{}

\item{\code{value}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
