% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrefetchDashboardFilterValue.r
\docType{data}
\name{PrefetchDashboardFilterValue}
\alias{PrefetchDashboardFilterValue}
\title{PrefetchDashboardFilterValue Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PrefetchDashboardFilterValue
}
\description{
PrefetchDashboardFilterValue Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{value}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
