% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrefetchDashboardRequest.r
\name{PrefetchDashboardRequest}
\alias{PrefetchDashboardRequest}
\title{PrefetchDashboardRequest Class}
\description{
PrefetchDashboardRequest Class

PrefetchDashboardRequest Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ttl}}{}

\item{\code{access_filters}}{}

\item{\code{dashboard_filters}}{}

\item{\code{created_at}}{}

\item{\code{computation_time}}{}

\item{\code{result_size_bytes}}{}

\item{\code{hit_count}}{}

\item{\code{touched_at}}{}

\item{\code{can}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PrefetchDashboardRequest$new()}}
\item \href{#method-toJSON}{\code{PrefetchDashboardRequest$toJSON()}}
\item \href{#method-fromJSON}{\code{PrefetchDashboardRequest$fromJSON()}}
\item \href{#method-toJSONString}{\code{PrefetchDashboardRequest$toJSONString()}}
\item \href{#method-fromJSONString}{\code{PrefetchDashboardRequest$fromJSONString()}}
\item \href{#method-clone}{\code{PrefetchDashboardRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrefetchDashboardRequest$new(
  ttl,
  access_filters,
  dashboard_filters,
  created_at,
  computation_time,
  result_size_bytes,
  hit_count,
  touched_at,
  can
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrefetchDashboardRequest$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrefetchDashboardRequest$fromJSON(PrefetchDashboardRequestJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrefetchDashboardRequest$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrefetchDashboardRequest$fromJSONString(PrefetchDashboardRequestJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrefetchDashboardRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
