% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project.r
\name{Project}
\alias{Project}
\title{Project Class}
\description{
Project Class

Project Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{can}}{}

\item{\code{id}}{}

\item{\code{name}}{}

\item{\code{uses_git}}{}

\item{\code{git_remote_url}}{}

\item{\code{git_username}}{}

\item{\code{git_password}}{}

\item{\code{git_production_branch_name}}{}

\item{\code{use_git_cookie_auth}}{}

\item{\code{git_username_user_attribute}}{}

\item{\code{git_password_user_attribute}}{}

\item{\code{git_service_name}}{}

\item{\code{git_application_server_http_port}}{}

\item{\code{git_application_server_http_scheme}}{}

\item{\code{deploy_secret}}{}

\item{\code{unset_deploy_secret}}{}

\item{\code{pull_request_mode}}{}

\item{\code{validation_required}}{}

\item{\code{git_release_mgmt_enabled}}{}

\item{\code{allow_warnings}}{}

\item{\code{is_example}}{}

\item{\code{dependency_status}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Project$new()}}
\item \href{#method-toJSON}{\code{Project$toJSON()}}
\item \href{#method-fromJSONObject}{\code{Project$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{Project$fromJSON()}}
\item \href{#method-toJSONString}{\code{Project$toJSONString()}}
\item \href{#method-fromJSONString}{\code{Project$fromJSONString()}}
\item \href{#method-clone}{\code{Project$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$new(
  can,
  id,
  name,
  uses_git,
  git_remote_url,
  git_username,
  git_password,
  git_production_branch_name,
  use_git_cookie_auth,
  git_username_user_attribute,
  git_password_user_attribute,
  git_service_name,
  git_application_server_http_port,
  git_application_server_http_scheme,
  deploy_secret,
  unset_deploy_secret,
  pull_request_mode,
  validation_required,
  git_release_mgmt_enabled,
  allow_warnings,
  is_example,
  dependency_status
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$fromJSONObject(ProjectJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$fromJSON(ProjectJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$fromJSONString(ProjectJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
