% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectApi.r
\docType{data}
\name{ProjectApi}
\alias{ProjectApi}
\title{Project operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ProjectApi
}
\description{
looker.Project
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

all_git_branches Get All Git Branchs 


all_git_connection_tests Get All Git Connection Tests 


all_project_files Get All Project Files 


all_projects Get All Projects 


create_git_deploy_key Create Deploy Key 


create_project Create Project 


git_deploy_key Git Deploy Key 


project Get Project 


project_file Get Project File 


project_validation_results Cached Project Validation Results 


project_workspace Get Project Workspace 


reset_project_to_production Reset To Production 


reset_project_to_remote Reset To Remote 


run_git_connection_test Run Git Connection Test 


update_project Update Project 


validate_project Validate Project 

}
}

\keyword{datasets}
