% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectError.r
\docType{data}
\name{ProjectError}
\alias{ProjectError}
\title{ProjectError Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ProjectError
}
\description{
ProjectError Class
}
\section{Fields}{

\describe{
\item{\code{code}}{}

\item{\code{severity}}{}

\item{\code{kind}}{}

\item{\code{message}}{}

\item{\code{field_name}}{}

\item{\code{file_path}}{}

\item{\code{line_number}}{}

\item{\code{model_id}}{}

\item{\code{explore}}{}

\item{\code{help_url}}{}

\item{\code{params}}{}

\item{\code{sanitized_message}}{}
}}

\keyword{datasets}
