% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectValidation.r
\docType{data}
\name{ProjectValidation}
\alias{ProjectValidation}
\title{ProjectValidation Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ProjectValidation
}
\description{
ProjectValidation Class
}
\section{Fields}{

\describe{
\item{\code{errors}}{}

\item{\code{project_digest}}{}

\item{\code{models_not_validated}}{}
}}

\keyword{datasets}
