% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectValidationCache.r
\docType{data}
\name{ProjectValidationCache}
\alias{ProjectValidationCache}
\title{ProjectValidationCache Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ProjectValidationCache
}
\description{
ProjectValidationCache Class
}
\section{Fields}{

\describe{
\item{\code{errors}}{}

\item{\code{project_digest}}{}

\item{\code{models_not_validated}}{}

\item{\code{stale}}{}
}}

\keyword{datasets}
