% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectWorkspace.r
\docType{data}
\name{ProjectWorkspace}
\alias{ProjectWorkspace}
\title{ProjectWorkspace Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ProjectWorkspace
}
\description{
ProjectWorkspace Class
}
\section{Fields}{

\describe{
\item{\code{project_id}}{}

\item{\code{workspace_id}}{}

\item{\code{git_status}}{}

\item{\code{git_head}}{}

\item{\code{git_branch}}{}

\item{\code{lookml_type}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
