% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.r
\name{Query}
\alias{Query}
\title{Query Class}
\description{
Query Class

Query Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{}

\item{\code{model}}{}

\item{\code{view}}{}

\item{\code{fields}}{}

\item{\code{pivots}}{}

\item{\code{fill_fields}}{}

\item{\code{filters}}{}

\item{\code{filter_expression}}{}

\item{\code{sorts}}{}

\item{\code{limit}}{}

\item{\code{column_limit}}{}

\item{\code{total}}{}

\item{\code{row_total}}{}

\item{\code{runtime}}{}

\item{\code{vis_config}}{}

\item{\code{filter_config}}{}

\item{\code{visible_ui_sections}}{}

\item{\code{slug}}{}

\item{\code{dynamic_fields}}{}

\item{\code{client_id}}{}

\item{\code{share_url}}{}

\item{\code{expanded_share_url}}{}

\item{\code{url}}{}

\item{\code{query_timezone}}{}

\item{\code{has_table_calculations}}{}

\item{\code{can}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Query$new()}}
\item \href{#method-toJSON}{\code{Query$toJSON()}}
\item \href{#method-fromJSON}{\code{Query$fromJSON()}}
\item \href{#method-toJSONString}{\code{Query$toJSONString()}}
\item \href{#method-fromJSONString}{\code{Query$fromJSONString()}}
\item \href{#method-clone}{\code{Query$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Query$new(
  id,
  model,
  view,
  fields,
  pivots,
  fill_fields,
  filters,
  filter_expression,
  sorts,
  limit,
  column_limit,
  total,
  row_total,
  runtime,
  vis_config,
  filter_config,
  visible_ui_sections,
  slug,
  dynamic_fields,
  client_id,
  share_url,
  expanded_share_url,
  url,
  query_timezone,
  has_table_calculations,
  can
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Query$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Query$fromJSON(QueryJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Query$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Query$fromJSONString(QueryJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Query$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
