% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryApi.r
\docType{data}
\name{QueryApi}
\alias{QueryApi}
\title{Query operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
QueryApi
}
\description{
looker.Query
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

create_query Create Query 


create_query_task Run Query Async 


query Get Query 


query_for_slug Get Query for Slug 


query_task Get Async Query Info 


query_task_multi_results Get Multiple Async Query Results 


query_task_results Get Async Query Results 


run_inline_query Run Inline Query 


run_query Run Query 


run_url_encoded_query Run Url Encoded Query 

}
}

\keyword{datasets}
