% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryTask.r
\docType{data}
\name{QueryTask}
\alias{QueryTask}
\title{QueryTask Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
QueryTask
}
\description{
QueryTask Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{query_id}}{}

\item{\code{query}}{}

\item{\code{generate_links}}{}

\item{\code{force_production}}{}

\item{\code{path_prefix}}{}

\item{\code{cache}}{}

\item{\code{server_table_calcs}}{}

\item{\code{cache_only}}{}

\item{\code{cache_key}}{}

\item{\code{status}}{}

\item{\code{source}}{}

\item{\code{runtime}}{}

\item{\code{rebuild_pdts}}{}

\item{\code{result_source}}{}

\item{\code{look_id}}{}

\item{\code{dashboard_id}}{}

\item{\code{result_format}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
