% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{ReduceTrials}
\alias{ReduceTrials}
\title{Reduce trial data using conventional steps and default options}
\usage{
ReduceTrials(trials)
}
\arguments{
\item{trials}{a TrialList}
}
\value{
the trials with times adjusted, time-sliced to interval of an
  interest, AOIs added, interpolated frames, and missing data calculated.
}
\description{
Reduce trial data using conventional steps and default options
}
