% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenderTaskApi.r
\name{RenderTaskApi}
\alias{RenderTaskApi}
\title{RenderTask operations}
\description{
looker.RenderTask
}
\section{Methods}{

\describe{

create_dashboard_render_task Create Dashboard Render Task 


create_look_render_task Create Look Render Task 


create_lookml_dashboard_render_task Create Lookml Dashboard Render Task 


create_query_render_task Create Query Render Task 


render_task Get Render Task 


render_task_results Render Task Results 

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RenderTaskApi$new()}}
\item \href{#method-create_dashboard_render_task}{\code{RenderTaskApi$create_dashboard_render_task()}}
\item \href{#method-create_look_render_task}{\code{RenderTaskApi$create_look_render_task()}}
\item \href{#method-create_lookml_dashboard_render_task}{\code{RenderTaskApi$create_lookml_dashboard_render_task()}}
\item \href{#method-create_query_render_task}{\code{RenderTaskApi$create_query_render_task()}}
\item \href{#method-render_task}{\code{RenderTaskApi$render_task()}}
\item \href{#method-render_task_results}{\code{RenderTaskApi$render_task_results()}}
\item \href{#method-clone}{\code{RenderTaskApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RenderTaskApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_dashboard_render_task"></a>}}
\if{latex}{\out{\hypertarget{method-create_dashboard_render_task}{}}}
\subsection{Method \code{create_dashboard_render_task()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RenderTaskApi$create_dashboard_render_task(
  dashboard_id,
  result_format,
  body,
  width,
  height,
  fields,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_look_render_task"></a>}}
\if{latex}{\out{\hypertarget{method-create_look_render_task}{}}}
\subsection{Method \code{create_look_render_task()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RenderTaskApi$create_look_render_task(
  look_id,
  result_format,
  width,
  height,
  fields,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_lookml_dashboard_render_task"></a>}}
\if{latex}{\out{\hypertarget{method-create_lookml_dashboard_render_task}{}}}
\subsection{Method \code{create_lookml_dashboard_render_task()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RenderTaskApi$create_lookml_dashboard_render_task(
  dashboard_id,
  result_format,
  body,
  width,
  height,
  fields,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_query_render_task"></a>}}
\if{latex}{\out{\hypertarget{method-create_query_render_task}{}}}
\subsection{Method \code{create_query_render_task()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RenderTaskApi$create_query_render_task(
  query_id,
  result_format,
  width,
  height,
  fields,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_task"></a>}}
\if{latex}{\out{\hypertarget{method-render_task}{}}}
\subsection{Method \code{render_task()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RenderTaskApi$render_task(render_task_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_task_results"></a>}}
\if{latex}{\out{\hypertarget{method-render_task_results}{}}}
\subsection{Method \code{render_task_results()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RenderTaskApi$render_task_results(render_task_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RenderTaskApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
