% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepositoryCredential.r
\name{RepositoryCredential}
\alias{RepositoryCredential}
\title{RepositoryCredential Class}
\description{
RepositoryCredential Class

RepositoryCredential Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{can}}{}

\item{\code{id}}{}

\item{\code{root_project_id}}{}

\item{\code{remote_url}}{}

\item{\code{git_username}}{}

\item{\code{git_password}}{}

\item{\code{ssh_public_key}}{}

\item{\code{is_configured}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RepositoryCredential$new()}}
\item \href{#method-toJSON}{\code{RepositoryCredential$toJSON()}}
\item \href{#method-fromJSONObject}{\code{RepositoryCredential$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{RepositoryCredential$fromJSON()}}
\item \href{#method-toJSONString}{\code{RepositoryCredential$toJSONString()}}
\item \href{#method-fromJSONString}{\code{RepositoryCredential$fromJSONString()}}
\item \href{#method-clone}{\code{RepositoryCredential$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepositoryCredential$new(
  can,
  id,
  root_project_id,
  remote_url,
  git_username,
  git_password,
  ssh_public_key,
  is_configured
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepositoryCredential$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepositoryCredential$fromJSONObject(RepositoryCredentialJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepositoryCredential$fromJSON(RepositoryCredentialJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepositoryCredential$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepositoryCredential$fromJSONString(RepositoryCredentialJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepositoryCredential$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
