% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultMakerFilterablesListen.r
\name{ResultMakerFilterablesListen}
\alias{ResultMakerFilterablesListen}
\title{ResultMakerFilterablesListen Class}
\description{
ResultMakerFilterablesListen Class

ResultMakerFilterablesListen Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dashboard_filter_name}}{}

\item{\code{field}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ResultMakerFilterablesListen$new()}}
\item \href{#method-toJSON}{\code{ResultMakerFilterablesListen$toJSON()}}
\item \href{#method-fromJSONObject}{\code{ResultMakerFilterablesListen$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{ResultMakerFilterablesListen$fromJSON()}}
\item \href{#method-toJSONString}{\code{ResultMakerFilterablesListen$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ResultMakerFilterablesListen$fromJSONString()}}
\item \href{#method-clone}{\code{ResultMakerFilterablesListen$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultMakerFilterablesListen$new(dashboard_filter_name, field)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultMakerFilterablesListen$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultMakerFilterablesListen$fromJSONObject(
  ResultMakerFilterablesListenJsonObject
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultMakerFilterablesListen$fromJSON(ResultMakerFilterablesListenJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultMakerFilterablesListen$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultMakerFilterablesListen$fromJSONString(ResultMakerFilterablesListenJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultMakerFilterablesListen$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
