% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultMakerWithIdVisConfigAndDynamicFields.r
\name{ResultMakerWithIdVisConfigAndDynamicFields}
\alias{ResultMakerWithIdVisConfigAndDynamicFields}
\title{ResultMakerWithIdVisConfigAndDynamicFields Class}
\description{
ResultMakerWithIdVisConfigAndDynamicFields Class

ResultMakerWithIdVisConfigAndDynamicFields Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{}

\item{\code{dynamic_fields}}{}

\item{\code{filterables}}{}

\item{\code{sorts}}{}

\item{\code{merge_result_id}}{}

\item{\code{total}}{}

\item{\code{query_id}}{}

\item{\code{sql_query_id}}{}

\item{\code{query}}{}

\item{\code{vis_config}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ResultMakerWithIdVisConfigAndDynamicFields$new()}}
\item \href{#method-toJSON}{\code{ResultMakerWithIdVisConfigAndDynamicFields$toJSON()}}
\item \href{#method-fromJSONObject}{\code{ResultMakerWithIdVisConfigAndDynamicFields$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{ResultMakerWithIdVisConfigAndDynamicFields$fromJSON()}}
\item \href{#method-toJSONString}{\code{ResultMakerWithIdVisConfigAndDynamicFields$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ResultMakerWithIdVisConfigAndDynamicFields$fromJSONString()}}
\item \href{#method-clone}{\code{ResultMakerWithIdVisConfigAndDynamicFields$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultMakerWithIdVisConfigAndDynamicFields$new(
  id,
  dynamic_fields,
  filterables,
  sorts,
  merge_result_id,
  total,
  query_id,
  sql_query_id,
  query,
  vis_config
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultMakerWithIdVisConfigAndDynamicFields$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultMakerWithIdVisConfigAndDynamicFields$fromJSONObject(
  ResultMakerWithIdVisConfigAndDynamicFieldsJsonObject
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultMakerWithIdVisConfigAndDynamicFields$fromJSON(
  ResultMakerWithIdVisConfigAndDynamicFieldsJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultMakerWithIdVisConfigAndDynamicFields$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultMakerWithIdVisConfigAndDynamicFields$fromJSONString(
  ResultMakerWithIdVisConfigAndDynamicFieldsJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultMakerWithIdVisConfigAndDynamicFields$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
