% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Role.r
\docType{data}
\name{Role}
\alias{Role}
\title{Role Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Role
}
\description{
Role Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{name}}{}

\item{\code{permission_set}}{}

\item{\code{permission_set_id}}{}

\item{\code{model_set}}{}

\item{\code{model_set_id}}{}

\item{\code{url}}{}

\item{\code{users_url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
