% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoleApi.r
\docType{data}
\name{RoleApi}
\alias{RoleApi}
\title{Role operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RoleApi
}
\description{
looker.Role
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

all_model_sets Get All Model Sets 


all_permission_sets Get All Permission Sets 


all_permissions Get All Permissions 


all_roles Get All Roles 


create_model_set Create Model Set 


create_permission_set Create Permission Set 


create_role Create Role 


delete_model_set Delete Model Set 


delete_permission_set Delete Permission Set 


delete_role Delete Role 


model_set Get Model Set 


permission_set Get Permission Set 


role Get Role 


role_groups Get Role Groups 


role_users Get Role Users 


set_role_groups Update Role Groups 


set_role_users Update Role Users 


update_model_set Update Model Set 


update_permission_set Update Permission Set 


update_role Update Role 

}
}

\keyword{datasets}
