% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunningQueries.r
\name{RunningQueries}
\alias{RunningQueries}
\title{RunningQueries Class}
\description{
RunningQueries Class

RunningQueries Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{}

\item{\code{user}}{}

\item{\code{query}}{}

\item{\code{sql_query}}{}

\item{\code{look}}{}

\item{\code{created_at}}{}

\item{\code{completed_at}}{}

\item{\code{query_id}}{}

\item{\code{source}}{}

\item{\code{node_id}}{}

\item{\code{slug}}{}

\item{\code{query_task_id}}{}

\item{\code{cache_key}}{}

\item{\code{connection_name}}{}

\item{\code{dialect}}{}

\item{\code{connection_id}}{}

\item{\code{message}}{}

\item{\code{status}}{}

\item{\code{runtime}}{}

\item{\code{sql}}{}

\item{\code{can}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RunningQueries$new()}}
\item \href{#method-toJSON}{\code{RunningQueries$toJSON()}}
\item \href{#method-fromJSON}{\code{RunningQueries$fromJSON()}}
\item \href{#method-toJSONString}{\code{RunningQueries$toJSONString()}}
\item \href{#method-fromJSONString}{\code{RunningQueries$fromJSONString()}}
\item \href{#method-clone}{\code{RunningQueries$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunningQueries$new(
  id,
  user,
  query,
  sql_query,
  look,
  created_at,
  completed_at,
  query_id,
  source,
  node_id,
  slug,
  query_task_id,
  cache_key,
  connection_name,
  dialect,
  connection_id,
  message,
  status,
  runtime,
  sql,
  can
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunningQueries$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunningQueries$fromJSON(RunningQueriesJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunningQueries$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunningQueries$fromJSONString(RunningQueriesJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunningQueries$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
