% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunningQueriesApi.r
\name{RunningQueriesApi}
\alias{RunningQueriesApi}
\title{RunningQueries operations}
\description{
looker.RunningQueries
}
\section{Methods}{

\describe{

all_running_queries Get All Running Queries 


kill_query Kill Running Query 

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RunningQueriesApi$new()}}
\item \href{#method-all_running_queries}{\code{RunningQueriesApi$all_running_queries()}}
\item \href{#method-kill_query}{\code{RunningQueriesApi$kill_query()}}
\item \href{#method-clone}{\code{RunningQueriesApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunningQueriesApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_running_queries"></a>}}
\if{latex}{\out{\hypertarget{method-all_running_queries}{}}}
\subsection{Method \code{all_running_queries()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunningQueriesApi$all_running_queries(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kill_query"></a>}}
\if{latex}{\out{\hypertarget{method-kill_query}{}}}
\subsection{Method \code{kill_query()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunningQueriesApi$kill_query(query_task_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunningQueriesApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
