% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamlConfig.r
\docType{data}
\name{SamlConfig}
\alias{SamlConfig}
\title{SamlConfig Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SamlConfig
}
\description{
SamlConfig Class
}
\section{Fields}{

\describe{
\item{\code{enabled}}{}

\item{\code{idp_cert}}{}

\item{\code{idp_url}}{}

\item{\code{idp_issuer}}{}

\item{\code{idp_audience}}{}

\item{\code{allowed_clock_drift}}{}

\item{\code{user_attribute_map_email}}{}

\item{\code{user_attribute_map_first_name}}{}

\item{\code{user_attribute_map_last_name}}{}

\item{\code{new_user_migration_types}}{}

\item{\code{alternate_email_login_allowed}}{}

\item{\code{test_slug}}{}

\item{\code{modified_at}}{}

\item{\code{modified_by}}{}

\item{\code{default_new_user_roles}}{}

\item{\code{default_new_user_groups}}{}

\item{\code{default_new_user_role_ids}}{}

\item{\code{default_new_user_group_ids}}{}

\item{\code{set_roles_from_groups}}{}

\item{\code{groups_attribute}}{}

\item{\code{groups}}{}

\item{\code{groups_with_role_ids}}{}

\item{\code{auth_requires_role}}{}

\item{\code{user_attributes}}{}

\item{\code{user_attributes_with_ids}}{}

\item{\code{groups_finder_type}}{}

\item{\code{groups_member_value}}{}

\item{\code{bypass_login_page}}{}

\item{\code{url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
