% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamlGroupRead.r
\docType{data}
\name{SamlGroupRead}
\alias{SamlGroupRead}
\title{SamlGroupRead Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SamlGroupRead
}
\description{
SamlGroupRead Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{roles}}{}

\item{\code{url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
