% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamlGroupWrite.r
\docType{data}
\name{SamlGroupWrite}
\alias{SamlGroupWrite}
\title{SamlGroupWrite Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SamlGroupWrite
}
\description{
SamlGroupWrite Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{role_ids}}{}

\item{\code{url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
