% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamlMetadataParseResult.r
\docType{data}
\name{SamlMetadataParseResult}
\alias{SamlMetadataParseResult}
\title{SamlMetadataParseResult Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SamlMetadataParseResult
}
\description{
SamlMetadataParseResult Class
}
\section{Fields}{

\describe{
\item{\code{idp_issuer}}{}

\item{\code{idp_url}}{}

\item{\code{idp_cert}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
