% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamlUserAttributeWrite.r
\docType{data}
\name{SamlUserAttributeWrite}
\alias{SamlUserAttributeWrite}
\title{SamlUserAttributeWrite Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SamlUserAttributeWrite
}
\description{
SamlUserAttributeWrite Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{required}}{}

\item{\code{user_attribute_ids}}{}

\item{\code{url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
