% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScheduledPlan.r
\docType{data}
\name{ScheduledPlan}
\alias{ScheduledPlan}
\title{ScheduledPlan Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ScheduledPlan
}
\description{
ScheduledPlan Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{name}}{}

\item{\code{created_at}}{}

\item{\code{updated_at}}{}

\item{\code{title}}{}

\item{\code{user_id}}{}

\item{\code{user}}{}

\item{\code{run_as_recipient}}{}

\item{\code{enabled}}{}

\item{\code{next_run_at}}{}

\item{\code{last_run_at}}{}

\item{\code{look_id}}{}

\item{\code{dashboard_id}}{}

\item{\code{lookml_dashboard_id}}{}

\item{\code{filters_string}}{}

\item{\code{dashboard_filters}}{}

\item{\code{require_results}}{}

\item{\code{require_no_results}}{}

\item{\code{require_change}}{}

\item{\code{send_all_results}}{}

\item{\code{crontab}}{}

\item{\code{datagroup}}{}

\item{\code{timezone}}{}

\item{\code{query_id}}{}

\item{\code{scheduled_plan_destination}}{}

\item{\code{run_once}}{}

\item{\code{include_links}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
