% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScheduledPlanApi.r
\docType{data}
\name{ScheduledPlanApi}
\alias{ScheduledPlanApi}
\title{ScheduledPlan operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ScheduledPlanApi
}
\description{
looker.ScheduledPlan
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

all_scheduled_plans Get All Scheduled Plans 


create_scheduled_plan Create Scheduled Plan 


delete_scheduled_plan Delete Scheduled Plan 


scheduled_plan Get Scheduled Plan 


scheduled_plan_run_once Run Scheduled Plan Once 


scheduled_plans_for_dashboard Scheduled Plans for Dashboard 


scheduled_plans_for_look Scheduled Plans for Look 


scheduled_plans_for_lookml_dashboard Scheduled Plans for LookML Dashboard 


scheduled_plans_for_space Scheduled Plans for Space 


update_scheduled_plan Update Scheduled Plan 

}
}

\keyword{datasets}
