% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScheduledPlanDestination.r
\docType{data}
\name{ScheduledPlanDestination}
\alias{ScheduledPlanDestination}
\title{ScheduledPlanDestination Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ScheduledPlanDestination
}
\description{
ScheduledPlanDestination Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{scheduled_plan_id}}{}

\item{\code{format}}{}

\item{\code{apply_formatting}}{}

\item{\code{apply_vis}}{}

\item{\code{address}}{}

\item{\code{looker_recipient}}{}

\item{\code{type}}{}

\item{\code{parameters}}{}

\item{\code{secret_parameters}}{}

\item{\code{message}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
