% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SchemaColumn.r
\name{SchemaColumn}
\alias{SchemaColumn}
\title{SchemaColumn Class}
\description{
SchemaColumn Class

SchemaColumn Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{}

\item{\code{sql_escaped_name}}{}

\item{\code{schema_name}}{}

\item{\code{data_type_database}}{}

\item{\code{data_type}}{}

\item{\code{data_type_looker}}{}

\item{\code{description}}{}

\item{\code{column_size}}{}

\item{\code{snippets}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SchemaColumn$new()}}
\item \href{#method-toJSON}{\code{SchemaColumn$toJSON()}}
\item \href{#method-fromJSONObject}{\code{SchemaColumn$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{SchemaColumn$fromJSON()}}
\item \href{#method-toJSONString}{\code{SchemaColumn$toJSONString()}}
\item \href{#method-fromJSONString}{\code{SchemaColumn$fromJSONString()}}
\item \href{#method-clone}{\code{SchemaColumn$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SchemaColumn$new(
  name,
  sql_escaped_name,
  schema_name,
  data_type_database,
  data_type,
  data_type_looker,
  description,
  column_size,
  snippets
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SchemaColumn$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SchemaColumn$fromJSONObject(SchemaColumnJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SchemaColumn$fromJSON(SchemaColumnJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SchemaColumn$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SchemaColumn$fromJSONString(SchemaColumnJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SchemaColumn$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
