% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gca.R
\name{se_prop}
\alias{se_prop}
\title{Standard error for proportions}
\usage{
se_prop(proportion, n_possible)
}
\arguments{
\item{proportion}{proportions of hits}

\item{n_possible}{numbers of total events}
}
\value{
the standard errors of the proportion estimates
}
\description{
See http://www.r-tutor.com/elementary-statistics/interval-estimation/interval-estimate-population-proportion
}
