% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{Session}
\alias{Session}
\title{Create a Session from multiple LWL Blocks}
\usage{
Session(...)
}
\arguments{
\item{...}{Additional arguments passed onto S3 methods. Currently ignored.}

\item{session_path}{The directory path of both pairs of .gazedata and .txt
files for the session---one pair for each block.}

\item{list_of_blocks}{A list of Block objects, each Block object a block of
the same experimental session. The Block objects in list_of_blocks should
be ordered as they were presented during the experiment. That is, Block 1
of the experiment should be the first element of list_of_blocks, etc.}
}
\value{
A Session object.
}
\description{
\code{Session()} is a generic function for initializing a Session object,
which represents the gazedata and stimdata for each trial in an experimental
session. The representation of each trial in a Session object is the same as
in a Block object.
}
\details{
\code{Session.character} creates a Session object from the path of the
directory that contains the \code{.gazedata} and \code{.txt} (stimdata) files
for blocks of an experimental session.
\code{Session.list} is a (largely internal) method for initializing a Session
object from a list of Block objects.
}
