% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SessionApi.r
\name{SessionApi}
\alias{SessionApi}
\title{Session operations}
\description{
looker.Session
}
\section{Methods}{

\describe{

session Get Session


update_session Update Session

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SessionApi$new()}}
\item \href{#method-session}{\code{SessionApi$session()}}
\item \href{#method-update_session}{\code{SessionApi$update_session()}}
\item \href{#method-clone}{\code{SessionApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-session"></a>}}
\if{latex}{\out{\hypertarget{method-session}{}}}
\subsection{Method \code{session()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionApi$session(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_session"></a>}}
\if{latex}{\out{\hypertarget{method-update_session}{}}}
\subsection{Method \code{update_session()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionApi$update_session(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
