% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{slice}
\alias{slice}
\title{Divide a vector into equal-sized chunks, except for maybe the last one}
\usage{
slice(xs, step = 5)
}
\arguments{
\item{xs}{a vector}

\item{step}{the length of a chunk}
}
\value{
a list with the elements of `xs` partitioned into `step`-sized chunks
}
\description{
Divide a vector into equal-sized chunks, except for maybe the last one
}

