% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Snippet.r
\docType{data}
\name{Snippet}
\alias{Snippet}
\title{Snippet Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Snippet
}
\description{
Snippet Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{label}}{}

\item{\code{sql}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
