% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceApi.r
\docType{data}
\name{SpaceApi}
\alias{SpaceApi}
\title{Space operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SpaceApi
}
\description{
looker.Space
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

all_spaces Get All Spaces 


create_space Create Space 


delete_space Delete Space 


search_spaces Search Spaces 


space Get Space 


space_ancestors Get Space Ancestors 


space_children Get Space Children 


space_children_search Search Space Children 


space_dashboards Get Space Dashboards 


space_looks Get Space Looks 


space_parent Get Space Parent 


update_space Update Space 

}
}

\keyword{datasets}
