% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceBase.r
\docType{data}
\name{SpaceBase}
\alias{SpaceBase}
\title{SpaceBase Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SpaceBase
}
\description{
SpaceBase Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{content_metadata_id}}{}

\item{\code{creator_id}}{}

\item{\code{name}}{}

\item{\code{is_personal}}{}

\item{\code{is_personal_descendant}}{}

\item{\code{is_shared_root}}{}

\item{\code{is_users_root}}{}

\item{\code{is_root}}{}

\item{\code{is_user_root}}{}

\item{\code{is_embed}}{}

\item{\code{is_embed_shared_root}}{}

\item{\code{is_embed_users_root}}{}

\item{\code{external_id}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
