% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SqlQuery.r
\name{SqlQuery}
\alias{SqlQuery}
\title{SqlQuery Class}
\description{
SqlQuery Class

SqlQuery Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{can}}{}

\item{\code{slug}}{}

\item{\code{last_runtime}}{}

\item{\code{run_count}}{}

\item{\code{browser_limit}}{}

\item{\code{sql}}{}

\item{\code{last_run_at}}{}

\item{\code{connection}}{}

\item{\code{model_name}}{}

\item{\code{creator}}{}

\item{\code{explore_url}}{}

\item{\code{plaintext}}{}

\item{\code{vis_config}}{}

\item{\code{result_maker_id}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SqlQuery$new()}}
\item \href{#method-toJSON}{\code{SqlQuery$toJSON()}}
\item \href{#method-fromJSONObject}{\code{SqlQuery$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{SqlQuery$fromJSON()}}
\item \href{#method-toJSONString}{\code{SqlQuery$toJSONString()}}
\item \href{#method-fromJSONString}{\code{SqlQuery$fromJSONString()}}
\item \href{#method-clone}{\code{SqlQuery$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SqlQuery$new(
  can,
  slug,
  last_runtime,
  run_count,
  browser_limit,
  sql,
  last_run_at,
  connection,
  model_name,
  creator,
  explore_url,
  plaintext,
  vis_config,
  result_maker_id
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SqlQuery$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SqlQuery$fromJSONObject(SqlQueryJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SqlQuery$fromJSON(SqlQueryJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SqlQuery$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SqlQuery$fromJSONString(SqlQueryJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SqlQuery$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
