% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SqlQueryApi.r
\docType{data}
\name{SqlQueryApi}
\alias{SqlQueryApi}
\title{SqlQuery operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SqlQueryApi
}
\description{
looker.SqlQuery
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

create_sql_query Create SQL Runner Query 


sql_query Get SQL Runner Query 

}
}

\keyword{datasets}
