% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SqlQueryCreate.r
\docType{data}
\name{SqlQueryCreate}
\alias{SqlQueryCreate}
\title{SqlQueryCreate Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SqlQueryCreate
}
\description{
SqlQueryCreate Class
}
\section{Fields}{

\describe{
\item{\code{connection_id}}{}

\item{\code{sql}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
