% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SshTunnel.r
\name{SshTunnel}
\alias{SshTunnel}
\title{SshTunnel Class}
\description{
SshTunnel Class

SshTunnel Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tunnel_id}}{}

\item{\code{ssh_server_id}}{}

\item{\code{ssh_server_name}}{}

\item{\code{ssh_server_host}}{}

\item{\code{ssh_server_port}}{}

\item{\code{ssh_server_user}}{}

\item{\code{last_attempt}}{}

\item{\code{local_host_port}}{}

\item{\code{database_host}}{}

\item{\code{database_port}}{}

\item{\code{status}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SshTunnel$new()}}
\item \href{#method-toJSON}{\code{SshTunnel$toJSON()}}
\item \href{#method-fromJSONObject}{\code{SshTunnel$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{SshTunnel$fromJSON()}}
\item \href{#method-toJSONString}{\code{SshTunnel$toJSONString()}}
\item \href{#method-fromJSONString}{\code{SshTunnel$fromJSONString()}}
\item \href{#method-clone}{\code{SshTunnel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SshTunnel$new(
  tunnel_id,
  ssh_server_id,
  ssh_server_name,
  ssh_server_host,
  ssh_server_port,
  ssh_server_user,
  last_attempt,
  local_host_port,
  database_host,
  database_port,
  status
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SshTunnel$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SshTunnel$fromJSONObject(SshTunnelJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SshTunnel$fromJSON(SshTunnelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SshTunnel$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SshTunnel$fromJSONString(SshTunnelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SshTunnel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
