% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stimdata.R
\name{Stimdata}
\alias{Stimdata}
\title{Parse an LWL stimulus data file outputted by Eprime}
\usage{
Stimdata(stimdata_path, output_file = lwl_opts$get("write_stimdata"))
}
\arguments{
\item{stimdata_path}{Either the full or relative path to the \code{.txt} file
that is to be parsed.}

\item{output_file}{Whether to write the parsed stimdata to a csv file.
Defaults to FALSE. If TRUE, that stimdata is saved to
\code{[folder]/[basename]_stim.csv}, where the folder and basename are
extracted from the path used in \code{stimdata_path}}
}
\value{
A dataframe containing the parsed stimdata.  Each row of the
  dataframe is the stimdata for a single experimental trial.
}
\description{
\code{Stimdata} is used to read and parse the \code{.txt} file that is output
by E-prime during a session of a Looking While Listening experiment.
}
\details{
A stimdata file in a Looking While Listening task file the naming
convention: \code{[Task]_[BlockNo]_[SubjectID]}. The stimdata file is
assigned a class based on the task in the filename, and then methods for
extracting the stimdata are dispatched based on that class value. Valid task
names include "RWL", "MP", and "Coartic".
}

