% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThemeApi.r
\name{ThemeApi}
\alias{ThemeApi}
\title{Theme operations}
\description{
looker.Theme
}
\section{Methods}{

\describe{

active_themes Get Active Themes


all_themes Get All Themes


create_theme Create Theme


default_theme Get Default Theme


delete_theme Delete Theme


search_themes Search Themes


set_default_theme Set Default Theme


theme Get Theme


theme_or_default Get Theme or Default


update_theme Update Theme


validate_theme Validate Theme

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ThemeApi$new()}}
\item \href{#method-active_themes}{\code{ThemeApi$active_themes()}}
\item \href{#method-all_themes}{\code{ThemeApi$all_themes()}}
\item \href{#method-create_theme}{\code{ThemeApi$create_theme()}}
\item \href{#method-default_theme}{\code{ThemeApi$default_theme()}}
\item \href{#method-delete_theme}{\code{ThemeApi$delete_theme()}}
\item \href{#method-search_themes}{\code{ThemeApi$search_themes()}}
\item \href{#method-set_default_theme}{\code{ThemeApi$set_default_theme()}}
\item \href{#method-theme}{\code{ThemeApi$theme()}}
\item \href{#method-theme_or_default}{\code{ThemeApi$theme_or_default()}}
\item \href{#method-update_theme}{\code{ThemeApi$update_theme()}}
\item \href{#method-validate_theme}{\code{ThemeApi$validate_theme()}}
\item \href{#method-clone}{\code{ThemeApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-active_themes"></a>}}
\if{latex}{\out{\hypertarget{method-active_themes}{}}}
\subsection{Method \code{active_themes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeApi$active_themes(name, ts, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_themes"></a>}}
\if{latex}{\out{\hypertarget{method-all_themes}{}}}
\subsection{Method \code{all_themes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeApi$all_themes(fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_theme"></a>}}
\if{latex}{\out{\hypertarget{method-create_theme}{}}}
\subsection{Method \code{create_theme()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeApi$create_theme(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-default_theme"></a>}}
\if{latex}{\out{\hypertarget{method-default_theme}{}}}
\subsection{Method \code{default_theme()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeApi$default_theme(ts, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_theme"></a>}}
\if{latex}{\out{\hypertarget{method-delete_theme}{}}}
\subsection{Method \code{delete_theme()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeApi$delete_theme(theme_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_themes"></a>}}
\if{latex}{\out{\hypertarget{method-search_themes}{}}}
\subsection{Method \code{search_themes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeApi$search_themes(
  id,
  name,
  begin_at,
  end_at,
  limit,
  offset,
  sorts,
  fields,
  filter_or,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_default_theme"></a>}}
\if{latex}{\out{\hypertarget{method-set_default_theme}{}}}
\subsection{Method \code{set_default_theme()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeApi$set_default_theme(name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-theme"></a>}}
\if{latex}{\out{\hypertarget{method-theme}{}}}
\subsection{Method \code{theme()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeApi$theme(theme_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-theme_or_default"></a>}}
\if{latex}{\out{\hypertarget{method-theme_or_default}{}}}
\subsection{Method \code{theme_or_default()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeApi$theme_or_default(name, ts, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_theme"></a>}}
\if{latex}{\out{\hypertarget{method-update_theme}{}}}
\subsection{Method \code{update_theme()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeApi$update_theme(theme_id, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate_theme"></a>}}
\if{latex}{\out{\hypertarget{method-validate_theme}{}}}
\subsection{Method \code{validate_theme()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeApi$validate_theme(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
