% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThemeSettings.r
\name{ThemeSettings}
\alias{ThemeSettings}
\title{ThemeSettings Class}
\description{
ThemeSettings Class

ThemeSettings Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{background_color}}{}

\item{\code{base_font_size}}{}

\item{\code{color_collection_id}}{}

\item{\code{font_color}}{}

\item{\code{font_family}}{}

\item{\code{font_source}}{}

\item{\code{info_button_color}}{}

\item{\code{primary_button_color}}{}

\item{\code{show_filters_bar}}{}

\item{\code{show_title}}{}

\item{\code{text_tile_text_color}}{}

\item{\code{tile_background_color}}{}

\item{\code{tile_text_color}}{}

\item{\code{title_color}}{}

\item{\code{warn_button_color}}{}

\item{\code{tile_title_alignment}}{}

\item{\code{tile_shadow}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ThemeSettings$new()}}
\item \href{#method-toJSON}{\code{ThemeSettings$toJSON()}}
\item \href{#method-fromJSONObject}{\code{ThemeSettings$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{ThemeSettings$fromJSON()}}
\item \href{#method-toJSONString}{\code{ThemeSettings$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ThemeSettings$fromJSONString()}}
\item \href{#method-clone}{\code{ThemeSettings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeSettings$new(
  background_color,
  base_font_size,
  color_collection_id,
  font_color,
  font_family,
  font_source,
  info_button_color,
  primary_button_color,
  show_filters_bar,
  show_title,
  text_tile_text_color,
  tile_background_color,
  tile_text_color,
  title_color,
  warn_button_color,
  tile_title_alignment,
  tile_shadow
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeSettings$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeSettings$fromJSONObject(ThemeSettingsJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeSettings$fromJSON(ThemeSettingsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeSettings$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeSettings$fromJSONString(ThemeSettingsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThemeSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
