% Generated by roxygen2 (4.0.2): do not edit by hand
\name{trial_lapply}
\alias{trial_lapply}
\title{Apply a function to a list of trials, preserving classes of the list}
\usage{
trial_lapply(trials, trial_func, ...)
}
\arguments{
\item{trials}{a list of Trial objects}

\item{trial_func}{a function to apply on each Trial object}

\item{...}{additional arguments passed to the function}
}
\description{
This function is just a wrapper for \code{lapply} that makes sure that the
classes of the inputted list are preserved in the outputted list.
}

