% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User.r
\docType{data}
\name{User}
\alias{User}
\title{User Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
User
}
\description{
User Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{first_name}}{}

\item{\code{last_name}}{}

\item{\code{display_name}}{}

\item{\code{email}}{}

\item{\code{is_disabled}}{}

\item{\code{avatar_url}}{}

\item{\code{home_space_id}}{}

\item{\code{personal_space_id}}{}

\item{\code{embed_group_space_id}}{}

\item{\code{access_filters}}{}

\item{\code{credentials_email}}{}

\item{\code{credentials_totp}}{}

\item{\code{credentials_ldap}}{}

\item{\code{credentials_google}}{}

\item{\code{credentials_saml}}{}

\item{\code{credentials_oidc}}{}

\item{\code{credentials_api}}{}

\item{\code{credentials_api3}}{}

\item{\code{credentials_embed}}{}

\item{\code{credentials_looker_openid}}{}

\item{\code{sessions}}{}

\item{\code{role_ids}}{}

\item{\code{group_ids}}{}

\item{\code{presumed_looker_employee}}{}

\item{\code{verified_looker_employee}}{}

\item{\code{looker_versions}}{}

\item{\code{ui_state}}{}

\item{\code{url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
