% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserApi.r
\docType{data}
\name{UserApi}
\alias{UserApi}
\title{User operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
UserApi
}
\description{
looker.User
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

all_user_access_filters Get All Access Filters 


all_user_credentials_api3s Get All API 3 Credentials 


all_user_credentials_embeds Get All Embedding Credentials 


all_user_sessions Get All Web Login Sessions 


all_users Get All Users 


create_user Create User 


create_user_access_filter Create Access Filter 


create_user_credentials_api Create API Credential 


create_user_credentials_api3 Create API 3 Credential 


create_user_credentials_email Create Email/Password Credential 


create_user_credentials_email_password_reset Create Password Reset Token 


create_user_credentials_totp Create Two-Factor Credential 


delete_user Delete User 


delete_user_access_filter Delete Access Filter 


delete_user_attribute_user_value Delete User Attribute User Value 


delete_user_credentials_api Delete API Credential 


delete_user_credentials_api3 Delete API 3 Credential 


delete_user_credentials_email Delete Email/Password Credential 


delete_user_credentials_embed Delete Embedding Credential 


delete_user_credentials_google Delete Google Auth Credential 


delete_user_credentials_ldap Delete LDAP Credential 


delete_user_credentials_looker_openid Delete Looker OpenId Credential 


delete_user_credentials_oidc Delete OIDC Auth Credential 


delete_user_credentials_saml Delete Saml Auth Credential 


delete_user_credentials_totp Delete Two-Factor Credential 


delete_user_session Delete Web Login Session 


me Get Current User 


search_users Search Users 


search_users_names Search User Names 


set_user_attribute_user_value Set User Attribute User Value 


set_user_roles Set User Roles 


update_user Update User 


update_user_access_filter Update Access Filter 


update_user_credentials_email Update Email/Password Credential 


user Get User by Id 


user_access_filter Get Access Filter 


user_attribute_user_values Get User Attribute Values 


user_credentials_api Get API Credential 


user_credentials_api3 Get API 3 Credential 


user_credentials_email Get Email/Password Credential 


user_credentials_embed Get Embedding Credential 


user_credentials_google Get Google Auth Credential 


user_credentials_ldap Get LDAP Credential 


user_credentials_looker_openid Get Looker OpenId Credential 


user_credentials_oidc Get OIDC Auth Credential 


user_credentials_saml Get Saml Auth Credential 


user_credentials_totp Get Two-Factor Credential 


user_for_credential Get User by Credential Id 


user_roles Get User Roles 


user_session Get Web Login Session 

}
}

\keyword{datasets}
