% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserAttributeApi.r
\name{UserAttributeApi}
\alias{UserAttributeApi}
\title{UserAttribute operations}
\description{
looker.UserAttribute
}
\section{Methods}{

\describe{

all_user_attribute_group_values Get User Attribute Group Values


all_user_attributes Get All User Attributes


create_user_attribute Create User Attribute


delete_user_attribute Delete User Attribute


set_user_attribute_group_values Set User Attribute Group Values


update_user_attribute Update User Attribute


user_attribute Get User Attribute

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UserAttributeApi$new()}}
\item \href{#method-all_user_attribute_group_values}{\code{UserAttributeApi$all_user_attribute_group_values()}}
\item \href{#method-all_user_attributes}{\code{UserAttributeApi$all_user_attributes()}}
\item \href{#method-create_user_attribute}{\code{UserAttributeApi$create_user_attribute()}}
\item \href{#method-delete_user_attribute}{\code{UserAttributeApi$delete_user_attribute()}}
\item \href{#method-set_user_attribute_group_values}{\code{UserAttributeApi$set_user_attribute_group_values()}}
\item \href{#method-update_user_attribute}{\code{UserAttributeApi$update_user_attribute()}}
\item \href{#method-user_attribute}{\code{UserAttributeApi$user_attribute()}}
\item \href{#method-clone}{\code{UserAttributeApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserAttributeApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_user_attribute_group_values"></a>}}
\if{latex}{\out{\hypertarget{method-all_user_attribute_group_values}{}}}
\subsection{Method \code{all_user_attribute_group_values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserAttributeApi$all_user_attribute_group_values(
  user_attribute_id,
  fields,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_user_attributes"></a>}}
\if{latex}{\out{\hypertarget{method-all_user_attributes}{}}}
\subsection{Method \code{all_user_attributes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserAttributeApi$all_user_attributes(fields, sorts, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_user_attribute"></a>}}
\if{latex}{\out{\hypertarget{method-create_user_attribute}{}}}
\subsection{Method \code{create_user_attribute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserAttributeApi$create_user_attribute(body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_user_attribute"></a>}}
\if{latex}{\out{\hypertarget{method-delete_user_attribute}{}}}
\subsection{Method \code{delete_user_attribute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserAttributeApi$delete_user_attribute(user_attribute_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_user_attribute_group_values"></a>}}
\if{latex}{\out{\hypertarget{method-set_user_attribute_group_values}{}}}
\subsection{Method \code{set_user_attribute_group_values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserAttributeApi$set_user_attribute_group_values(user_attribute_id, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_user_attribute"></a>}}
\if{latex}{\out{\hypertarget{method-update_user_attribute}{}}}
\subsection{Method \code{update_user_attribute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserAttributeApi$update_user_attribute(user_attribute_id, body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-user_attribute"></a>}}
\if{latex}{\out{\hypertarget{method-user_attribute}{}}}
\subsection{Method \code{user_attribute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserAttributeApi$user_attribute(user_attribute_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserAttributeApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
