% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserAttributeGroupValue.r
\docType{data}
\name{UserAttributeGroupValue}
\alias{UserAttributeGroupValue}
\title{UserAttributeGroupValue Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
UserAttributeGroupValue
}
\description{
UserAttributeGroupValue Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{group_id}}{}

\item{\code{user_attribute_id}}{}

\item{\code{value_is_hidden}}{}

\item{\code{rank}}{}

\item{\code{value}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
