% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserAttributeWithValue.r
\name{UserAttributeWithValue}
\alias{UserAttributeWithValue}
\title{UserAttributeWithValue Class}
\description{
UserAttributeWithValue Class

UserAttributeWithValue Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{}

\item{\code{label}}{}

\item{\code{rank}}{}

\item{\code{value}}{}

\item{\code{user_id}}{}

\item{\code{user_can_edit}}{}

\item{\code{value_is_hidden}}{}

\item{\code{user_attribute_id}}{}

\item{\code{source}}{}

\item{\code{can}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UserAttributeWithValue$new()}}
\item \href{#method-toJSON}{\code{UserAttributeWithValue$toJSON()}}
\item \href{#method-fromJSON}{\code{UserAttributeWithValue$fromJSON()}}
\item \href{#method-toJSONString}{\code{UserAttributeWithValue$toJSONString()}}
\item \href{#method-fromJSONString}{\code{UserAttributeWithValue$fromJSONString()}}
\item \href{#method-clone}{\code{UserAttributeWithValue$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserAttributeWithValue$new(
  name,
  label,
  rank,
  value,
  user_id,
  user_can_edit,
  value_is_hidden,
  user_attribute_id,
  source,
  can
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserAttributeWithValue$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserAttributeWithValue$fromJSON(UserAttributeWithValueJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserAttributeWithValue$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserAttributeWithValue$fromJSONString(UserAttributeWithValueJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserAttributeWithValue$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
