% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserLoginLockout.r
\name{UserLoginLockout}
\alias{UserLoginLockout}
\title{UserLoginLockout Class}
\description{
UserLoginLockout Class

UserLoginLockout Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{can}}{}

\item{\code{key}}{}

\item{\code{auth_type}}{}

\item{\code{ip}}{}

\item{\code{user_id}}{}

\item{\code{remote_id}}{}

\item{\code{full_name}}{}

\item{\code{email}}{}

\item{\code{fail_count}}{}

\item{\code{lockout_at}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UserLoginLockout$new()}}
\item \href{#method-toJSON}{\code{UserLoginLockout$toJSON()}}
\item \href{#method-fromJSONObject}{\code{UserLoginLockout$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{UserLoginLockout$fromJSON()}}
\item \href{#method-toJSONString}{\code{UserLoginLockout$toJSONString()}}
\item \href{#method-fromJSONString}{\code{UserLoginLockout$fromJSONString()}}
\item \href{#method-clone}{\code{UserLoginLockout$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserLoginLockout$new(
  can,
  key,
  auth_type,
  ip,
  user_id,
  remote_id,
  full_name,
  email,
  fail_count,
  lockout_at
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserLoginLockout$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserLoginLockout$fromJSONObject(UserLoginLockoutJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserLoginLockout$fromJSON(UserLoginLockoutJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserLoginLockout$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserLoginLockout$fromJSONString(UserLoginLockoutJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserLoginLockout$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
