% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserPublic.r
\docType{data}
\name{UserPublic}
\alias{UserPublic}
\title{UserPublic Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
UserPublic
}
\description{
UserPublic Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{first_name}}{}

\item{\code{last_name}}{}

\item{\code{display_name}}{}

\item{\code{avatar_url}}{}

\item{\code{url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
