% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserSession.R
\name{UserSession}
\alias{UserSession}
\title{UserSession}
\description{
Keeps track of the client objects for each binding (LookApi, QueryApi, etc).
This shouldn't generally be called directly.
A new \code{UserSession} will be generated when you created a new \code{LookerSDK}.
}
