% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidationError.r
\docType{data}
\name{ValidationError}
\alias{ValidationError}
\title{ValidationError Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ValidationError
}
\description{
ValidationError Class
}
\section{Fields}{

\describe{
\item{\code{message}}{}

\item{\code{errors}}{}

\item{\code{documentation_url}}{}
}}

\keyword{datasets}
