% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidationErrorDetail.r
\docType{data}
\name{ValidationErrorDetail}
\alias{ValidationErrorDetail}
\title{ValidationErrorDetail Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ValidationErrorDetail
}
\description{
ValidationErrorDetail Class
}
\section{Fields}{

\describe{
\item{\code{field}}{}

\item{\code{code}}{}

\item{\code{message}}{}

\item{\code{documentation_url}}{}
}}

\keyword{datasets}
