% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Whitelabel.r
\docType{data}
\name{Whitelabel}
\alias{Whitelabel}
\title{Whitelabel Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Whitelabel
}
\description{
Whitelabel Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{logo_file}}{}

\item{\code{logo_url}}{}

\item{\code{favicon_file}}{}

\item{\code{favicon_url}}{}

\item{\code{default_title}}{}

\item{\code{show_help_menu}}{}

\item{\code{show_docs}}{}

\item{\code{show_email_sub_options}}{}

\item{\code{allow_looker_mentions}}{}

\item{\code{allow_looker_links}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
