% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkspaceApi.r
\docType{data}
\name{WorkspaceApi}
\alias{WorkspaceApi}
\title{Workspace operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
WorkspaceApi
}
\description{
looker.Workspace
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

all_workspaces Get All Workspaces 


workspace Get Workspace 

}
}

\keyword{datasets}
