% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteScheduledPlan.r
\name{WriteScheduledPlan}
\alias{WriteScheduledPlan}
\title{WriteScheduledPlan Class}
\description{
WriteScheduledPlan Class

WriteScheduledPlan Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{}

\item{\code{user_id}}{}

\item{\code{run_as_recipient}}{}

\item{\code{enabled}}{}

\item{\code{look_id}}{}

\item{\code{dashboard_id}}{}

\item{\code{lookml_dashboard_id}}{}

\item{\code{filters_string}}{}

\item{\code{dashboard_filters}}{}

\item{\code{require_results}}{}

\item{\code{require_no_results}}{}

\item{\code{require_change}}{}

\item{\code{send_all_results}}{}

\item{\code{crontab}}{}

\item{\code{datagroup}}{}

\item{\code{timezone}}{}

\item{\code{query_id}}{}

\item{\code{scheduled_plan_destination}}{}

\item{\code{run_once}}{}

\item{\code{include_links}}{}

\item{\code{pdf_paper_size}}{}

\item{\code{pdf_landscape}}{}

\item{\code{embed}}{}

\item{\code{color_theme}}{}

\item{\code{long_tables}}{}

\item{\code{inline_table_width}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{WriteScheduledPlan$new()}}
\item \href{#method-toJSON}{\code{WriteScheduledPlan$toJSON()}}
\item \href{#method-fromJSONObject}{\code{WriteScheduledPlan$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{WriteScheduledPlan$fromJSON()}}
\item \href{#method-toJSONString}{\code{WriteScheduledPlan$toJSONString()}}
\item \href{#method-fromJSONString}{\code{WriteScheduledPlan$fromJSONString()}}
\item \href{#method-clone}{\code{WriteScheduledPlan$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WriteScheduledPlan$new(
  name,
  user_id,
  run_as_recipient,
  enabled,
  look_id,
  dashboard_id,
  lookml_dashboard_id,
  filters_string,
  dashboard_filters,
  require_results,
  require_no_results,
  require_change,
  send_all_results,
  crontab,
  datagroup,
  timezone,
  query_id,
  scheduled_plan_destination,
  run_once,
  include_links,
  pdf_paper_size,
  pdf_landscape,
  embed,
  color_theme,
  long_tables,
  inline_table_width
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WriteScheduledPlan$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WriteScheduledPlan$fromJSONObject(WriteScheduledPlanJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WriteScheduledPlan$fromJSON(WriteScheduledPlanJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WriteScheduledPlan$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WriteScheduledPlan$fromJSONString(WriteScheduledPlanJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WriteScheduledPlan$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
