% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{AdjustTimingMarks}
\alias{AdjustTimingMarks}
\title{Adjust the event-timing attributes of a trial}
\usage{
AdjustTimingMarks(trial, alignment_event)
}
\arguments{
\item{trial}{a \code{Trial} object}

\item{alignment_event}{the name of the Trial attribute to be used as time 0.
The default is \code{"TargetOnset"}.}
}
\value{
the inputted trial object with updated timing attributes
}
\description{
Adjust the event-timing attributes of a trial
}
\keyword{internal}
