% Generated by roxygen2 (4.0.0): do not edit by hand
\name{AOI}
\alias{AOI}
\title{Convert pixel locations of an image to proportion-on-screen location}
\usage{
AOI(x_pix, y_pix)
}
\arguments{
\item{x_pix}{A numeric vector whose elements determine the left and right
boundaries of the image.}

\item{y_pix}{A numeric vector whose elements determine the lower and upper
boundaries of the image.}
}
\value{
A list of the form \code{list(x = (left_prop, right_prop), y =
  (lower_prop, upper_prop))}, where \code{left_prop}, \code{right_prop},
  \code{lower_prop}, and \code{upper_prop} are the screen proportion of the
  left, right, lower, and upper boundaries, respectively.
}
\description{
AOI is a utility function for converting the pixel locations of the x- and
y-boundaries of an image to screen proportions that can be passed to
functions like AddAOIData.
}
\keyword{internal}

