% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aoi_logodds.R
\name{AOILogOdds}
\alias{AOILogOdds}
\title{Compute Log-Odds of Looking to Target versus Non-target Areas of Interest.}
\usage{
AOILogOdds(trials, aoi_column = "GazeByImageAOI", target_aoi = c("Target"),
  distractor_aoi = c("SemanticFoil", "PhonologicalFoil", "Unrelated"),
  bin_width = 3)
}
\arguments{
\item{trials}{A list of \code{Trial} objects that have been aligned (via
\code{AlignTrials}) and had discretized AOI data added (via
\code{AddAOIData}).}

\item{aoi_column}{A character vector whose elements each name a column of
discretized AOI data in all the \code{Trial} objects in \code{trials}.
Default is \code{GazeByImageAOI}.}

\item{target_aoi}{A vector that specifies which AOI codes should be counted
as looks to target---i.e., which AOI codes should be summed as the
numerator of the log odds. Default is \code{Target}, which counts only the
looks to the target image in the \code{GazeByImageAOI} discretization.}

\item{distractor_aoi}{Vector that specifies which AOIs should be counted as
looks to the "distractor" image, or perhaps more correctly, looks away from
target---i.e., which AOI codes should be summed as the denominator of the
log odds.}

\item{bin_width}{An integer specifying the number of adjacent frames that
should be binned and used to compute the log odds of the discretized data.}
}
\value{
A 2--column \code{data.frame} of binned gaze data enumerating (1) the
  mean time of each bin and (2) the log-odds of looks to each type of
  discretized AOI data specified by \code{aoi_column}.
}
\description{
Compute Log-Odds of Looking to Target versus Non-target Areas of Interest.
}
\details{
The value of the log-odds for each time-point is found by:

\enumerate{
  \item binning (down-sampling) the time-points in each trial;
  \item summing all the looks to the "target" AOI in each bin across all
    trials in \code{trials} (call this \code{target_sum});
  \item summing all the looks to the "distractor" AOI in each bin across all
    trials (call this \code{distract_sum});
  \item correcting zero values of \code{target_sum} and \code{distract_sum};
  \item computing \code{log(target_sum / distract_sum)} for each bin.
}

The times at which each log-odds is defined are found by taking the mean time
value of each bin. The times and values of the log odds function(s) are
collected into a \code{data.frame}. The first column of the \code{data.frame}
is the times, each other column is a log-odds.
}

