% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApiAuthApi.r
\name{ApiAuthApi}
\alias{ApiAuthApi}
\title{ApiAuth operations}
\description{
looker.ApiAuth
}
\section{Methods}{

\describe{

login Login


login_user Login user


logout Logout

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ApiAuthApi$new()}}
\item \href{#method-login}{\code{ApiAuthApi$login()}}
\item \href{#method-login_user}{\code{ApiAuthApi$login_user()}}
\item \href{#method-logout}{\code{ApiAuthApi$logout()}}
\item \href{#method-clone}{\code{ApiAuthApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiAuthApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-login"></a>}}
\if{latex}{\out{\hypertarget{method-login}{}}}
\subsection{Method \code{login()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiAuthApi$login(client_id, client_secret, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-login_user"></a>}}
\if{latex}{\out{\hypertarget{method-login_user}{}}}
\subsection{Method \code{login_user()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiAuthApi$login_user(user_id, associative, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logout"></a>}}
\if{latex}{\out{\hypertarget{method-logout}{}}}
\subsection{Method \code{logout()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiAuthApi$logout(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiAuthApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
