% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/time.R
\name{AssignNewTimes}
\alias{AssignNewTimes}
\title{Assign new times to a trial once it has been aligned at a reference frame}
\usage{
AssignNewTimes(trial, zero_frame, frame_rate = lwl_constants$ms_per_frame)
}
\arguments{
\item{trial}{a Trial object}

\item{frame_rate}{the eye-tracker sampling rate. Defaults to
\code{lwl_constants$ms_per_frame} which is 16.6546ms.}

\item{the}{index of the reference frame that will mark time = 0 in the
aligned trial}
}
\value{
the inputted trial object with updated values in its \code{Time}
  column and an added \code{FrameRate} attribute
}
\description{
Assign new times to a trial once it has been aligned at a reference frame
}

