% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthApi.r
\name{AuthApi}
\alias{AuthApi}
\title{Auth operations}
\description{
looker.Auth
}
\section{Methods}{

\describe{

activate_app_user Activate OAuth App User


all_oauth_client_apps Get All OAuth Client Apps


all_user_login_lockouts Get All User Login Lockouts


create_embed_url_as_me Create Embed URL


create_oidc_test_config Create OIDC Test Configuration


create_saml_test_config Create SAML Test Configuration


create_sso_embed_url Create SSO Embed Url


deactivate_app_user Deactivate OAuth App User


delete_oauth_client_app Delete OAuth Client App


delete_oidc_test_config Delete OIDC Test Configuration


delete_saml_test_config Delete SAML Test Configuration


delete_user_login_lockout Delete User Login Lockout


fetch_and_parse_saml_idp_metadata Parse SAML IdP Url


force_password_reset_at_next_login_for_all_users Force password reset


invalidate_tokens Invalidate Tokens


ldap_config Get LDAP Configuration


oauth_client_app Get OAuth Client App


oidc_config Get OIDC Configuration


oidc_test_config Get OIDC Test Configuration


parse_saml_idp_metadata Parse SAML IdP XML


password_config Get Password Config


register_oauth_client_app Register OAuth App


saml_config Get SAML Configuration


saml_test_config Get SAML Test Configuration


search_user_login_lockouts Search User Login Lockouts


session_config Get Session Config


test_ldap_config_auth Test LDAP Auth


test_ldap_config_connection Test LDAP Connection


test_ldap_config_user_auth Test LDAP User Auth


test_ldap_config_user_info Test LDAP User Info


update_ldap_config Update LDAP Configuration


update_oauth_client_app Update OAuth App


update_oidc_config Update OIDC Configuration


update_password_config Update Password Config


update_saml_config Update SAML Configuration


update_session_config Update Session Config

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AuthApi$new()}}
\item \href{#method-activate_app_user}{\code{AuthApi$activate_app_user()}}
\item \href{#method-all_oauth_client_apps}{\code{AuthApi$all_oauth_client_apps()}}
\item \href{#method-all_user_login_lockouts}{\code{AuthApi$all_user_login_lockouts()}}
\item \href{#method-create_embed_url_as_me}{\code{AuthApi$create_embed_url_as_me()}}
\item \href{#method-create_oidc_test_config}{\code{AuthApi$create_oidc_test_config()}}
\item \href{#method-create_saml_test_config}{\code{AuthApi$create_saml_test_config()}}
\item \href{#method-create_sso_embed_url}{\code{AuthApi$create_sso_embed_url()}}
\item \href{#method-deactivate_app_user}{\code{AuthApi$deactivate_app_user()}}
\item \href{#method-delete_oauth_client_app}{\code{AuthApi$delete_oauth_client_app()}}
\item \href{#method-delete_oidc_test_config}{\code{AuthApi$delete_oidc_test_config()}}
\item \href{#method-delete_saml_test_config}{\code{AuthApi$delete_saml_test_config()}}
\item \href{#method-delete_user_login_lockout}{\code{AuthApi$delete_user_login_lockout()}}
\item \href{#method-fetch_and_parse_saml_idp_metadata}{\code{AuthApi$fetch_and_parse_saml_idp_metadata()}}
\item \href{#method-force_password_reset_at_next_login_for_all_users}{\code{AuthApi$force_password_reset_at_next_login_for_all_users()}}
\item \href{#method-invalidate_tokens}{\code{AuthApi$invalidate_tokens()}}
\item \href{#method-ldap_config}{\code{AuthApi$ldap_config()}}
\item \href{#method-oauth_client_app}{\code{AuthApi$oauth_client_app()}}
\item \href{#method-oidc_config}{\code{AuthApi$oidc_config()}}
\item \href{#method-oidc_test_config}{\code{AuthApi$oidc_test_config()}}
\item \href{#method-parse_saml_idp_metadata}{\code{AuthApi$parse_saml_idp_metadata()}}
\item \href{#method-password_config}{\code{AuthApi$password_config()}}
\item \href{#method-register_oauth_client_app}{\code{AuthApi$register_oauth_client_app()}}
\item \href{#method-saml_config}{\code{AuthApi$saml_config()}}
\item \href{#method-saml_test_config}{\code{AuthApi$saml_test_config()}}
\item \href{#method-search_user_login_lockouts}{\code{AuthApi$search_user_login_lockouts()}}
\item \href{#method-session_config}{\code{AuthApi$session_config()}}
\item \href{#method-test_ldap_config_auth}{\code{AuthApi$test_ldap_config_auth()}}
\item \href{#method-test_ldap_config_connection}{\code{AuthApi$test_ldap_config_connection()}}
\item \href{#method-test_ldap_config_user_auth}{\code{AuthApi$test_ldap_config_user_auth()}}
\item \href{#method-test_ldap_config_user_info}{\code{AuthApi$test_ldap_config_user_info()}}
\item \href{#method-update_ldap_config}{\code{AuthApi$update_ldap_config()}}
\item \href{#method-update_oauth_client_app}{\code{AuthApi$update_oauth_client_app()}}
\item \href{#method-update_oidc_config}{\code{AuthApi$update_oidc_config()}}
\item \href{#method-update_password_config}{\code{AuthApi$update_password_config()}}
\item \href{#method-update_saml_config}{\code{AuthApi$update_saml_config()}}
\item \href{#method-update_session_config}{\code{AuthApi$update_session_config()}}
\item \href{#method-clone}{\code{AuthApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-activate_app_user"></a>}}
\if{latex}{\out{\hypertarget{method-activate_app_user}{}}}
\subsection{Method \code{activate_app_user()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$activate_app_user(client_guid, user_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_oauth_client_apps"></a>}}
\if{latex}{\out{\hypertarget{method-all_oauth_client_apps}{}}}
\subsection{Method \code{all_oauth_client_apps()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$all_oauth_client_apps(fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_user_login_lockouts"></a>}}
\if{latex}{\out{\hypertarget{method-all_user_login_lockouts}{}}}
\subsection{Method \code{all_user_login_lockouts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$all_user_login_lockouts(fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_embed_url_as_me"></a>}}
\if{latex}{\out{\hypertarget{method-create_embed_url_as_me}{}}}
\subsection{Method \code{create_embed_url_as_me()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$create_embed_url_as_me(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_oidc_test_config"></a>}}
\if{latex}{\out{\hypertarget{method-create_oidc_test_config}{}}}
\subsection{Method \code{create_oidc_test_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$create_oidc_test_config(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_saml_test_config"></a>}}
\if{latex}{\out{\hypertarget{method-create_saml_test_config}{}}}
\subsection{Method \code{create_saml_test_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$create_saml_test_config(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_sso_embed_url"></a>}}
\if{latex}{\out{\hypertarget{method-create_sso_embed_url}{}}}
\subsection{Method \code{create_sso_embed_url()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$create_sso_embed_url(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deactivate_app_user"></a>}}
\if{latex}{\out{\hypertarget{method-deactivate_app_user}{}}}
\subsection{Method \code{deactivate_app_user()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$deactivate_app_user(client_guid, user_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_oauth_client_app"></a>}}
\if{latex}{\out{\hypertarget{method-delete_oauth_client_app}{}}}
\subsection{Method \code{delete_oauth_client_app()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$delete_oauth_client_app(client_guid, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_oidc_test_config"></a>}}
\if{latex}{\out{\hypertarget{method-delete_oidc_test_config}{}}}
\subsection{Method \code{delete_oidc_test_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$delete_oidc_test_config(test_slug, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_saml_test_config"></a>}}
\if{latex}{\out{\hypertarget{method-delete_saml_test_config}{}}}
\subsection{Method \code{delete_saml_test_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$delete_saml_test_config(test_slug, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_user_login_lockout"></a>}}
\if{latex}{\out{\hypertarget{method-delete_user_login_lockout}{}}}
\subsection{Method \code{delete_user_login_lockout()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$delete_user_login_lockout(key, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fetch_and_parse_saml_idp_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-fetch_and_parse_saml_idp_metadata}{}}}
\subsection{Method \code{fetch_and_parse_saml_idp_metadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$fetch_and_parse_saml_idp_metadata(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-force_password_reset_at_next_login_for_all_users"></a>}}
\if{latex}{\out{\hypertarget{method-force_password_reset_at_next_login_for_all_users}{}}}
\subsection{Method \code{force_password_reset_at_next_login_for_all_users()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$force_password_reset_at_next_login_for_all_users(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-invalidate_tokens"></a>}}
\if{latex}{\out{\hypertarget{method-invalidate_tokens}{}}}
\subsection{Method \code{invalidate_tokens()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$invalidate_tokens(client_guid, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ldap_config"></a>}}
\if{latex}{\out{\hypertarget{method-ldap_config}{}}}
\subsection{Method \code{ldap_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$ldap_config(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-oauth_client_app"></a>}}
\if{latex}{\out{\hypertarget{method-oauth_client_app}{}}}
\subsection{Method \code{oauth_client_app()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$oauth_client_app(client_guid, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-oidc_config"></a>}}
\if{latex}{\out{\hypertarget{method-oidc_config}{}}}
\subsection{Method \code{oidc_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$oidc_config(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-oidc_test_config"></a>}}
\if{latex}{\out{\hypertarget{method-oidc_test_config}{}}}
\subsection{Method \code{oidc_test_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$oidc_test_config(test_slug, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-parse_saml_idp_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-parse_saml_idp_metadata}{}}}
\subsection{Method \code{parse_saml_idp_metadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$parse_saml_idp_metadata(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-password_config"></a>}}
\if{latex}{\out{\hypertarget{method-password_config}{}}}
\subsection{Method \code{password_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$password_config(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-register_oauth_client_app"></a>}}
\if{latex}{\out{\hypertarget{method-register_oauth_client_app}{}}}
\subsection{Method \code{register_oauth_client_app()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$register_oauth_client_app(client_guid, body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saml_config"></a>}}
\if{latex}{\out{\hypertarget{method-saml_config}{}}}
\subsection{Method \code{saml_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$saml_config(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saml_test_config"></a>}}
\if{latex}{\out{\hypertarget{method-saml_test_config}{}}}
\subsection{Method \code{saml_test_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$saml_test_config(test_slug, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_user_login_lockouts"></a>}}
\if{latex}{\out{\hypertarget{method-search_user_login_lockouts}{}}}
\subsection{Method \code{search_user_login_lockouts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$search_user_login_lockouts(
  fields,
  page,
  per_page,
  sorts,
  auth_type,
  full_name,
  email,
  remote_id,
  filter_or,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-session_config"></a>}}
\if{latex}{\out{\hypertarget{method-session_config}{}}}
\subsection{Method \code{session_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$session_config(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-test_ldap_config_auth"></a>}}
\if{latex}{\out{\hypertarget{method-test_ldap_config_auth}{}}}
\subsection{Method \code{test_ldap_config_auth()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$test_ldap_config_auth(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-test_ldap_config_connection"></a>}}
\if{latex}{\out{\hypertarget{method-test_ldap_config_connection}{}}}
\subsection{Method \code{test_ldap_config_connection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$test_ldap_config_connection(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-test_ldap_config_user_auth"></a>}}
\if{latex}{\out{\hypertarget{method-test_ldap_config_user_auth}{}}}
\subsection{Method \code{test_ldap_config_user_auth()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$test_ldap_config_user_auth(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-test_ldap_config_user_info"></a>}}
\if{latex}{\out{\hypertarget{method-test_ldap_config_user_info}{}}}
\subsection{Method \code{test_ldap_config_user_info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$test_ldap_config_user_info(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_ldap_config"></a>}}
\if{latex}{\out{\hypertarget{method-update_ldap_config}{}}}
\subsection{Method \code{update_ldap_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$update_ldap_config(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_oauth_client_app"></a>}}
\if{latex}{\out{\hypertarget{method-update_oauth_client_app}{}}}
\subsection{Method \code{update_oauth_client_app()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$update_oauth_client_app(client_guid, body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_oidc_config"></a>}}
\if{latex}{\out{\hypertarget{method-update_oidc_config}{}}}
\subsection{Method \code{update_oidc_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$update_oidc_config(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_password_config"></a>}}
\if{latex}{\out{\hypertarget{method-update_password_config}{}}}
\subsection{Method \code{update_password_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$update_password_config(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_saml_config"></a>}}
\if{latex}{\out{\hypertarget{method-update_saml_config}{}}}
\subsection{Method \code{update_saml_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$update_saml_config(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_session_config"></a>}}
\if{latex}{\out{\hypertarget{method-update_session_config}{}}}
\subsection{Method \code{update_session_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$update_session_config(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
