% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BackupConfiguration.r
\docType{data}
\name{BackupConfiguration}
\alias{BackupConfiguration}
\title{BackupConfiguration Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BackupConfiguration
}
\description{
BackupConfiguration Class
}
\section{Fields}{

\describe{
\item{\code{type}}{}

\item{\code{custom_s3_bucket}}{}

\item{\code{custom_s3_bucket_region}}{}

\item{\code{custom_s3_key}}{}

\item{\code{custom_s3_secret}}{}

\item{\code{url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
