% Generated by roxygen2 (4.0.1): do not edit by hand
\name{CalculateMistrackings}
\alias{CalculateMistrackings}
\title{Calculate the amount of missing data in a list of Trials}
\usage{
CalculateMistrackings(trials, start = NULL, end = NULL,
  column = "GazeByImageAOI")
}
\arguments{
\item{trials}{A list of Trial objects.}

\item{start}{A time value or the name of a time attribute of a Trial. Default
is to use the whole trial.}

\item{end}{A time value or the name of a time attribute of a Trial. Default
is to use the whole trial.}

\item{column}{A character string that names one of the columns of Gazedata in
the trial object. Default is `"GazeByImageAOI"`.}
}
\value{
A list of Trial objects that has been updated to include attributes containing percentage of mistracked data.
}
\description{
`start` and `end` can use the following strings to pull time attributes such
as `ImageOnset`, `CarrierOnset`, `CarrierEnd`, `TargetOnset`, `TargetEnd`,
`AttentionOnset`, and `AttentionEnd`.
}

