% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{CalculateMistrackings}
\alias{CalculateMistrackings}
\title{Calculate the amount of missing data in a list of Trials}
\usage{
CalculateMistrackings(x, column = "GazeByImageAOI")
}
\arguments{
\item{x}{a Trial with a \code{GazeByImageAOI} column or a TrialList where
each Trial has a \code{GazeByImageAOI} column}

\item{column}{A character string that names one of the columns of Gazedata in
the trial object. Default is `"GazeByImageAOI"`.}
}
\value{
the Trial object(s) updated to include attributes about the amount of
  missing data
}
\description{
Calculate the amount of missing data in a list of Trials
}
