% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{ComputeGazeContact}
\alias{ComputeGazeContact}
\title{Compute the average duration of tracked gazes}
\usage{
ComputeGazeContact(x)
}
\arguments{
\item{x}{a Trial with a \code{GazeByImageAOI} column or a TrialList where
each Trial has a \code{GazeByImageAOI} column}
}
\value{
the Trial object(s) updated to include the attribute
  \code{AvgGazeContact}
}
\description{
To quantify "flicker" in eye-tracking data, Wass, Smith and Johnson (2013)
found the average duration of tracked segments of eye-tracking data. This
function performs this calculation, updating trials with a new attribute
\code{AvgGazeContact}. This value measures the average duration (ms) of each
segment of uninterrupted eye-tracking data in a trial.
}

