% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfigApi.r
\name{ConfigApi}
\alias{ConfigApi}
\title{Config operations}
\description{
looker.Config
}
\section{Methods}{

\describe{

all_legacy_features Get All Legacy Features


all_locales Get All Locales


all_timezones Get All Timezones


api_spec Get an API specification


cloud_storage_configuration Get Cloud Storage


create_digest_email_send Deliver digest email contents


custom_welcome_email Get Custom Welcome Email


digest_emails_enabled Get Digest_emails


get_setting Get Setting


internal_help_resources Get Internal Help Resources


internal_help_resources_content Get Internal Help Resources Content


legacy_feature Get Legacy Feature


mobile_settings Get Mobile_Settings


set_setting Set Setting


update_cloud_storage_configuration Update Cloud Storage


update_custom_welcome_email Update Custom Welcome Email Content


update_custom_welcome_email_test Send a test welcome email to the currently logged in user with the supplied content 


update_digest_emails_enabled Update Digest_emails


update_internal_help_resources Update internal help resources configuration


update_internal_help_resources_content Update internal help resources content


update_legacy_feature Update Legacy Feature


update_whitelabel_configuration Update Whitelabel configuration


versions Get ApiVersion


whitelabel_configuration Get Whitelabel configuration

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ConfigApi$new()}}
\item \href{#method-all_legacy_features}{\code{ConfigApi$all_legacy_features()}}
\item \href{#method-all_locales}{\code{ConfigApi$all_locales()}}
\item \href{#method-all_timezones}{\code{ConfigApi$all_timezones()}}
\item \href{#method-api_spec}{\code{ConfigApi$api_spec()}}
\item \href{#method-cloud_storage_configuration}{\code{ConfigApi$cloud_storage_configuration()}}
\item \href{#method-create_digest_email_send}{\code{ConfigApi$create_digest_email_send()}}
\item \href{#method-custom_welcome_email}{\code{ConfigApi$custom_welcome_email()}}
\item \href{#method-digest_emails_enabled}{\code{ConfigApi$digest_emails_enabled()}}
\item \href{#method-get_setting}{\code{ConfigApi$get_setting()}}
\item \href{#method-internal_help_resources}{\code{ConfigApi$internal_help_resources()}}
\item \href{#method-internal_help_resources_content}{\code{ConfigApi$internal_help_resources_content()}}
\item \href{#method-legacy_feature}{\code{ConfigApi$legacy_feature()}}
\item \href{#method-mobile_settings}{\code{ConfigApi$mobile_settings()}}
\item \href{#method-set_setting}{\code{ConfigApi$set_setting()}}
\item \href{#method-update_cloud_storage_configuration}{\code{ConfigApi$update_cloud_storage_configuration()}}
\item \href{#method-update_custom_welcome_email}{\code{ConfigApi$update_custom_welcome_email()}}
\item \href{#method-update_custom_welcome_email_test}{\code{ConfigApi$update_custom_welcome_email_test()}}
\item \href{#method-update_digest_emails_enabled}{\code{ConfigApi$update_digest_emails_enabled()}}
\item \href{#method-update_internal_help_resources}{\code{ConfigApi$update_internal_help_resources()}}
\item \href{#method-update_internal_help_resources_content}{\code{ConfigApi$update_internal_help_resources_content()}}
\item \href{#method-update_legacy_feature}{\code{ConfigApi$update_legacy_feature()}}
\item \href{#method-update_whitelabel_configuration}{\code{ConfigApi$update_whitelabel_configuration()}}
\item \href{#method-versions}{\code{ConfigApi$versions()}}
\item \href{#method-whitelabel_configuration}{\code{ConfigApi$whitelabel_configuration()}}
\item \href{#method-clone}{\code{ConfigApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_legacy_features"></a>}}
\if{latex}{\out{\hypertarget{method-all_legacy_features}{}}}
\subsection{Method \code{all_legacy_features()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$all_legacy_features(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_locales"></a>}}
\if{latex}{\out{\hypertarget{method-all_locales}{}}}
\subsection{Method \code{all_locales()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$all_locales(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_timezones"></a>}}
\if{latex}{\out{\hypertarget{method-all_timezones}{}}}
\subsection{Method \code{all_timezones()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$all_timezones(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-api_spec"></a>}}
\if{latex}{\out{\hypertarget{method-api_spec}{}}}
\subsection{Method \code{api_spec()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$api_spec(api_version, specification, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cloud_storage_configuration"></a>}}
\if{latex}{\out{\hypertarget{method-cloud_storage_configuration}{}}}
\subsection{Method \code{cloud_storage_configuration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$cloud_storage_configuration(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_digest_email_send"></a>}}
\if{latex}{\out{\hypertarget{method-create_digest_email_send}{}}}
\subsection{Method \code{create_digest_email_send()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$create_digest_email_send(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-custom_welcome_email"></a>}}
\if{latex}{\out{\hypertarget{method-custom_welcome_email}{}}}
\subsection{Method \code{custom_welcome_email()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$custom_welcome_email(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-digest_emails_enabled"></a>}}
\if{latex}{\out{\hypertarget{method-digest_emails_enabled}{}}}
\subsection{Method \code{digest_emails_enabled()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$digest_emails_enabled(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_setting"></a>}}
\if{latex}{\out{\hypertarget{method-get_setting}{}}}
\subsection{Method \code{get_setting()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$get_setting(fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-internal_help_resources"></a>}}
\if{latex}{\out{\hypertarget{method-internal_help_resources}{}}}
\subsection{Method \code{internal_help_resources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$internal_help_resources(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-internal_help_resources_content"></a>}}
\if{latex}{\out{\hypertarget{method-internal_help_resources_content}{}}}
\subsection{Method \code{internal_help_resources_content()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$internal_help_resources_content(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-legacy_feature"></a>}}
\if{latex}{\out{\hypertarget{method-legacy_feature}{}}}
\subsection{Method \code{legacy_feature()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$legacy_feature(legacy_feature_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mobile_settings"></a>}}
\if{latex}{\out{\hypertarget{method-mobile_settings}{}}}
\subsection{Method \code{mobile_settings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$mobile_settings(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_setting"></a>}}
\if{latex}{\out{\hypertarget{method-set_setting}{}}}
\subsection{Method \code{set_setting()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$set_setting(body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_cloud_storage_configuration"></a>}}
\if{latex}{\out{\hypertarget{method-update_cloud_storage_configuration}{}}}
\subsection{Method \code{update_cloud_storage_configuration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$update_cloud_storage_configuration(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_custom_welcome_email"></a>}}
\if{latex}{\out{\hypertarget{method-update_custom_welcome_email}{}}}
\subsection{Method \code{update_custom_welcome_email()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$update_custom_welcome_email(body, send_test_welcome_email, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_custom_welcome_email_test"></a>}}
\if{latex}{\out{\hypertarget{method-update_custom_welcome_email_test}{}}}
\subsection{Method \code{update_custom_welcome_email_test()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$update_custom_welcome_email_test(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_digest_emails_enabled"></a>}}
\if{latex}{\out{\hypertarget{method-update_digest_emails_enabled}{}}}
\subsection{Method \code{update_digest_emails_enabled()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$update_digest_emails_enabled(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_internal_help_resources"></a>}}
\if{latex}{\out{\hypertarget{method-update_internal_help_resources}{}}}
\subsection{Method \code{update_internal_help_resources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$update_internal_help_resources(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_internal_help_resources_content"></a>}}
\if{latex}{\out{\hypertarget{method-update_internal_help_resources_content}{}}}
\subsection{Method \code{update_internal_help_resources_content()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$update_internal_help_resources_content(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_legacy_feature"></a>}}
\if{latex}{\out{\hypertarget{method-update_legacy_feature}{}}}
\subsection{Method \code{update_legacy_feature()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$update_legacy_feature(legacy_feature_id, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_whitelabel_configuration"></a>}}
\if{latex}{\out{\hypertarget{method-update_whitelabel_configuration}{}}}
\subsection{Method \code{update_whitelabel_configuration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$update_whitelabel_configuration(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-versions"></a>}}
\if{latex}{\out{\hypertarget{method-versions}{}}}
\subsection{Method \code{versions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$versions(fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-whitelabel_configuration"></a>}}
\if{latex}{\out{\hypertarget{method-whitelabel_configuration}{}}}
\subsection{Method \code{whitelabel_configuration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$whitelabel_configuration(fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
