% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfigApi.r
\name{ConfigApi}
\alias{ConfigApi}
\title{Config operations}
\description{
looker.Config
}
\section{Methods}{

\describe{

all_legacy_features Get All Legacy Features 


all_timezones Get All Timezones 


backup_configuration Get Backup Configuration 


legacy_feature Get Legacy Feature 


update_backup_configuration Update Backup Configuration 


update_legacy_feature Update Legacy Feature 


update_whitelabel_configuration Update Whitelabel configuration 


versions Get ApiVersion 


whitelabel_configuration Get Whitelabel configuration 

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ConfigApi$new()}}
\item \href{#method-all_legacy_features}{\code{ConfigApi$all_legacy_features()}}
\item \href{#method-all_timezones}{\code{ConfigApi$all_timezones()}}
\item \href{#method-backup_configuration}{\code{ConfigApi$backup_configuration()}}
\item \href{#method-legacy_feature}{\code{ConfigApi$legacy_feature()}}
\item \href{#method-update_backup_configuration}{\code{ConfigApi$update_backup_configuration()}}
\item \href{#method-update_legacy_feature}{\code{ConfigApi$update_legacy_feature()}}
\item \href{#method-update_whitelabel_configuration}{\code{ConfigApi$update_whitelabel_configuration()}}
\item \href{#method-versions}{\code{ConfigApi$versions()}}
\item \href{#method-whitelabel_configuration}{\code{ConfigApi$whitelabel_configuration()}}
\item \href{#method-clone}{\code{ConfigApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_legacy_features"></a>}}
\if{latex}{\out{\hypertarget{method-all_legacy_features}{}}}
\subsection{Method \code{all_legacy_features()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$all_legacy_features(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_timezones"></a>}}
\if{latex}{\out{\hypertarget{method-all_timezones}{}}}
\subsection{Method \code{all_timezones()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$all_timezones(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-backup_configuration"></a>}}
\if{latex}{\out{\hypertarget{method-backup_configuration}{}}}
\subsection{Method \code{backup_configuration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$backup_configuration(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-legacy_feature"></a>}}
\if{latex}{\out{\hypertarget{method-legacy_feature}{}}}
\subsection{Method \code{legacy_feature()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$legacy_feature(legacy_feature_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_backup_configuration"></a>}}
\if{latex}{\out{\hypertarget{method-update_backup_configuration}{}}}
\subsection{Method \code{update_backup_configuration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$update_backup_configuration(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_legacy_feature"></a>}}
\if{latex}{\out{\hypertarget{method-update_legacy_feature}{}}}
\subsection{Method \code{update_legacy_feature()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$update_legacy_feature(legacy_feature_id, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_whitelabel_configuration"></a>}}
\if{latex}{\out{\hypertarget{method-update_whitelabel_configuration}{}}}
\subsection{Method \code{update_whitelabel_configuration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$update_whitelabel_configuration(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-versions"></a>}}
\if{latex}{\out{\hypertarget{method-versions}{}}}
\subsection{Method \code{versions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$versions(fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-whitelabel_configuration"></a>}}
\if{latex}{\out{\hypertarget{method-whitelabel_configuration}{}}}
\subsection{Method \code{whitelabel_configuration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$whitelabel_configuration(fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
