% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectionApi.r
\name{ConnectionApi}
\alias{ConnectionApi}
\title{Connection operations}
\description{
looker.Connection
}
\section{Methods}{

\describe{

all_connections Get All Connections 


all_dialect_infos Get All Dialect Infos 


connection Get Connection 


create_connection Create Connection 


delete_connection Delete Connection 


test_connection Test Connection 


test_connection_config Test Connection Configuration 


update_connection Update Connection 

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ConnectionApi$new()}}
\item \href{#method-all_connections}{\code{ConnectionApi$all_connections()}}
\item \href{#method-all_dialect_infos}{\code{ConnectionApi$all_dialect_infos()}}
\item \href{#method-connection}{\code{ConnectionApi$connection()}}
\item \href{#method-create_connection}{\code{ConnectionApi$create_connection()}}
\item \href{#method-delete_connection}{\code{ConnectionApi$delete_connection()}}
\item \href{#method-test_connection}{\code{ConnectionApi$test_connection()}}
\item \href{#method-test_connection_config}{\code{ConnectionApi$test_connection_config()}}
\item \href{#method-update_connection}{\code{ConnectionApi$update_connection()}}
\item \href{#method-clone}{\code{ConnectionApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_connections"></a>}}
\if{latex}{\out{\hypertarget{method-all_connections}{}}}
\subsection{Method \code{all_connections()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$all_connections(fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_dialect_infos"></a>}}
\if{latex}{\out{\hypertarget{method-all_dialect_infos}{}}}
\subsection{Method \code{all_dialect_infos()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$all_dialect_infos(fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connection"></a>}}
\if{latex}{\out{\hypertarget{method-connection}{}}}
\subsection{Method \code{connection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$connection(connection_name, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_connection"></a>}}
\if{latex}{\out{\hypertarget{method-create_connection}{}}}
\subsection{Method \code{create_connection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$create_connection(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_connection"></a>}}
\if{latex}{\out{\hypertarget{method-delete_connection}{}}}
\subsection{Method \code{delete_connection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$delete_connection(connection_name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-test_connection"></a>}}
\if{latex}{\out{\hypertarget{method-test_connection}{}}}
\subsection{Method \code{test_connection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$test_connection(connection_name, tests, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-test_connection_config"></a>}}
\if{latex}{\out{\hypertarget{method-test_connection_config}{}}}
\subsection{Method \code{test_connection_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$test_connection_config(body, tests, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_connection"></a>}}
\if{latex}{\out{\hypertarget{method-update_connection}{}}}
\subsection{Method \code{update_connection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$update_connection(connection_name, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
