% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectionApi.r
\name{ConnectionApi}
\alias{ConnectionApi}
\title{Connection operations}
\description{
looker.Connection
}
\section{Methods}{

\describe{

all_connections Get All Connections


all_dialect_infos Get All Dialect Infos


all_external_oauth_applications Get All External OAuth Applications


all_ssh_servers Get All SSH Servers


all_ssh_tunnels Get All SSH Tunnels


connection Get Connection


create_connection Create Connection


create_external_oauth_application Create External OAuth Application


create_oauth_application_user_state Create Create OAuth user state.


create_ssh_server Create SSH Server


create_ssh_tunnel Create SSH Tunnel


delete_connection Delete Connection


delete_connection_override Delete Connection Override


delete_ssh_server Delete SSH Server


delete_ssh_tunnel Delete SSH Tunnel


ssh_public_key Get SSH Public Key


ssh_server Get SSH Server


ssh_tunnel Get SSH Tunnel


test_connection Test Connection


test_connection_config Test Connection Configuration


test_ssh_server Test SSH Server


test_ssh_tunnel Test SSH Tunnel


update_connection Update Connection


update_ssh_server Update SSH Server


update_ssh_tunnel Update SSH Tunnel

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ConnectionApi$new()}}
\item \href{#method-all_connections}{\code{ConnectionApi$all_connections()}}
\item \href{#method-all_dialect_infos}{\code{ConnectionApi$all_dialect_infos()}}
\item \href{#method-all_external_oauth_applications}{\code{ConnectionApi$all_external_oauth_applications()}}
\item \href{#method-all_ssh_servers}{\code{ConnectionApi$all_ssh_servers()}}
\item \href{#method-all_ssh_tunnels}{\code{ConnectionApi$all_ssh_tunnels()}}
\item \href{#method-connection}{\code{ConnectionApi$connection()}}
\item \href{#method-create_connection}{\code{ConnectionApi$create_connection()}}
\item \href{#method-create_external_oauth_application}{\code{ConnectionApi$create_external_oauth_application()}}
\item \href{#method-create_oauth_application_user_state}{\code{ConnectionApi$create_oauth_application_user_state()}}
\item \href{#method-create_ssh_server}{\code{ConnectionApi$create_ssh_server()}}
\item \href{#method-create_ssh_tunnel}{\code{ConnectionApi$create_ssh_tunnel()}}
\item \href{#method-delete_connection}{\code{ConnectionApi$delete_connection()}}
\item \href{#method-delete_connection_override}{\code{ConnectionApi$delete_connection_override()}}
\item \href{#method-delete_ssh_server}{\code{ConnectionApi$delete_ssh_server()}}
\item \href{#method-delete_ssh_tunnel}{\code{ConnectionApi$delete_ssh_tunnel()}}
\item \href{#method-ssh_public_key}{\code{ConnectionApi$ssh_public_key()}}
\item \href{#method-ssh_server}{\code{ConnectionApi$ssh_server()}}
\item \href{#method-ssh_tunnel}{\code{ConnectionApi$ssh_tunnel()}}
\item \href{#method-test_connection}{\code{ConnectionApi$test_connection()}}
\item \href{#method-test_connection_config}{\code{ConnectionApi$test_connection_config()}}
\item \href{#method-test_ssh_server}{\code{ConnectionApi$test_ssh_server()}}
\item \href{#method-test_ssh_tunnel}{\code{ConnectionApi$test_ssh_tunnel()}}
\item \href{#method-update_connection}{\code{ConnectionApi$update_connection()}}
\item \href{#method-update_ssh_server}{\code{ConnectionApi$update_ssh_server()}}
\item \href{#method-update_ssh_tunnel}{\code{ConnectionApi$update_ssh_tunnel()}}
\item \href{#method-clone}{\code{ConnectionApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_connections"></a>}}
\if{latex}{\out{\hypertarget{method-all_connections}{}}}
\subsection{Method \code{all_connections()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$all_connections(fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_dialect_infos"></a>}}
\if{latex}{\out{\hypertarget{method-all_dialect_infos}{}}}
\subsection{Method \code{all_dialect_infos()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$all_dialect_infos(fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_external_oauth_applications"></a>}}
\if{latex}{\out{\hypertarget{method-all_external_oauth_applications}{}}}
\subsection{Method \code{all_external_oauth_applications()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$all_external_oauth_applications(name, client_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_ssh_servers"></a>}}
\if{latex}{\out{\hypertarget{method-all_ssh_servers}{}}}
\subsection{Method \code{all_ssh_servers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$all_ssh_servers(fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_ssh_tunnels"></a>}}
\if{latex}{\out{\hypertarget{method-all_ssh_tunnels}{}}}
\subsection{Method \code{all_ssh_tunnels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$all_ssh_tunnels(fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connection"></a>}}
\if{latex}{\out{\hypertarget{method-connection}{}}}
\subsection{Method \code{connection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$connection(connection_name, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_connection"></a>}}
\if{latex}{\out{\hypertarget{method-create_connection}{}}}
\subsection{Method \code{create_connection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$create_connection(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_external_oauth_application"></a>}}
\if{latex}{\out{\hypertarget{method-create_external_oauth_application}{}}}
\subsection{Method \code{create_external_oauth_application()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$create_external_oauth_application(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_oauth_application_user_state"></a>}}
\if{latex}{\out{\hypertarget{method-create_oauth_application_user_state}{}}}
\subsection{Method \code{create_oauth_application_user_state()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$create_oauth_application_user_state(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_ssh_server"></a>}}
\if{latex}{\out{\hypertarget{method-create_ssh_server}{}}}
\subsection{Method \code{create_ssh_server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$create_ssh_server(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_ssh_tunnel"></a>}}
\if{latex}{\out{\hypertarget{method-create_ssh_tunnel}{}}}
\subsection{Method \code{create_ssh_tunnel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$create_ssh_tunnel(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_connection"></a>}}
\if{latex}{\out{\hypertarget{method-delete_connection}{}}}
\subsection{Method \code{delete_connection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$delete_connection(connection_name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_connection_override"></a>}}
\if{latex}{\out{\hypertarget{method-delete_connection_override}{}}}
\subsection{Method \code{delete_connection_override()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$delete_connection_override(
  connection_name,
  override_context,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_ssh_server"></a>}}
\if{latex}{\out{\hypertarget{method-delete_ssh_server}{}}}
\subsection{Method \code{delete_ssh_server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$delete_ssh_server(ssh_server_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_ssh_tunnel"></a>}}
\if{latex}{\out{\hypertarget{method-delete_ssh_tunnel}{}}}
\subsection{Method \code{delete_ssh_tunnel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$delete_ssh_tunnel(ssh_tunnel_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ssh_public_key"></a>}}
\if{latex}{\out{\hypertarget{method-ssh_public_key}{}}}
\subsection{Method \code{ssh_public_key()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$ssh_public_key(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ssh_server"></a>}}
\if{latex}{\out{\hypertarget{method-ssh_server}{}}}
\subsection{Method \code{ssh_server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$ssh_server(ssh_server_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ssh_tunnel"></a>}}
\if{latex}{\out{\hypertarget{method-ssh_tunnel}{}}}
\subsection{Method \code{ssh_tunnel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$ssh_tunnel(ssh_tunnel_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-test_connection"></a>}}
\if{latex}{\out{\hypertarget{method-test_connection}{}}}
\subsection{Method \code{test_connection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$test_connection(connection_name, tests, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-test_connection_config"></a>}}
\if{latex}{\out{\hypertarget{method-test_connection_config}{}}}
\subsection{Method \code{test_connection_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$test_connection_config(body, tests, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-test_ssh_server"></a>}}
\if{latex}{\out{\hypertarget{method-test_ssh_server}{}}}
\subsection{Method \code{test_ssh_server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$test_ssh_server(ssh_server_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-test_ssh_tunnel"></a>}}
\if{latex}{\out{\hypertarget{method-test_ssh_tunnel}{}}}
\subsection{Method \code{test_ssh_tunnel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$test_ssh_tunnel(ssh_tunnel_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_connection"></a>}}
\if{latex}{\out{\hypertarget{method-update_connection}{}}}
\subsection{Method \code{update_connection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$update_connection(connection_name, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_ssh_server"></a>}}
\if{latex}{\out{\hypertarget{method-update_ssh_server}{}}}
\subsection{Method \code{update_ssh_server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$update_ssh_server(ssh_server_id, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_ssh_tunnel"></a>}}
\if{latex}{\out{\hypertarget{method-update_ssh_tunnel}{}}}
\subsection{Method \code{update_ssh_tunnel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$update_ssh_tunnel(ssh_tunnel_id, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
