% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContentApi.r
\docType{data}
\name{ContentApi}
\alias{ContentApi}
\title{Content operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ContentApi
}
\description{
looker.Content
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

all_content_metadata_accesss Get All Content Metadata Accesss 


all_content_metadatas Get All Content Metadatas 


content_favorite Get Favorite Content 


content_metadata Get Content Metadata 


create_content_favorite Create Favorite Content 


create_content_metadata_access Create Content Metadata Access 


delete_content_favorite Delete Favorite Content 


delete_content_metadata_access Delete Content Metadata Access 


search_content_favorites Search Favorite Contents 


search_content_views Search Content Views 


update_content_metadata Update Content Metadata 


update_content_metadata_access Update Content Metadata Access 

}
}

\keyword{datasets}
