% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContentView.r
\docType{data}
\name{ContentView}
\alias{ContentView}
\title{ContentView Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ContentView
}
\description{
ContentView Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{look_id}}{}

\item{\code{dashboard_id}}{}

\item{\code{content_metadata_id}}{}

\item{\code{user_id}}{}

\item{\code{group_id}}{}

\item{\code{view_count}}{}

\item{\code{favorite_count}}{}

\item{\code{last_viewed_at}}{}

\item{\code{start_of_week_date}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
