% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CredentialsApi3.r
\docType{data}
\name{CredentialsApi3}
\alias{CredentialsApi3}
\title{CredentialsApi3 Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CredentialsApi3
}
\description{
CredentialsApi3 Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{client_id}}{}

\item{\code{created_at}}{}

\item{\code{is_disabled}}{}

\item{\code{type}}{}

\item{\code{url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
