% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CredentialsLDAP.r
\name{CredentialsLDAP}
\alias{CredentialsLDAP}
\title{CredentialsLDAP Class}
\description{
CredentialsLDAP Class

CredentialsLDAP Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{can}}{}

\item{\code{created_at}}{}

\item{\code{email}}{}

\item{\code{is_disabled}}{}

\item{\code{ldap_dn}}{}

\item{\code{ldap_id}}{}

\item{\code{logged_in_at}}{}

\item{\code{type}}{}

\item{\code{url}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CredentialsLDAP$new()}}
\item \href{#method-toJSON}{\code{CredentialsLDAP$toJSON()}}
\item \href{#method-fromJSONObject}{\code{CredentialsLDAP$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{CredentialsLDAP$fromJSON()}}
\item \href{#method-toJSONString}{\code{CredentialsLDAP$toJSONString()}}
\item \href{#method-fromJSONString}{\code{CredentialsLDAP$fromJSONString()}}
\item \href{#method-clone}{\code{CredentialsLDAP$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsLDAP$new(
  can,
  created_at,
  email,
  is_disabled,
  ldap_dn,
  ldap_id,
  logged_in_at,
  type,
  url
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsLDAP$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsLDAP$fromJSONObject(CredentialsLDAPJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsLDAP$fromJSON(CredentialsLDAPJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsLDAP$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsLDAP$fromJSONString(CredentialsLDAPJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsLDAP$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
