% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CredentialsLookerOpenid.r
\docType{data}
\name{CredentialsLookerOpenid}
\alias{CredentialsLookerOpenid}
\title{CredentialsLookerOpenid Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CredentialsLookerOpenid
}
\description{
CredentialsLookerOpenid Class
}
\section{Fields}{

\describe{
\item{\code{email}}{}

\item{\code{created_at}}{}

\item{\code{logged_in_at}}{}

\item{\code{logged_in_ip}}{}

\item{\code{is_disabled}}{}

\item{\code{type}}{}

\item{\code{url}}{}

\item{\code{user_url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
