% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CredentialsLookerOpenid.r
\name{CredentialsLookerOpenid}
\alias{CredentialsLookerOpenid}
\title{CredentialsLookerOpenid Class}
\description{
CredentialsLookerOpenid Class

CredentialsLookerOpenid Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{email}}{}

\item{\code{created_at}}{}

\item{\code{logged_in_at}}{}

\item{\code{logged_in_ip}}{}

\item{\code{is_disabled}}{}

\item{\code{type}}{}

\item{\code{url}}{}

\item{\code{user_url}}{}

\item{\code{can}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CredentialsLookerOpenid$new()}}
\item \href{#method-toJSON}{\code{CredentialsLookerOpenid$toJSON()}}
\item \href{#method-fromJSON}{\code{CredentialsLookerOpenid$fromJSON()}}
\item \href{#method-toJSONString}{\code{CredentialsLookerOpenid$toJSONString()}}
\item \href{#method-fromJSONString}{\code{CredentialsLookerOpenid$fromJSONString()}}
\item \href{#method-clone}{\code{CredentialsLookerOpenid$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsLookerOpenid$new(
  email,
  created_at,
  logged_in_at,
  logged_in_ip,
  is_disabled,
  type,
  url,
  user_url,
  can
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsLookerOpenid$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsLookerOpenid$fromJSON(CredentialsLookerOpenidJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsLookerOpenid$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsLookerOpenid$fromJSONString(CredentialsLookerOpenidJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsLookerOpenid$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
