% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CredentialsSaml.r
\docType{data}
\name{CredentialsSaml}
\alias{CredentialsSaml}
\title{CredentialsSaml Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CredentialsSaml
}
\description{
CredentialsSaml Class
}
\section{Fields}{

\describe{
\item{\code{email}}{}

\item{\code{saml_user_id}}{}

\item{\code{created_at}}{}

\item{\code{logged_in_at}}{}

\item{\code{is_disabled}}{}

\item{\code{type}}{}

\item{\code{url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
