% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dashboard.r
\name{Dashboard}
\alias{Dashboard}
\title{Dashboard Class}
\description{
Dashboard Class

Dashboard Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{can}}{}

\item{\code{content_favorite_id}}{}

\item{\code{content_metadata_id}}{}

\item{\code{description}}{}

\item{\code{hidden}}{}

\item{\code{id}}{}

\item{\code{model}}{}

\item{\code{query_timezone}}{}

\item{\code{readonly}}{}

\item{\code{refresh_interval}}{}

\item{\code{refresh_interval_to_i}}{}

\item{\code{folder}}{}

\item{\code{title}}{}

\item{\code{user_id}}{}

\item{\code{slug}}{}

\item{\code{preferred_viewer}}{}

\item{\code{alert_sync_with_dashboard_filter_enabled}}{}

\item{\code{background_color}}{}

\item{\code{created_at}}{}

\item{\code{crossfilter_enabled}}{}

\item{\code{dashboard_elements}}{}

\item{\code{dashboard_filters}}{}

\item{\code{dashboard_layouts}}{}

\item{\code{deleted}}{}

\item{\code{deleted_at}}{}

\item{\code{deleter_id}}{}

\item{\code{edit_uri}}{}

\item{\code{favorite_count}}{}

\item{\code{filters_bar_collapsed}}{}

\item{\code{last_accessed_at}}{}

\item{\code{last_viewed_at}}{}

\item{\code{updated_at}}{}

\item{\code{last_updater_id}}{}

\item{\code{last_updater_name}}{}

\item{\code{user_name}}{}

\item{\code{load_configuration}}{}

\item{\code{lookml_link_id}}{}

\item{\code{show_filters_bar}}{}

\item{\code{show_title}}{}

\item{\code{folder_id}}{}

\item{\code{text_tile_text_color}}{}

\item{\code{tile_background_color}}{}

\item{\code{tile_text_color}}{}

\item{\code{title_color}}{}

\item{\code{view_count}}{}

\item{\code{appearance}}{}

\item{\code{url}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Dashboard$new()}}
\item \href{#method-toJSON}{\code{Dashboard$toJSON()}}
\item \href{#method-fromJSONObject}{\code{Dashboard$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{Dashboard$fromJSON()}}
\item \href{#method-toJSONString}{\code{Dashboard$toJSONString()}}
\item \href{#method-fromJSONString}{\code{Dashboard$fromJSONString()}}
\item \href{#method-clone}{\code{Dashboard$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dashboard$new(
  can,
  content_favorite_id,
  content_metadata_id,
  description,
  hidden,
  id,
  model,
  query_timezone,
  readonly,
  refresh_interval,
  refresh_interval_to_i,
  folder,
  title,
  user_id,
  slug,
  preferred_viewer,
  alert_sync_with_dashboard_filter_enabled,
  background_color,
  created_at,
  crossfilter_enabled,
  dashboard_elements,
  dashboard_filters,
  dashboard_layouts,
  deleted,
  deleted_at,
  deleter_id,
  edit_uri,
  favorite_count,
  filters_bar_collapsed,
  last_accessed_at,
  last_viewed_at,
  updated_at,
  last_updater_id,
  last_updater_name,
  user_name,
  load_configuration,
  lookml_link_id,
  show_filters_bar,
  show_title,
  folder_id,
  text_tile_text_color,
  tile_background_color,
  tile_text_color,
  title_color,
  view_count,
  appearance,
  url
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dashboard$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dashboard$fromJSONObject(DashboardJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dashboard$fromJSON(DashboardJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dashboard$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dashboard$fromJSONString(DashboardJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dashboard$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
