% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dashboard.r
\name{Dashboard}
\alias{Dashboard}
\title{Dashboard Class}
\description{
Dashboard Class

Dashboard Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{}

\item{\code{content_metadata_id}}{}

\item{\code{content_favorite_id}}{}

\item{\code{view_count}}{}

\item{\code{last_accessed_at}}{}

\item{\code{favorite_count}}{}

\item{\code{user_id}}{}

\item{\code{title}}{}

\item{\code{description}}{}

\item{\code{readonly}}{}

\item{\code{hidden}}{}

\item{\code{refresh_interval}}{}

\item{\code{refresh_interval_to_i}}{}

\item{\code{space}}{}

\item{\code{load_configuration}}{}

\item{\code{model}}{}

\item{\code{space_id}}{}

\item{\code{dashboard_elements}}{}

\item{\code{dashboard_layouts}}{}

\item{\code{dashboard_filters}}{}

\item{\code{last_viewed_at}}{}

\item{\code{background_color}}{}

\item{\code{show_title}}{}

\item{\code{title_color}}{}

\item{\code{show_filters_bar}}{}

\item{\code{tile_background_color}}{}

\item{\code{tile_text_color}}{}

\item{\code{text_tile_text_color}}{}

\item{\code{last_updater_id}}{}

\item{\code{deleter_id}}{}

\item{\code{deleted}}{}

\item{\code{created_at}}{}

\item{\code{deleted_at}}{}

\item{\code{query_timezone}}{}

\item{\code{edit_uri}}{}

\item{\code{can}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Dashboard$new()}}
\item \href{#method-toJSON}{\code{Dashboard$toJSON()}}
\item \href{#method-fromJSON}{\code{Dashboard$fromJSON()}}
\item \href{#method-toJSONString}{\code{Dashboard$toJSONString()}}
\item \href{#method-fromJSONString}{\code{Dashboard$fromJSONString()}}
\item \href{#method-clone}{\code{Dashboard$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dashboard$new(
  id,
  content_metadata_id,
  content_favorite_id,
  view_count,
  last_accessed_at,
  favorite_count,
  user_id,
  title,
  description,
  readonly,
  hidden,
  refresh_interval,
  refresh_interval_to_i,
  space,
  load_configuration,
  model,
  space_id,
  dashboard_elements,
  dashboard_layouts,
  dashboard_filters,
  last_viewed_at,
  background_color,
  show_title,
  title_color,
  show_filters_bar,
  tile_background_color,
  tile_text_color,
  text_tile_text_color,
  last_updater_id,
  deleter_id,
  deleted,
  created_at,
  deleted_at,
  query_timezone,
  edit_uri,
  can
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dashboard$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dashboard$fromJSON(DashboardJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dashboard$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dashboard$fromJSONString(DashboardJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dashboard$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
