% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DashboardApi.r
\name{DashboardApi}
\alias{DashboardApi}
\title{Dashboard operations}
\description{
looker.Dashboard
}
\section{Methods}{

\describe{

all_dashboards Get All Dashboards


copy_dashboard Copy Dashboard


create_dashboard Create Dashboard


create_dashboard_element Create DashboardElement


create_dashboard_filter Create Dashboard Filter


create_dashboard_layout Create DashboardLayout


dashboard Get Dashboard


dashboard_aggregate_table_lookml Get Aggregate Table LookML for a dashboard


dashboard_dashboard_elements Get All DashboardElements


dashboard_dashboard_filters Get All Dashboard Filters


dashboard_dashboard_layouts Get All DashboardLayouts


dashboard_element Get DashboardElement


dashboard_filter Get Dashboard Filter


dashboard_layout Get DashboardLayout


dashboard_layout_component Get DashboardLayoutComponent


dashboard_layout_dashboard_layout_components Get All DashboardLayoutComponents


dashboard_lookml Get lookml of a UDD


delete_dashboard Delete Dashboard


delete_dashboard_element Delete DashboardElement


delete_dashboard_filter Delete Dashboard Filter


delete_dashboard_layout Delete DashboardLayout


import_lookml_dashboard Import LookML Dashboard


move_dashboard Move Dashboard


search_dashboard_elements Search Dashboard Elements


search_dashboards Search Dashboards


sync_lookml_dashboard Sync LookML Dashboard


update_dashboard Update Dashboard


update_dashboard_element Update DashboardElement


update_dashboard_filter Update Dashboard Filter


update_dashboard_layout Update DashboardLayout


update_dashboard_layout_component Update DashboardLayoutComponent

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DashboardApi$new()}}
\item \href{#method-all_dashboards}{\code{DashboardApi$all_dashboards()}}
\item \href{#method-copy_dashboard}{\code{DashboardApi$copy_dashboard()}}
\item \href{#method-create_dashboard}{\code{DashboardApi$create_dashboard()}}
\item \href{#method-create_dashboard_element}{\code{DashboardApi$create_dashboard_element()}}
\item \href{#method-create_dashboard_filter}{\code{DashboardApi$create_dashboard_filter()}}
\item \href{#method-create_dashboard_layout}{\code{DashboardApi$create_dashboard_layout()}}
\item \href{#method-dashboard}{\code{DashboardApi$dashboard()}}
\item \href{#method-dashboard_aggregate_table_lookml}{\code{DashboardApi$dashboard_aggregate_table_lookml()}}
\item \href{#method-dashboard_dashboard_elements}{\code{DashboardApi$dashboard_dashboard_elements()}}
\item \href{#method-dashboard_dashboard_filters}{\code{DashboardApi$dashboard_dashboard_filters()}}
\item \href{#method-dashboard_dashboard_layouts}{\code{DashboardApi$dashboard_dashboard_layouts()}}
\item \href{#method-dashboard_element}{\code{DashboardApi$dashboard_element()}}
\item \href{#method-dashboard_filter}{\code{DashboardApi$dashboard_filter()}}
\item \href{#method-dashboard_layout}{\code{DashboardApi$dashboard_layout()}}
\item \href{#method-dashboard_layout_component}{\code{DashboardApi$dashboard_layout_component()}}
\item \href{#method-dashboard_layout_dashboard_layout_components}{\code{DashboardApi$dashboard_layout_dashboard_layout_components()}}
\item \href{#method-dashboard_lookml}{\code{DashboardApi$dashboard_lookml()}}
\item \href{#method-delete_dashboard}{\code{DashboardApi$delete_dashboard()}}
\item \href{#method-delete_dashboard_element}{\code{DashboardApi$delete_dashboard_element()}}
\item \href{#method-delete_dashboard_filter}{\code{DashboardApi$delete_dashboard_filter()}}
\item \href{#method-delete_dashboard_layout}{\code{DashboardApi$delete_dashboard_layout()}}
\item \href{#method-import_lookml_dashboard}{\code{DashboardApi$import_lookml_dashboard()}}
\item \href{#method-move_dashboard}{\code{DashboardApi$move_dashboard()}}
\item \href{#method-search_dashboard_elements}{\code{DashboardApi$search_dashboard_elements()}}
\item \href{#method-search_dashboards}{\code{DashboardApi$search_dashboards()}}
\item \href{#method-sync_lookml_dashboard}{\code{DashboardApi$sync_lookml_dashboard()}}
\item \href{#method-update_dashboard}{\code{DashboardApi$update_dashboard()}}
\item \href{#method-update_dashboard_element}{\code{DashboardApi$update_dashboard_element()}}
\item \href{#method-update_dashboard_filter}{\code{DashboardApi$update_dashboard_filter()}}
\item \href{#method-update_dashboard_layout}{\code{DashboardApi$update_dashboard_layout()}}
\item \href{#method-update_dashboard_layout_component}{\code{DashboardApi$update_dashboard_layout_component()}}
\item \href{#method-clone}{\code{DashboardApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_dashboards"></a>}}
\if{latex}{\out{\hypertarget{method-all_dashboards}{}}}
\subsection{Method \code{all_dashboards()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$all_dashboards(fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-copy_dashboard"></a>}}
\if{latex}{\out{\hypertarget{method-copy_dashboard}{}}}
\subsection{Method \code{copy_dashboard()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$copy_dashboard(dashboard_id, folder_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_dashboard"></a>}}
\if{latex}{\out{\hypertarget{method-create_dashboard}{}}}
\subsection{Method \code{create_dashboard()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$create_dashboard(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_dashboard_element"></a>}}
\if{latex}{\out{\hypertarget{method-create_dashboard_element}{}}}
\subsection{Method \code{create_dashboard_element()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$create_dashboard_element(body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_dashboard_filter"></a>}}
\if{latex}{\out{\hypertarget{method-create_dashboard_filter}{}}}
\subsection{Method \code{create_dashboard_filter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$create_dashboard_filter(body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_dashboard_layout"></a>}}
\if{latex}{\out{\hypertarget{method-create_dashboard_layout}{}}}
\subsection{Method \code{create_dashboard_layout()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$create_dashboard_layout(body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dashboard"></a>}}
\if{latex}{\out{\hypertarget{method-dashboard}{}}}
\subsection{Method \code{dashboard()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$dashboard(dashboard_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dashboard_aggregate_table_lookml"></a>}}
\if{latex}{\out{\hypertarget{method-dashboard_aggregate_table_lookml}{}}}
\subsection{Method \code{dashboard_aggregate_table_lookml()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$dashboard_aggregate_table_lookml(dashboard_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dashboard_dashboard_elements"></a>}}
\if{latex}{\out{\hypertarget{method-dashboard_dashboard_elements}{}}}
\subsection{Method \code{dashboard_dashboard_elements()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$dashboard_dashboard_elements(dashboard_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dashboard_dashboard_filters"></a>}}
\if{latex}{\out{\hypertarget{method-dashboard_dashboard_filters}{}}}
\subsection{Method \code{dashboard_dashboard_filters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$dashboard_dashboard_filters(dashboard_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dashboard_dashboard_layouts"></a>}}
\if{latex}{\out{\hypertarget{method-dashboard_dashboard_layouts}{}}}
\subsection{Method \code{dashboard_dashboard_layouts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$dashboard_dashboard_layouts(dashboard_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dashboard_element"></a>}}
\if{latex}{\out{\hypertarget{method-dashboard_element}{}}}
\subsection{Method \code{dashboard_element()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$dashboard_element(dashboard_element_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dashboard_filter"></a>}}
\if{latex}{\out{\hypertarget{method-dashboard_filter}{}}}
\subsection{Method \code{dashboard_filter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$dashboard_filter(dashboard_filter_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dashboard_layout"></a>}}
\if{latex}{\out{\hypertarget{method-dashboard_layout}{}}}
\subsection{Method \code{dashboard_layout()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$dashboard_layout(dashboard_layout_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dashboard_layout_component"></a>}}
\if{latex}{\out{\hypertarget{method-dashboard_layout_component}{}}}
\subsection{Method \code{dashboard_layout_component()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$dashboard_layout_component(
  dashboard_layout_component_id,
  fields,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dashboard_layout_dashboard_layout_components"></a>}}
\if{latex}{\out{\hypertarget{method-dashboard_layout_dashboard_layout_components}{}}}
\subsection{Method \code{dashboard_layout_dashboard_layout_components()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$dashboard_layout_dashboard_layout_components(
  dashboard_layout_id,
  fields,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dashboard_lookml"></a>}}
\if{latex}{\out{\hypertarget{method-dashboard_lookml}{}}}
\subsection{Method \code{dashboard_lookml()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$dashboard_lookml(dashboard_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_dashboard"></a>}}
\if{latex}{\out{\hypertarget{method-delete_dashboard}{}}}
\subsection{Method \code{delete_dashboard()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$delete_dashboard(dashboard_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_dashboard_element"></a>}}
\if{latex}{\out{\hypertarget{method-delete_dashboard_element}{}}}
\subsection{Method \code{delete_dashboard_element()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$delete_dashboard_element(dashboard_element_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_dashboard_filter"></a>}}
\if{latex}{\out{\hypertarget{method-delete_dashboard_filter}{}}}
\subsection{Method \code{delete_dashboard_filter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$delete_dashboard_filter(dashboard_filter_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_dashboard_layout"></a>}}
\if{latex}{\out{\hypertarget{method-delete_dashboard_layout}{}}}
\subsection{Method \code{delete_dashboard_layout()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$delete_dashboard_layout(dashboard_layout_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-import_lookml_dashboard"></a>}}
\if{latex}{\out{\hypertarget{method-import_lookml_dashboard}{}}}
\subsection{Method \code{import_lookml_dashboard()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$import_lookml_dashboard(
  lookml_dashboard_id,
  space_id,
  body,
  raw_locale,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-move_dashboard"></a>}}
\if{latex}{\out{\hypertarget{method-move_dashboard}{}}}
\subsection{Method \code{move_dashboard()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$move_dashboard(dashboard_id, folder_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_dashboard_elements"></a>}}
\if{latex}{\out{\hypertarget{method-search_dashboard_elements}{}}}
\subsection{Method \code{search_dashboard_elements()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$search_dashboard_elements(
  dashboard_id,
  look_id,
  title,
  deleted,
  fields,
  filter_or,
  sorts,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_dashboards"></a>}}
\if{latex}{\out{\hypertarget{method-search_dashboards}{}}}
\subsection{Method \code{search_dashboards()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$search_dashboards(
  id,
  slug,
  title,
  description,
  content_favorite_id,
  folder_id,
  deleted,
  user_id,
  view_count,
  content_metadata_id,
  curate,
  last_viewed_at,
  fields,
  page,
  per_page,
  limit,
  offset,
  sorts,
  filter_or,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sync_lookml_dashboard"></a>}}
\if{latex}{\out{\hypertarget{method-sync_lookml_dashboard}{}}}
\subsection{Method \code{sync_lookml_dashboard()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$sync_lookml_dashboard(lookml_dashboard_id, body, raw_locale, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_dashboard"></a>}}
\if{latex}{\out{\hypertarget{method-update_dashboard}{}}}
\subsection{Method \code{update_dashboard()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$update_dashboard(dashboard_id, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_dashboard_element"></a>}}
\if{latex}{\out{\hypertarget{method-update_dashboard_element}{}}}
\subsection{Method \code{update_dashboard_element()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$update_dashboard_element(dashboard_element_id, body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_dashboard_filter"></a>}}
\if{latex}{\out{\hypertarget{method-update_dashboard_filter}{}}}
\subsection{Method \code{update_dashboard_filter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$update_dashboard_filter(dashboard_filter_id, body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_dashboard_layout"></a>}}
\if{latex}{\out{\hypertarget{method-update_dashboard_layout}{}}}
\subsection{Method \code{update_dashboard_layout()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$update_dashboard_layout(dashboard_layout_id, body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_dashboard_layout_component"></a>}}
\if{latex}{\out{\hypertarget{method-update_dashboard_layout_component}{}}}
\subsection{Method \code{update_dashboard_layout_component()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$update_dashboard_layout_component(
  dashboard_layout_component_id,
  body,
  fields,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
