% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DashboardApi.r
\name{DashboardApi}
\alias{DashboardApi}
\title{Dashboard operations}
\description{
looker.Dashboard
}
\section{Methods}{

\describe{

all_dashboards Get All Dashboards 


create_dashboard Create Dashboard 


create_dashboard_prefetch Create Dashboard Prefetch 


dashboard Get Dashboard 


dashboard_prefetch Get Dashboard Prefetch 


delete_dashboard Delete Dashboard 


search_dashboards Search Dashboards 


update_dashboard Update Dashboard 

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DashboardApi$new()}}
\item \href{#method-all_dashboards}{\code{DashboardApi$all_dashboards()}}
\item \href{#method-create_dashboard}{\code{DashboardApi$create_dashboard()}}
\item \href{#method-create_dashboard_prefetch}{\code{DashboardApi$create_dashboard_prefetch()}}
\item \href{#method-dashboard}{\code{DashboardApi$dashboard()}}
\item \href{#method-dashboard_prefetch}{\code{DashboardApi$dashboard_prefetch()}}
\item \href{#method-delete_dashboard}{\code{DashboardApi$delete_dashboard()}}
\item \href{#method-search_dashboards}{\code{DashboardApi$search_dashboards()}}
\item \href{#method-update_dashboard}{\code{DashboardApi$update_dashboard()}}
\item \href{#method-clone}{\code{DashboardApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_dashboards"></a>}}
\if{latex}{\out{\hypertarget{method-all_dashboards}{}}}
\subsection{Method \code{all_dashboards()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$all_dashboards(fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_dashboard"></a>}}
\if{latex}{\out{\hypertarget{method-create_dashboard}{}}}
\subsection{Method \code{create_dashboard()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$create_dashboard(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_dashboard_prefetch"></a>}}
\if{latex}{\out{\hypertarget{method-create_dashboard_prefetch}{}}}
\subsection{Method \code{create_dashboard_prefetch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$create_dashboard_prefetch(dashboard_id, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dashboard"></a>}}
\if{latex}{\out{\hypertarget{method-dashboard}{}}}
\subsection{Method \code{dashboard()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$dashboard(dashboard_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dashboard_prefetch"></a>}}
\if{latex}{\out{\hypertarget{method-dashboard_prefetch}{}}}
\subsection{Method \code{dashboard_prefetch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$dashboard_prefetch(dashboard_id, dashboard_filters, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_dashboard"></a>}}
\if{latex}{\out{\hypertarget{method-delete_dashboard}{}}}
\subsection{Method \code{delete_dashboard()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$delete_dashboard(dashboard_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_dashboards"></a>}}
\if{latex}{\out{\hypertarget{method-search_dashboards}{}}}
\subsection{Method \code{search_dashboards()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$search_dashboards(
  fields,
  id,
  page,
  per_page,
  limit,
  offset,
  sorts,
  title,
  description,
  content_favorite_id,
  space_id,
  deleted,
  user_id,
  view_count,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_dashboard"></a>}}
\if{latex}{\out{\hypertarget{method-update_dashboard}{}}}
\subsection{Method \code{update_dashboard()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$update_dashboard(dashboard_id, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
