% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DashboardBase.r
\name{DashboardBase}
\alias{DashboardBase}
\title{DashboardBase Class}
\description{
DashboardBase Class

DashboardBase Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{can}}{}

\item{\code{content_favorite_id}}{}

\item{\code{content_metadata_id}}{}

\item{\code{description}}{}

\item{\code{hidden}}{}

\item{\code{id}}{}

\item{\code{model}}{}

\item{\code{query_timezone}}{}

\item{\code{readonly}}{}

\item{\code{refresh_interval}}{}

\item{\code{refresh_interval_to_i}}{}

\item{\code{folder}}{}

\item{\code{title}}{}

\item{\code{user_id}}{}

\item{\code{slug}}{}

\item{\code{preferred_viewer}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DashboardBase$new()}}
\item \href{#method-toJSON}{\code{DashboardBase$toJSON()}}
\item \href{#method-fromJSONObject}{\code{DashboardBase$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{DashboardBase$fromJSON()}}
\item \href{#method-toJSONString}{\code{DashboardBase$toJSONString()}}
\item \href{#method-fromJSONString}{\code{DashboardBase$fromJSONString()}}
\item \href{#method-clone}{\code{DashboardBase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardBase$new(
  can,
  content_favorite_id,
  content_metadata_id,
  description,
  hidden,
  id,
  model,
  query_timezone,
  readonly,
  refresh_interval,
  refresh_interval_to_i,
  folder,
  title,
  user_id,
  slug,
  preferred_viewer
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardBase$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardBase$fromJSONObject(DashboardBaseJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardBase$fromJSON(DashboardBaseJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardBase$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardBase$fromJSONString(DashboardBaseJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
