% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DashboardElement.r
\name{DashboardElement}
\alias{DashboardElement}
\title{DashboardElement Class}
\description{
DashboardElement Class

DashboardElement Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{can}}{}

\item{\code{body_text}}{}

\item{\code{body_text_as_html}}{}

\item{\code{dashboard_id}}{}

\item{\code{edit_uri}}{}

\item{\code{id}}{}

\item{\code{look}}{}

\item{\code{look_id}}{}

\item{\code{lookml_link_id}}{}

\item{\code{merge_result_id}}{}

\item{\code{note_display}}{}

\item{\code{note_state}}{}

\item{\code{note_text}}{}

\item{\code{note_text_as_html}}{}

\item{\code{query}}{}

\item{\code{query_id}}{}

\item{\code{refresh_interval}}{}

\item{\code{refresh_interval_to_i}}{}

\item{\code{result_maker}}{}

\item{\code{result_maker_id}}{}

\item{\code{subtitle_text}}{}

\item{\code{title}}{}

\item{\code{title_hidden}}{}

\item{\code{title_text}}{}

\item{\code{type}}{}

\item{\code{alert_count}}{}

\item{\code{title_text_as_html}}{}

\item{\code{subtitle_text_as_html}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DashboardElement$new()}}
\item \href{#method-toJSON}{\code{DashboardElement$toJSON()}}
\item \href{#method-fromJSONObject}{\code{DashboardElement$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{DashboardElement$fromJSON()}}
\item \href{#method-toJSONString}{\code{DashboardElement$toJSONString()}}
\item \href{#method-fromJSONString}{\code{DashboardElement$fromJSONString()}}
\item \href{#method-clone}{\code{DashboardElement$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardElement$new(
  can,
  body_text,
  body_text_as_html,
  dashboard_id,
  edit_uri,
  id,
  look,
  look_id,
  lookml_link_id,
  merge_result_id,
  note_display,
  note_state,
  note_text,
  note_text_as_html,
  query,
  query_id,
  refresh_interval,
  refresh_interval_to_i,
  result_maker,
  result_maker_id,
  subtitle_text,
  title,
  title_hidden,
  title_text,
  type,
  alert_count,
  title_text_as_html,
  subtitle_text_as_html
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardElement$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardElement$fromJSONObject(DashboardElementJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardElement$fromJSON(DashboardElementJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardElement$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardElement$fromJSONString(DashboardElementJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardElement$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
