% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DashboardFilter.r
\docType{data}
\name{DashboardFilter}
\alias{DashboardFilter}
\title{DashboardFilter Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DashboardFilter
}
\description{
DashboardFilter Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{dashboard_id}}{}

\item{\code{name}}{}

\item{\code{title}}{}

\item{\code{type}}{}

\item{\code{default_value}}{}

\item{\code{model}}{}

\item{\code{explore}}{}

\item{\code{dimension}}{}

\item{\code{field}}{}

\item{\code{row}}{}

\item{\code{listens_to_filters}}{}

\item{\code{allow_multiple_values}}{}

\item{\code{required}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
