% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DashboardLayout.r
\name{DashboardLayout}
\alias{DashboardLayout}
\title{DashboardLayout Class}
\description{
DashboardLayout Class

DashboardLayout Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{}

\item{\code{dashboard_id}}{}

\item{\code{type}}{}

\item{\code{active}}{}

\item{\code{column_width}}{}

\item{\code{width}}{}

\item{\code{deleted}}{}

\item{\code{dashboard_layout_components}}{}

\item{\code{can}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DashboardLayout$new()}}
\item \href{#method-toJSON}{\code{DashboardLayout$toJSON()}}
\item \href{#method-fromJSON}{\code{DashboardLayout$fromJSON()}}
\item \href{#method-toJSONString}{\code{DashboardLayout$toJSONString()}}
\item \href{#method-fromJSONString}{\code{DashboardLayout$fromJSONString()}}
\item \href{#method-clone}{\code{DashboardLayout$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardLayout$new(
  id,
  dashboard_id,
  type,
  active,
  column_width,
  width,
  deleted,
  dashboard_layout_components,
  can
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardLayout$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardLayout$fromJSON(DashboardLayoutJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardLayout$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardLayout$fromJSONString(DashboardLayoutJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardLayout$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
