% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DashboardLayoutComponent.r
\name{DashboardLayoutComponent}
\alias{DashboardLayoutComponent}
\title{DashboardLayoutComponent Class}
\description{
DashboardLayoutComponent Class

DashboardLayoutComponent Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{can}}{}

\item{\code{id}}{}

\item{\code{dashboard_layout_id}}{}

\item{\code{dashboard_element_id}}{}

\item{\code{row}}{}

\item{\code{column}}{}

\item{\code{width}}{}

\item{\code{height}}{}

\item{\code{deleted}}{}

\item{\code{element_title}}{}

\item{\code{element_title_hidden}}{}

\item{\code{vis_type}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DashboardLayoutComponent$new()}}
\item \href{#method-toJSON}{\code{DashboardLayoutComponent$toJSON()}}
\item \href{#method-fromJSONObject}{\code{DashboardLayoutComponent$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{DashboardLayoutComponent$fromJSON()}}
\item \href{#method-toJSONString}{\code{DashboardLayoutComponent$toJSONString()}}
\item \href{#method-fromJSONString}{\code{DashboardLayoutComponent$fromJSONString()}}
\item \href{#method-clone}{\code{DashboardLayoutComponent$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardLayoutComponent$new(
  can,
  id,
  dashboard_layout_id,
  dashboard_element_id,
  row,
  column,
  width,
  height,
  deleted,
  element_title,
  element_title_hidden,
  vis_type
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardLayoutComponent$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardLayoutComponent$fromJSONObject(DashboardLayoutComponentJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardLayoutComponent$fromJSON(DashboardLayoutComponentJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardLayoutComponent$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardLayoutComponent$fromJSONString(DashboardLayoutComponentJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DashboardLayoutComponent$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
