% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataActionApi.r
\docType{data}
\name{DataActionApi}
\alias{DataActionApi}
\title{DataAction operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataActionApi
}
\description{
looker.DataAction
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

fetch_remote_data_action_form Fetch Remote Data Action Form 


perform_data_action Send a Data Action 

}
}

\keyword{datasets}
