% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataActionResponse.r
\docType{data}
\name{DataActionResponse}
\alias{DataActionResponse}
\title{DataActionResponse Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataActionResponse
}
\description{
DataActionResponse Class
}
\section{Fields}{

\describe{
\item{\code{webhook_id}}{}

\item{\code{success}}{}

\item{\code{refresh_query}}{}

\item{\code{validation_errors}}{}

\item{\code{message}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
