% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatagroupApi.r
\docType{data}
\name{DatagroupApi}
\alias{DatagroupApi}
\title{Datagroup operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DatagroupApi
}
\description{
looker.Datagroup
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

all_datagroups Get All Datagroups 


datagroup Get Datagroup 


update_datagroup Update Datagroup 

}
}

\keyword{datasets}
