% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBConnection.r
\docType{data}
\name{DBConnection}
\alias{DBConnection}
\title{DBConnection Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DBConnection
}
\description{
DBConnection Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{dialect}}{}

\item{\code{snippets}}{}

\item{\code{host}}{}

\item{\code{port}}{}

\item{\code{username}}{}

\item{\code{password}}{}

\item{\code{certificate}}{}

\item{\code{file_type}}{}

\item{\code{database}}{}

\item{\code{db_timezone}}{}

\item{\code{query_timezone}}{}

\item{\code{schema}}{}

\item{\code{max_connections}}{}

\item{\code{max_billing_gigabytes}}{}

\item{\code{ssl}}{}

\item{\code{verify_ssl}}{}

\item{\code{tmp_db_name}}{}

\item{\code{jdbc_additional_params}}{}

\item{\code{pool_timeout}}{}

\item{\code{dialect_name}}{}

\item{\code{created_at}}{}

\item{\code{user_id}}{}

\item{\code{example}}{}

\item{\code{user_db_credentials}}{}

\item{\code{user_attribute_fields}}{}

\item{\code{maintenance_cron}}{}

\item{\code{last_regen_at}}{}

\item{\code{last_reap_at}}{}

\item{\code{sql_runner_precache_tables}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
