% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dialect.r
\name{Dialect}
\alias{Dialect}
\title{Dialect Class}
\description{
Dialect Class

Dialect Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{}

\item{\code{label}}{}

\item{\code{supports_cost_estimate}}{}

\item{\code{persistent_table_indexes}}{}

\item{\code{persistent_table_sortkeys}}{}

\item{\code{persistent_table_distkey}}{}

\item{\code{supports_streaming}}{}

\item{\code{automatically_run_sql_runner_snippets}}{}

\item{\code{connection_tests}}{}

\item{\code{supports_inducer}}{}

\item{\code{supports_multiple_databases}}{}

\item{\code{supports_persistent_derived_tables}}{}

\item{\code{has_ssl_support}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Dialect$new()}}
\item \href{#method-toJSON}{\code{Dialect$toJSON()}}
\item \href{#method-fromJSONObject}{\code{Dialect$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{Dialect$fromJSON()}}
\item \href{#method-toJSONString}{\code{Dialect$toJSONString()}}
\item \href{#method-fromJSONString}{\code{Dialect$fromJSONString()}}
\item \href{#method-clone}{\code{Dialect$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dialect$new(
  name,
  label,
  supports_cost_estimate,
  persistent_table_indexes,
  persistent_table_sortkeys,
  persistent_table_distkey,
  supports_streaming,
  automatically_run_sql_runner_snippets,
  connection_tests,
  supports_inducer,
  supports_multiple_databases,
  supports_persistent_derived_tables,
  has_ssl_support
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dialect$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dialect$fromJSONObject(DialectJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dialect$fromJSON(DialectJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dialect$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dialect$fromJSONString(DialectJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dialect$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
