% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DialectInfo.r
\docType{data}
\name{DialectInfo}
\alias{DialectInfo}
\title{DialectInfo Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DialectInfo
}
\description{
DialectInfo Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{label}}{}

\item{\code{label_for_database_equivalent}}{}

\item{\code{default_port}}{}

\item{\code{default_max_connections}}{}

\item{\code{supported_options}}{}

\item{\code{installed}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
