% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DialectInfoOptions.r
\name{DialectInfoOptions}
\alias{DialectInfoOptions}
\title{DialectInfoOptions Class}
\description{
DialectInfoOptions Class

DialectInfoOptions Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{timezone}}{}

\item{\code{schema}}{}

\item{\code{ssl}}{}

\item{\code{auth}}{}

\item{\code{host}}{}

\item{\code{tmp_table}}{}

\item{\code{project_name}}{}

\item{\code{oauth_credentials}}{}

\item{\code{additional_params}}{}

\item{\code{username_required}}{}

\item{\code{can}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DialectInfoOptions$new()}}
\item \href{#method-toJSON}{\code{DialectInfoOptions$toJSON()}}
\item \href{#method-fromJSON}{\code{DialectInfoOptions$fromJSON()}}
\item \href{#method-toJSONString}{\code{DialectInfoOptions$toJSONString()}}
\item \href{#method-fromJSONString}{\code{DialectInfoOptions$fromJSONString()}}
\item \href{#method-clone}{\code{DialectInfoOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DialectInfoOptions$new(
  timezone,
  schema,
  ssl,
  auth,
  host,
  tmp_table,
  project_name,
  oauth_credentials,
  additional_params,
  username_required,
  can
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DialectInfoOptions$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DialectInfoOptions$fromJSON(DialectInfoOptionsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DialectInfoOptions$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DialectInfoOptions$fromJSONString(DialectInfoOptionsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DialectInfoOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
