% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DialectInfoOptions.r
\docType{data}
\name{DialectInfoOptions}
\alias{DialectInfoOptions}
\title{DialectInfoOptions Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DialectInfoOptions
}
\description{
DialectInfoOptions Class
}
\section{Fields}{

\describe{
\item{\code{timezone}}{}

\item{\code{schema}}{}

\item{\code{ssl}}{}

\item{\code{auth}}{}

\item{\code{host}}{}

\item{\code{tmp_table}}{}

\item{\code{project_name}}{}

\item{\code{oauth_credentials}}{}

\item{\code{additional_params}}{}

\item{\code{username_required}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
