% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetDwellTimes}
\alias{GetDwellTimes}
\title{Compute dwell times}
\usage{
GetDwellTimes(...)
}
\value{
the Trial object(s) with a \code{DwellSummary} attribute which
  contains a dataframe describing the maximum and total dwell times for each
  image location and stimulus.
}
\description{
A "dwell" is an uninterrupted gaze. This function calculates the longest
dwell time per image and the total time dwelling in each trial.
}

