% Generated by roxygen2 (4.0.0): do not edit by hand
\name{GetImageAOI}
\alias{GetImageAOI}
\title{Get the boundaries of an Area of Interest from its name}
\usage{
GetImageAOI(image_location)
}
\arguments{
\item{image_location}{a string naming an image location. It may be:
\code{UpperLeftImage}, \code{UpperRightImage}, \code{LowerRightImage},
\code{LowerLeftImage}, \code{FixationImage}, \code{ImageL}, or
\code{ImageR}.}
}
\value{
a list specifying the boundaries of the named AOI
}
\description{
Get the boundaries of an Area of Interest from its name
}
\keyword{internal}

