% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitBranch.r
\name{GitBranch}
\alias{GitBranch}
\title{GitBranch Class}
\description{
GitBranch Class

GitBranch Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{can}}{}

\item{\code{name}}{}

\item{\code{remote}}{}

\item{\code{remote_name}}{}

\item{\code{error}}{}

\item{\code{message}}{}

\item{\code{owner_name}}{}

\item{\code{readonly}}{}

\item{\code{personal}}{}

\item{\code{is_local}}{}

\item{\code{is_remote}}{}

\item{\code{is_production}}{}

\item{\code{ahead_count}}{}

\item{\code{behind_count}}{}

\item{\code{commit_at}}{}

\item{\code{ref}}{}

\item{\code{remote_ref}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GitBranch$new()}}
\item \href{#method-toJSON}{\code{GitBranch$toJSON()}}
\item \href{#method-fromJSONObject}{\code{GitBranch$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{GitBranch$fromJSON()}}
\item \href{#method-toJSONString}{\code{GitBranch$toJSONString()}}
\item \href{#method-fromJSONString}{\code{GitBranch$fromJSONString()}}
\item \href{#method-clone}{\code{GitBranch$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitBranch$new(
  can,
  name,
  remote,
  remote_name,
  error,
  message,
  owner_name,
  readonly,
  personal,
  is_local,
  is_remote,
  is_production,
  ahead_count,
  behind_count,
  commit_at,
  ref,
  remote_ref
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitBranch$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitBranch$fromJSONObject(GitBranchJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitBranch$fromJSON(GitBranchJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitBranch$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitBranch$fromJSONString(GitBranchJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GitBranch$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
