% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitConnectionTest.r
\docType{data}
\name{GitConnectionTest}
\alias{GitConnectionTest}
\title{GitConnectionTest Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GitConnectionTest
}
\description{
GitConnectionTest Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{description}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
