% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupApi.r
\name{GroupApi}
\alias{GroupApi}
\title{Group operations}
\description{
looker.Group
}
\section{Methods}{

\describe{

add_group_group Add a Group to Group 


add_group_user Add a User to Group 


all_group_groups Get All Groups in Group 


all_group_users Get All Users in Group 


all_groups Get All Groups 


create_group Create Group 


delete_group Delete Group 


delete_group_from_group Deletes a Group from Group 


delete_group_user Remove a User from Group 


delete_user_attribute_group_value Delete User Attribute Group Value 


group Get Group 


update_group Update Group 


update_user_attribute_group_value Set User Attribute Group Value 

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GroupApi$new()}}
\item \href{#method-add_group_group}{\code{GroupApi$add_group_group()}}
\item \href{#method-add_group_user}{\code{GroupApi$add_group_user()}}
\item \href{#method-all_group_groups}{\code{GroupApi$all_group_groups()}}
\item \href{#method-all_group_users}{\code{GroupApi$all_group_users()}}
\item \href{#method-all_groups}{\code{GroupApi$all_groups()}}
\item \href{#method-create_group}{\code{GroupApi$create_group()}}
\item \href{#method-delete_group}{\code{GroupApi$delete_group()}}
\item \href{#method-delete_group_from_group}{\code{GroupApi$delete_group_from_group()}}
\item \href{#method-delete_group_user}{\code{GroupApi$delete_group_user()}}
\item \href{#method-delete_user_attribute_group_value}{\code{GroupApi$delete_user_attribute_group_value()}}
\item \href{#method-group}{\code{GroupApi$group()}}
\item \href{#method-update_group}{\code{GroupApi$update_group()}}
\item \href{#method-update_user_attribute_group_value}{\code{GroupApi$update_user_attribute_group_value()}}
\item \href{#method-clone}{\code{GroupApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_group_group"></a>}}
\if{latex}{\out{\hypertarget{method-add_group_group}{}}}
\subsection{Method \code{add_group_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupApi$add_group_group(group_id, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_group_user"></a>}}
\if{latex}{\out{\hypertarget{method-add_group_user}{}}}
\subsection{Method \code{add_group_user()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupApi$add_group_user(group_id, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_group_groups"></a>}}
\if{latex}{\out{\hypertarget{method-all_group_groups}{}}}
\subsection{Method \code{all_group_groups()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupApi$all_group_groups(group_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_group_users"></a>}}
\if{latex}{\out{\hypertarget{method-all_group_users}{}}}
\subsection{Method \code{all_group_users()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupApi$all_group_users(group_id, fields, page, per_page, sorts, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_groups"></a>}}
\if{latex}{\out{\hypertarget{method-all_groups}{}}}
\subsection{Method \code{all_groups()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupApi$all_groups(
  fields,
  page,
  per_page,
  sorts,
  ids,
  content_metadata_id,
  can_add_to_content_metadata,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_group"></a>}}
\if{latex}{\out{\hypertarget{method-create_group}{}}}
\subsection{Method \code{create_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupApi$create_group(body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_group"></a>}}
\if{latex}{\out{\hypertarget{method-delete_group}{}}}
\subsection{Method \code{delete_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupApi$delete_group(group_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_group_from_group"></a>}}
\if{latex}{\out{\hypertarget{method-delete_group_from_group}{}}}
\subsection{Method \code{delete_group_from_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupApi$delete_group_from_group(group_id, deleting_group_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_group_user"></a>}}
\if{latex}{\out{\hypertarget{method-delete_group_user}{}}}
\subsection{Method \code{delete_group_user()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupApi$delete_group_user(group_id, user_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_user_attribute_group_value"></a>}}
\if{latex}{\out{\hypertarget{method-delete_user_attribute_group_value}{}}}
\subsection{Method \code{delete_user_attribute_group_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupApi$delete_user_attribute_group_value(group_id, user_attribute_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-group"></a>}}
\if{latex}{\out{\hypertarget{method-group}{}}}
\subsection{Method \code{group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupApi$group(group_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_group"></a>}}
\if{latex}{\out{\hypertarget{method-update_group}{}}}
\subsection{Method \code{update_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupApi$update_group(group_id, body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_user_attribute_group_value"></a>}}
\if{latex}{\out{\hypertarget{method-update_user_attribute_group_value}{}}}
\subsection{Method \code{update_user_attribute_group_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupApi$update_user_attribute_group_value(
  group_id,
  user_attribute_id,
  body,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
