% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupIdForGroupUserInclusion.r
\name{GroupIdForGroupUserInclusion}
\alias{GroupIdForGroupUserInclusion}
\title{GroupIdForGroupUserInclusion Class}
\description{
GroupIdForGroupUserInclusion Class

GroupIdForGroupUserInclusion Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{user_id}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GroupIdForGroupUserInclusion$new()}}
\item \href{#method-toJSON}{\code{GroupIdForGroupUserInclusion$toJSON()}}
\item \href{#method-fromJSONObject}{\code{GroupIdForGroupUserInclusion$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{GroupIdForGroupUserInclusion$fromJSON()}}
\item \href{#method-toJSONString}{\code{GroupIdForGroupUserInclusion$toJSONString()}}
\item \href{#method-fromJSONString}{\code{GroupIdForGroupUserInclusion$fromJSONString()}}
\item \href{#method-clone}{\code{GroupIdForGroupUserInclusion$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupIdForGroupUserInclusion$new(user_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupIdForGroupUserInclusion$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupIdForGroupUserInclusion$fromJSONObject(
  GroupIdForGroupUserInclusionJsonObject
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupIdForGroupUserInclusion$fromJSON(GroupIdForGroupUserInclusionJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupIdForGroupUserInclusion$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupIdForGroupUserInclusion$fromJSONString(GroupIdForGroupUserInclusionJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupIdForGroupUserInclusion$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
