% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomepageApi.r
\name{HomepageApi}
\alias{HomepageApi}
\title{Homepage operations}
\description{
looker.Homepage
}
\section{Methods}{

\describe{

all_homepage_items Get All Homepage Items 


all_homepage_sections Get All Homepage sections 


create_homepage_item Create Homepage Item 


create_homepage_section Create Homepage section 


delete_homepage_item Delete Homepage Item 


delete_homepage_section Delete Homepage section 


homepage_item Get Homepage Item 


homepage_section Get Homepage section 


update_homepage_item Update Homepage Item 


update_homepage_section Update Homepage section 

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{HomepageApi$new()}}
\item \href{#method-all_homepage_items}{\code{HomepageApi$all_homepage_items()}}
\item \href{#method-all_homepage_sections}{\code{HomepageApi$all_homepage_sections()}}
\item \href{#method-create_homepage_item}{\code{HomepageApi$create_homepage_item()}}
\item \href{#method-create_homepage_section}{\code{HomepageApi$create_homepage_section()}}
\item \href{#method-delete_homepage_item}{\code{HomepageApi$delete_homepage_item()}}
\item \href{#method-delete_homepage_section}{\code{HomepageApi$delete_homepage_section()}}
\item \href{#method-homepage_item}{\code{HomepageApi$homepage_item()}}
\item \href{#method-homepage_section}{\code{HomepageApi$homepage_section()}}
\item \href{#method-update_homepage_item}{\code{HomepageApi$update_homepage_item()}}
\item \href{#method-update_homepage_section}{\code{HomepageApi$update_homepage_section()}}
\item \href{#method-clone}{\code{HomepageApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HomepageApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_homepage_items"></a>}}
\if{latex}{\out{\hypertarget{method-all_homepage_items}{}}}
\subsection{Method \code{all_homepage_items()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HomepageApi$all_homepage_items(fields, sorts, homepage_section_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_homepage_sections"></a>}}
\if{latex}{\out{\hypertarget{method-all_homepage_sections}{}}}
\subsection{Method \code{all_homepage_sections()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HomepageApi$all_homepage_sections(fields, sorts, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_homepage_item"></a>}}
\if{latex}{\out{\hypertarget{method-create_homepage_item}{}}}
\subsection{Method \code{create_homepage_item()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HomepageApi$create_homepage_item(body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_homepage_section"></a>}}
\if{latex}{\out{\hypertarget{method-create_homepage_section}{}}}
\subsection{Method \code{create_homepage_section()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HomepageApi$create_homepage_section(body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_homepage_item"></a>}}
\if{latex}{\out{\hypertarget{method-delete_homepage_item}{}}}
\subsection{Method \code{delete_homepage_item()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HomepageApi$delete_homepage_item(homepage_item_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_homepage_section"></a>}}
\if{latex}{\out{\hypertarget{method-delete_homepage_section}{}}}
\subsection{Method \code{delete_homepage_section()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HomepageApi$delete_homepage_section(homepage_section_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-homepage_item"></a>}}
\if{latex}{\out{\hypertarget{method-homepage_item}{}}}
\subsection{Method \code{homepage_item()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HomepageApi$homepage_item(homepage_item_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-homepage_section"></a>}}
\if{latex}{\out{\hypertarget{method-homepage_section}{}}}
\subsection{Method \code{homepage_section()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HomepageApi$homepage_section(homepage_section_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_homepage_item"></a>}}
\if{latex}{\out{\hypertarget{method-update_homepage_item}{}}}
\subsection{Method \code{update_homepage_item()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HomepageApi$update_homepage_item(homepage_item_id, body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_homepage_section"></a>}}
\if{latex}{\out{\hypertarget{method-update_homepage_section}{}}}
\subsection{Method \code{update_homepage_section()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HomepageApi$update_homepage_section(homepage_section_id, body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HomepageApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
