% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomepageApi.r
\name{HomepageApi}
\alias{HomepageApi}
\title{Homepage operations}
\description{
looker.Homepage
}
\section{Methods}{

\describe{

all_primary_homepage_sections Get All Primary homepage sections

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{HomepageApi$new()}}
\item \href{#method-all_primary_homepage_sections}{\code{HomepageApi$all_primary_homepage_sections()}}
\item \href{#method-clone}{\code{HomepageApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HomepageApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_primary_homepage_sections"></a>}}
\if{latex}{\out{\hypertarget{method-all_primary_homepage_sections}{}}}
\subsection{Method \code{all_primary_homepage_sections()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HomepageApi$all_primary_homepage_sections(fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HomepageApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
