% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomepageItem.r
\name{HomepageItem}
\alias{HomepageItem}
\title{HomepageItem Class}
\description{
HomepageItem Class

HomepageItem Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{can}}{}

\item{\code{content_created_by}}{}

\item{\code{content_favorite_id}}{}

\item{\code{content_metadata_id}}{}

\item{\code{content_updated_at}}{}

\item{\code{custom_description}}{}

\item{\code{custom_image_data_base64}}{}

\item{\code{custom_image_url}}{}

\item{\code{custom_title}}{}

\item{\code{custom_url}}{}

\item{\code{dashboard_id}}{}

\item{\code{description}}{}

\item{\code{favorite_count}}{}

\item{\code{homepage_section_id}}{}

\item{\code{id}}{}

\item{\code{image_url}}{}

\item{\code{location}}{}

\item{\code{look_id}}{}

\item{\code{lookml_dashboard_id}}{}

\item{\code{order}}{}

\item{\code{section_fetch_time}}{}

\item{\code{title}}{}

\item{\code{url}}{}

\item{\code{use_custom_description}}{}

\item{\code{use_custom_image}}{}

\item{\code{use_custom_title}}{}

\item{\code{use_custom_url}}{}

\item{\code{view_count}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{HomepageItem$new()}}
\item \href{#method-toJSON}{\code{HomepageItem$toJSON()}}
\item \href{#method-fromJSONObject}{\code{HomepageItem$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{HomepageItem$fromJSON()}}
\item \href{#method-toJSONString}{\code{HomepageItem$toJSONString()}}
\item \href{#method-fromJSONString}{\code{HomepageItem$fromJSONString()}}
\item \href{#method-clone}{\code{HomepageItem$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HomepageItem$new(
  can,
  content_created_by,
  content_favorite_id,
  content_metadata_id,
  content_updated_at,
  custom_description,
  custom_image_data_base64,
  custom_image_url,
  custom_title,
  custom_url,
  dashboard_id,
  description,
  favorite_count,
  homepage_section_id,
  id,
  image_url,
  location,
  look_id,
  lookml_dashboard_id,
  order,
  section_fetch_time,
  title,
  url,
  use_custom_description,
  use_custom_image,
  use_custom_title,
  use_custom_url,
  view_count
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HomepageItem$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HomepageItem$fromJSONObject(HomepageItemJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HomepageItem$fromJSON(HomepageItemJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HomepageItem$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HomepageItem$fromJSONString(HomepageItemJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HomepageItem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
