% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomepageSection.r
\docType{data}
\name{HomepageSection}
\alias{HomepageSection}
\title{HomepageSection Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
HomepageSection
}
\description{
HomepageSection Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{title}}{}

\item{\code{is_dynamic}}{}

\item{\code{is_header}}{}

\item{\code{order}}{}

\item{\code{detail_url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
