% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Integration.r
\name{Integration}
\alias{Integration}
\title{Integration Class}
\description{
Integration Class

Integration Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{can}}{}

\item{\code{id}}{}

\item{\code{integration_hub_id}}{}

\item{\code{label}}{}

\item{\code{description}}{}

\item{\code{enabled}}{}

\item{\code{params}}{}

\item{\code{supported_formats}}{}

\item{\code{supported_action_types}}{}

\item{\code{supported_formattings}}{}

\item{\code{supported_visualization_formattings}}{}

\item{\code{supported_download_settings}}{}

\item{\code{icon_url}}{}

\item{\code{uses_oauth}}{}

\item{\code{required_fields}}{}

\item{\code{delegate_oauth}}{}

\item{\code{installed_delegate_oauth_targets}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Integration$new()}}
\item \href{#method-toJSON}{\code{Integration$toJSON()}}
\item \href{#method-fromJSONObject}{\code{Integration$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{Integration$fromJSON()}}
\item \href{#method-toJSONString}{\code{Integration$toJSONString()}}
\item \href{#method-fromJSONString}{\code{Integration$fromJSONString()}}
\item \href{#method-clone}{\code{Integration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Integration$new(
  can,
  id,
  integration_hub_id,
  label,
  description,
  enabled,
  params,
  supported_formats,
  supported_action_types,
  supported_formattings,
  supported_visualization_formattings,
  supported_download_settings,
  icon_url,
  uses_oauth,
  required_fields,
  delegate_oauth,
  installed_delegate_oauth_targets
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Integration$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Integration$fromJSONObject(IntegrationJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Integration$fromJSON(IntegrationJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Integration$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Integration$fromJSONString(IntegrationJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Integration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
