% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrationApi.r
\docType{data}
\name{IntegrationApi}
\alias{IntegrationApi}
\title{Integration operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
IntegrationApi
}
\description{
looker.Integration
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

accept_integration_hub_legal_agreement Accept Integration Hub Legal Agreement 


all_integration_hubs Get All Integration Hubs 


all_integrations Get All Integrations 


create_integration_hub Create Integration Hub 


delete_integration_hub Delete Integration Hub 


fetch_integration_form Fetch Remote Integration Form 


integration Get Integration 


integration_hub Get Integration Hub 


update_integration Update Integration 


update_integration_hub Update Integration Hub 

}
}

\keyword{datasets}
