% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrationHub.r
\name{IntegrationHub}
\alias{IntegrationHub}
\title{IntegrationHub Class}
\description{
IntegrationHub Class

IntegrationHub Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{}

\item{\code{url}}{}

\item{\code{label}}{}

\item{\code{official}}{}

\item{\code{fetch_error_message}}{}

\item{\code{authorization_token}}{}

\item{\code{has_authorization_token}}{}

\item{\code{legal_agreement_signed}}{}

\item{\code{legal_agreement_required}}{}

\item{\code{legal_agreement_text}}{}

\item{\code{can}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{IntegrationHub$new()}}
\item \href{#method-toJSON}{\code{IntegrationHub$toJSON()}}
\item \href{#method-fromJSON}{\code{IntegrationHub$fromJSON()}}
\item \href{#method-toJSONString}{\code{IntegrationHub$toJSONString()}}
\item \href{#method-fromJSONString}{\code{IntegrationHub$fromJSONString()}}
\item \href{#method-clone}{\code{IntegrationHub$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntegrationHub$new(
  id,
  url,
  label,
  official,
  fetch_error_message,
  authorization_token,
  has_authorization_token,
  legal_agreement_signed,
  legal_agreement_required,
  legal_agreement_text,
  can
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntegrationHub$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntegrationHub$fromJSON(IntegrationHubJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntegrationHub$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntegrationHub$fromJSONString(IntegrationHubJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntegrationHub$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
