% Generated by roxygen2 (4.0.2): do not edit by hand
\name{InterpolateMissingFrames}
\alias{InterpolateMissingFrames}
\title{Interpolate missing gaze data}
\usage{
InterpolateMissingFrames(x, window = lwl_opts$get("interpolation_window"))
}
\arguments{
\item{x}{a Trial with a \code{GazeByImageAOI} column or a TrialList where
each Trial has a \code{GazeByImageAOI} column}

\item{window}{the time window (in ms.) to interpolate over}
}
\value{
the Trial object(s) with missing data interpolated
}
\description{
Interpolate missing gaze data
}

