% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LegacyFeature.r
\name{LegacyFeature}
\alias{LegacyFeature}
\title{LegacyFeature Class}
\description{
LegacyFeature Class

LegacyFeature Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{can}}{}

\item{\code{id}}{}

\item{\code{name}}{}

\item{\code{description}}{}

\item{\code{enabled_locally}}{}

\item{\code{enabled}}{}

\item{\code{disallowed_as_of_version}}{}

\item{\code{disable_on_upgrade_to_version}}{}

\item{\code{end_of_life_version}}{}

\item{\code{documentation_url}}{}

\item{\code{approximate_disable_date}}{}

\item{\code{approximate_end_of_life_date}}{}

\item{\code{has_disabled_on_upgrade}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LegacyFeature$new()}}
\item \href{#method-toJSON}{\code{LegacyFeature$toJSON()}}
\item \href{#method-fromJSONObject}{\code{LegacyFeature$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{LegacyFeature$fromJSON()}}
\item \href{#method-toJSONString}{\code{LegacyFeature$toJSONString()}}
\item \href{#method-fromJSONString}{\code{LegacyFeature$fromJSONString()}}
\item \href{#method-clone}{\code{LegacyFeature$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LegacyFeature$new(
  can,
  id,
  name,
  description,
  enabled_locally,
  enabled,
  disallowed_as_of_version,
  disable_on_upgrade_to_version,
  end_of_life_version,
  documentation_url,
  approximate_disable_date,
  approximate_end_of_life_date,
  has_disabled_on_upgrade
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LegacyFeature$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LegacyFeature$fromJSONObject(LegacyFeatureJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LegacyFeature$fromJSON(LegacyFeatureJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LegacyFeature$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LegacyFeature$fromJSONString(LegacyFeatureJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LegacyFeature$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
