% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Look.r
\name{Look}
\alias{Look}
\title{Look Class}
\description{
Look Class

Look Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{can}}{}

\item{\code{content_metadata_id}}{}

\item{\code{id}}{}

\item{\code{title}}{}

\item{\code{user_id}}{}

\item{\code{content_favorite_id}}{}

\item{\code{created_at}}{}

\item{\code{deleted}}{}

\item{\code{deleted_at}}{}

\item{\code{deleter_id}}{}

\item{\code{description}}{}

\item{\code{embed_url}}{}

\item{\code{excel_file_url}}{}

\item{\code{favorite_count}}{}

\item{\code{google_spreadsheet_formula}}{}

\item{\code{image_embed_url}}{}

\item{\code{is_run_on_load}}{}

\item{\code{last_accessed_at}}{}

\item{\code{last_updater_id}}{}

\item{\code{last_viewed_at}}{}

\item{\code{model}}{}

\item{\code{public}}{}

\item{\code{public_slug}}{}

\item{\code{public_url}}{}

\item{\code{query_id}}{}

\item{\code{short_url}}{}

\item{\code{folder}}{}

\item{\code{folder_id}}{}

\item{\code{updated_at}}{}

\item{\code{view_count}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Look$new()}}
\item \href{#method-toJSON}{\code{Look$toJSON()}}
\item \href{#method-fromJSONObject}{\code{Look$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{Look$fromJSON()}}
\item \href{#method-toJSONString}{\code{Look$toJSONString()}}
\item \href{#method-fromJSONString}{\code{Look$fromJSONString()}}
\item \href{#method-clone}{\code{Look$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Look$new(
  can,
  content_metadata_id,
  id,
  title,
  user_id,
  content_favorite_id,
  created_at,
  deleted,
  deleted_at,
  deleter_id,
  description,
  embed_url,
  excel_file_url,
  favorite_count,
  google_spreadsheet_formula,
  image_embed_url,
  is_run_on_load,
  last_accessed_at,
  last_updater_id,
  last_viewed_at,
  model,
  public,
  public_slug,
  public_url,
  query_id,
  short_url,
  folder,
  folder_id,
  updated_at,
  view_count
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Look$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Look$fromJSONObject(LookJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Look$fromJSON(LookJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Look$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Look$fromJSONString(LookJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Look$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
