% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookApi.r
\name{LookApi}
\alias{LookApi}
\title{Look operations}
\description{
looker.Look
}
\section{Methods}{

\describe{

all_looks Get All Looks 


create_look Create Look 


delete_look Delete Look 


look Get Look 


run_look Run Look 


search_looks Search Looks 


update_look Update Look 

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LookApi$new()}}
\item \href{#method-all_looks}{\code{LookApi$all_looks()}}
\item \href{#method-create_look}{\code{LookApi$create_look()}}
\item \href{#method-delete_look}{\code{LookApi$delete_look()}}
\item \href{#method-look}{\code{LookApi$look()}}
\item \href{#method-run_look}{\code{LookApi$run_look()}}
\item \href{#method-search_looks}{\code{LookApi$search_looks()}}
\item \href{#method-update_look}{\code{LookApi$update_look()}}
\item \href{#method-clone}{\code{LookApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_looks"></a>}}
\if{latex}{\out{\hypertarget{method-all_looks}{}}}
\subsection{Method \code{all_looks()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookApi$all_looks(fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_look"></a>}}
\if{latex}{\out{\hypertarget{method-create_look}{}}}
\subsection{Method \code{create_look()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookApi$create_look(body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_look"></a>}}
\if{latex}{\out{\hypertarget{method-delete_look}{}}}
\subsection{Method \code{delete_look()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookApi$delete_look(look_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-look"></a>}}
\if{latex}{\out{\hypertarget{method-look}{}}}
\subsection{Method \code{look()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookApi$look(look_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_look"></a>}}
\if{latex}{\out{\hypertarget{method-run_look}{}}}
\subsection{Method \code{run_look()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookApi$run_look(
  look_id,
  result_format,
  limit,
  apply_formatting,
  apply_vis,
  cache,
  image_width,
  image_height,
  generate_drill_links,
  force_production,
  cache_only,
  path_prefix,
  rebuild_pdts,
  server_table_calcs,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_looks"></a>}}
\if{latex}{\out{\hypertarget{method-search_looks}{}}}
\subsection{Method \code{search_looks()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookApi$search_looks(
  fields,
  page,
  per_page,
  limit,
  offset,
  sorts,
  title,
  description,
  content_favorite_id,
  space_id,
  user_id,
  view_count,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_look"></a>}}
\if{latex}{\out{\hypertarget{method-update_look}{}}}
\subsection{Method \code{update_look()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookApi$update_look(look_id, body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
