% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookmlModelApi.r
\docType{data}
\name{LookmlModelApi}
\alias{LookmlModelApi}
\title{LookmlModel operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LookmlModelApi
}
\description{
looker.LookmlModel
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

all_lookml_models Get All LookML Models 


create_lookml_model Create LookML Model 


delete_lookml_model Delete LookML Model 


lookml_model Get LookML Model 


lookml_model_explore Get LookML Model Explore 


update_lookml_model Update LookML Model 

}
}

\keyword{datasets}
