% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookmlModelExplore.r
\name{LookmlModelExplore}
\alias{LookmlModelExplore}
\title{LookmlModelExplore Class}
\description{
LookmlModelExplore Class

LookmlModelExplore Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{}

\item{\code{name}}{}

\item{\code{description}}{}

\item{\code{label}}{}

\item{\code{scopes}}{}

\item{\code{can_total}}{}

\item{\code{can_save}}{}

\item{\code{can_explain}}{}

\item{\code{can_pivot_in_db}}{}

\item{\code{has_timezone_support}}{}

\item{\code{supports_cost_estimate}}{}

\item{\code{connection_name}}{}

\item{\code{null_sort_treatment}}{}

\item{\code{files}}{}

\item{\code{source_file}}{}

\item{\code{project_name}}{}

\item{\code{model_name}}{}

\item{\code{view_name}}{}

\item{\code{hidden}}{}

\item{\code{sql_table_name}}{}

\item{\code{access_filter_fields}}{}

\item{\code{access_filters}}{}

\item{\code{aliases}}{}

\item{\code{always_filter}}{}

\item{\code{conditionally_filter}}{}

\item{\code{index_fields}}{}

\item{\code{sets}}{}

\item{\code{errors}}{}

\item{\code{fields}}{}

\item{\code{joins}}{}

\item{\code{group_label}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LookmlModelExplore$new()}}
\item \href{#method-toJSON}{\code{LookmlModelExplore$toJSON()}}
\item \href{#method-fromJSON}{\code{LookmlModelExplore$fromJSON()}}
\item \href{#method-toJSONString}{\code{LookmlModelExplore$toJSONString()}}
\item \href{#method-fromJSONString}{\code{LookmlModelExplore$fromJSONString()}}
\item \href{#method-clone}{\code{LookmlModelExplore$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExplore$new(
  id,
  name,
  description,
  label,
  scopes,
  can_total,
  can_save,
  can_explain,
  can_pivot_in_db,
  has_timezone_support,
  supports_cost_estimate,
  connection_name,
  null_sort_treatment,
  files,
  source_file,
  project_name,
  model_name,
  view_name,
  hidden,
  sql_table_name,
  access_filter_fields,
  access_filters,
  aliases,
  always_filter,
  conditionally_filter,
  index_fields,
  sets,
  errors,
  fields,
  joins,
  group_label
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExplore$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExplore$fromJSON(LookmlModelExploreJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExplore$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExplore$fromJSONString(LookmlModelExploreJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExplore$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
