% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookmlModelExploreAccessFilter.r
\name{LookmlModelExploreAccessFilter}
\alias{LookmlModelExploreAccessFilter}
\title{LookmlModelExploreAccessFilter Class}
\description{
LookmlModelExploreAccessFilter Class

LookmlModelExploreAccessFilter Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{field}}{}

\item{\code{user_attribute}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LookmlModelExploreAccessFilter$new()}}
\item \href{#method-toJSON}{\code{LookmlModelExploreAccessFilter$toJSON()}}
\item \href{#method-fromJSONObject}{\code{LookmlModelExploreAccessFilter$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{LookmlModelExploreAccessFilter$fromJSON()}}
\item \href{#method-toJSONString}{\code{LookmlModelExploreAccessFilter$toJSONString()}}
\item \href{#method-fromJSONString}{\code{LookmlModelExploreAccessFilter$fromJSONString()}}
\item \href{#method-clone}{\code{LookmlModelExploreAccessFilter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreAccessFilter$new(field, user_attribute)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreAccessFilter$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreAccessFilter$fromJSONObject(
  LookmlModelExploreAccessFilterJsonObject
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreAccessFilter$fromJSON(LookmlModelExploreAccessFilterJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreAccessFilter$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreAccessFilter$fromJSONString(
  LookmlModelExploreAccessFilterJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreAccessFilter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
