% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookmlModelExploreField.r
\docType{data}
\name{LookmlModelExploreField}
\alias{LookmlModelExploreField}
\title{LookmlModelExploreField Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LookmlModelExploreField
}
\description{
LookmlModelExploreField Class
}
\section{Fields}{

\describe{
\item{\code{align}}{}

\item{\code{can_filter}}{}

\item{\code{category}}{}

\item{\code{default_filter_value}}{}

\item{\code{description}}{}

\item{\code{enumerations}}{}

\item{\code{error}}{}

\item{\code{field_group_label}}{}

\item{\code{field_group_variant}}{}

\item{\code{fill_style}}{}

\item{\code{fiscal_month_offset}}{}

\item{\code{has_allowed_values}}{}

\item{\code{hidden}}{}

\item{\code{is_filter}}{}

\item{\code{is_numeric}}{}

\item{\code{is_timeframe}}{}

\item{\code{can_time_filter}}{}

\item{\code{time_interval}}{}

\item{\code{label}}{}

\item{\code{label_from_parameter}}{}

\item{\code{label_short}}{}

\item{\code{lookml_link}}{}

\item{\code{map_layer}}{}

\item{\code{measure}}{}

\item{\code{name}}{}

\item{\code{parameter}}{}

\item{\code{primary_key}}{}

\item{\code{project_name}}{}

\item{\code{requires_refresh_on_sort}}{}

\item{\code{scope}}{}

\item{\code{sortable}}{}

\item{\code{source_file}}{}

\item{\code{source_file_path}}{}

\item{\code{sql}}{}

\item{\code{sql_case}}{}

\item{\code{suggest_dimension}}{}

\item{\code{suggest_explore}}{}

\item{\code{suggestable}}{}

\item{\code{suggestions}}{}

\item{\code{supported_measure_types}}{}

\item{\code{tags}}{}

\item{\code{type}}{}

\item{\code{user_attribute_filter_types}}{}

\item{\code{value_format}}{}

\item{\code{view}}{}

\item{\code{view_label}}{}

\item{\code{dynamic}}{}
}}

\keyword{datasets}
