% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookmlModelExploreFieldMapLayer.r
\name{LookmlModelExploreFieldMapLayer}
\alias{LookmlModelExploreFieldMapLayer}
\title{LookmlModelExploreFieldMapLayer Class}
\description{
LookmlModelExploreFieldMapLayer Class

LookmlModelExploreFieldMapLayer Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{url}}{}

\item{\code{feature_key}}{}

\item{\code{property_key}}{}

\item{\code{property_label_key}}{}

\item{\code{projection}}{}

\item{\code{format}}{}

\item{\code{extents_json_url}}{}

\item{\code{max_zoom_level}}{}

\item{\code{min_zoom_level}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LookmlModelExploreFieldMapLayer$new()}}
\item \href{#method-toJSON}{\code{LookmlModelExploreFieldMapLayer$toJSON()}}
\item \href{#method-fromJSON}{\code{LookmlModelExploreFieldMapLayer$fromJSON()}}
\item \href{#method-toJSONString}{\code{LookmlModelExploreFieldMapLayer$toJSONString()}}
\item \href{#method-fromJSONString}{\code{LookmlModelExploreFieldMapLayer$fromJSONString()}}
\item \href{#method-clone}{\code{LookmlModelExploreFieldMapLayer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreFieldMapLayer$new(
  url,
  feature_key,
  property_key,
  property_label_key,
  projection,
  format,
  extents_json_url,
  max_zoom_level,
  min_zoom_level
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreFieldMapLayer$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreFieldMapLayer$fromJSON(LookmlModelExploreFieldMapLayerJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreFieldMapLayer$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreFieldMapLayer$fromJSONString(
  LookmlModelExploreFieldMapLayerJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreFieldMapLayer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
