% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookmlModelExploreFieldTimeInterval.r
\docType{data}
\name{LookmlModelExploreFieldTimeInterval}
\alias{LookmlModelExploreFieldTimeInterval}
\title{LookmlModelExploreFieldTimeInterval Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LookmlModelExploreFieldTimeInterval
}
\description{
LookmlModelExploreFieldTimeInterval Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{count}}{}
}}

\keyword{datasets}
