% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookmlModelExploreJoins.r
\name{LookmlModelExploreJoins}
\alias{LookmlModelExploreJoins}
\title{LookmlModelExploreJoins Class}
\description{
LookmlModelExploreJoins Class

LookmlModelExploreJoins Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{}

\item{\code{dependent_fields}}{}

\item{\code{fields}}{}

\item{\code{foreign_key}}{}

\item{\code{from}}{}

\item{\code{outer_only}}{}

\item{\code{relationship}}{}

\item{\code{required_joins}}{}

\item{\code{sql_foreign_key}}{}

\item{\code{sql_on}}{}

\item{\code{sql_table_name}}{}

\item{\code{type}}{}

\item{\code{view_label}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LookmlModelExploreJoins$new()}}
\item \href{#method-toJSON}{\code{LookmlModelExploreJoins$toJSON()}}
\item \href{#method-fromJSON}{\code{LookmlModelExploreJoins$fromJSON()}}
\item \href{#method-toJSONString}{\code{LookmlModelExploreJoins$toJSONString()}}
\item \href{#method-fromJSONString}{\code{LookmlModelExploreJoins$fromJSONString()}}
\item \href{#method-clone}{\code{LookmlModelExploreJoins$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreJoins$new(
  name,
  dependent_fields,
  fields,
  foreign_key,
  from,
  outer_only,
  relationship,
  required_joins,
  sql_foreign_key,
  sql_on,
  sql_table_name,
  type,
  view_label
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreJoins$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreJoins$fromJSON(LookmlModelExploreJoinsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreJoins$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreJoins$fromJSONString(LookmlModelExploreJoinsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LookmlModelExploreJoins$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
