% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookmlModelNavExplore.r
\docType{data}
\name{LookmlModelNavExplore}
\alias{LookmlModelNavExplore}
\title{LookmlModelNavExplore Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LookmlModelNavExplore
}
\description{
LookmlModelNavExplore Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{label}}{}

\item{\code{hidden}}{}

\item{\code{group_label}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
