% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookWithDashboards.r
\docType{data}
\name{LookWithDashboards}
\alias{LookWithDashboards}
\title{LookWithDashboards Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LookWithDashboards
}
\description{
LookWithDashboards Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{content_metadata_id}}{}

\item{\code{view_count}}{}

\item{\code{favorite_count}}{}

\item{\code{content_favorite_id}}{}

\item{\code{title}}{}

\item{\code{user}}{}

\item{\code{query_id}}{}

\item{\code{description}}{}

\item{\code{short_url}}{}

\item{\code{space}}{}

\item{\code{public}}{}

\item{\code{public_slug}}{}

\item{\code{user_id}}{}

\item{\code{space_id}}{}

\item{\code{model}}{}

\item{\code{public_url}}{}

\item{\code{embed_url}}{}

\item{\code{google_spreadsheet_formula}}{}

\item{\code{excel_file_url}}{}

\item{\code{created_at}}{}

\item{\code{updated_at}}{}

\item{\code{last_updater_id}}{}

\item{\code{last_viewed_at}}{}

\item{\code{last_accessed_at}}{}

\item{\code{deleter_id}}{}

\item{\code{deleted}}{}

\item{\code{deleted_at}}{}

\item{\code{is_run_on_load}}{}

\item{\code{dashboards}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
