\docType{data}
\name{lwl_constants}
\alias{lwl_constants}
\title{Package-wide constants}
\format{\preformatted{List of 6
 $ ms_per_frame   : num 16.7
 $ screen_width   : num 1920
 $ screen_height  : num 1200
 $ pixels_per_mm  : num 3.71
 $ l2t_subject_dir: chr "^[0-9]\{3\}[CLPXclpx]"
 $ l2t_subject    : chr "[0-9]\{3\}[CLPDE][0-9]\{2\}[MFX][SA][1-9]"
}}
\usage{
lwl_constants
}
\description{
Package-wide constants
}
\keyword{datasets}

