% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSet.r
\docType{data}
\name{ModelSet}
\alias{ModelSet}
\title{ModelSet Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ModelSet
}
\description{
ModelSet Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{name}}{}

\item{\code{models}}{}

\item{\code{built_in}}{}

\item{\code{all_access}}{}

\item{\code{url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
