% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prefetch.r
\docType{data}
\name{Prefetch}
\alias{Prefetch}
\title{Prefetch Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Prefetch
}
\description{
Prefetch Class
}
\section{Fields}{

\describe{
\item{\code{ttl}}{}

\item{\code{created_at}}{}

\item{\code{computation_time}}{}

\item{\code{result_size_bytes}}{}

\item{\code{hit_count}}{}

\item{\code{touched_at}}{}

\item{\code{value}}{}

\item{\code{url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
