% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrefetchDashboardRequest.r
\docType{data}
\name{PrefetchDashboardRequest}
\alias{PrefetchDashboardRequest}
\title{PrefetchDashboardRequest Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PrefetchDashboardRequest
}
\description{
PrefetchDashboardRequest Class
}
\section{Fields}{

\describe{
\item{\code{ttl}}{}

\item{\code{access_filters}}{}

\item{\code{dashboard_filters}}{}

\item{\code{created_at}}{}

\item{\code{computation_time}}{}

\item{\code{result_size_bytes}}{}

\item{\code{hit_count}}{}

\item{\code{touched_at}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
