% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project.r
\docType{data}
\name{Project}
\alias{Project}
\title{Project Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Project
}
\description{
Project Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{name}}{}

\item{\code{uses_git}}{}

\item{\code{git_remote_url}}{}

\item{\code{git_service_name}}{}

\item{\code{pull_request_mode}}{}

\item{\code{validation_required}}{}

\item{\code{is_example}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
