% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectApi.r
\name{ProjectApi}
\alias{ProjectApi}
\title{Project operations}
\description{
looker.Project
}
\section{Methods}{

\describe{

all_git_branches Get All Git Branches


all_git_connection_tests Get All Git Connection Tests


all_lookml_tests Get All LookML Tests


all_project_files Get All Project Files


all_projects Get All Projects


create_git_branch Checkout New Git Branch


create_git_deploy_key Create Deploy Key


create_project Create Project


delete_git_branch Delete a Git Branch


delete_repository_credential Delete Repository Credential


deploy_ref_to_production Deploy Remote Branch or Ref to Production


deploy_to_production Deploy To Production


find_git_branch Find a Git Branch


get_all_repository_credentials Get All Repository Credentials


git_branch Get Active Git Branch


git_deploy_key Git Deploy Key


lock_all Lock All


manifest Get Manifest


project Get Project


project_file Get Project File


project_validation_results Cached Project Validation Results


project_workspace Get Project Workspace


reset_project_to_production Reset To Production


reset_project_to_remote Reset To Remote


run_git_connection_test Run Git Connection Test


run_lookml_test Run LookML Test


tag_ref Tag Ref


update_git_branch Update Project Git Branch


update_project Update Project


update_repository_credential Create Repository Credential


validate_project Validate Project

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ProjectApi$new()}}
\item \href{#method-all_git_branches}{\code{ProjectApi$all_git_branches()}}
\item \href{#method-all_git_connection_tests}{\code{ProjectApi$all_git_connection_tests()}}
\item \href{#method-all_lookml_tests}{\code{ProjectApi$all_lookml_tests()}}
\item \href{#method-all_project_files}{\code{ProjectApi$all_project_files()}}
\item \href{#method-all_projects}{\code{ProjectApi$all_projects()}}
\item \href{#method-create_git_branch}{\code{ProjectApi$create_git_branch()}}
\item \href{#method-create_git_deploy_key}{\code{ProjectApi$create_git_deploy_key()}}
\item \href{#method-create_project}{\code{ProjectApi$create_project()}}
\item \href{#method-delete_git_branch}{\code{ProjectApi$delete_git_branch()}}
\item \href{#method-delete_repository_credential}{\code{ProjectApi$delete_repository_credential()}}
\item \href{#method-deploy_ref_to_production}{\code{ProjectApi$deploy_ref_to_production()}}
\item \href{#method-deploy_to_production}{\code{ProjectApi$deploy_to_production()}}
\item \href{#method-find_git_branch}{\code{ProjectApi$find_git_branch()}}
\item \href{#method-get_all_repository_credentials}{\code{ProjectApi$get_all_repository_credentials()}}
\item \href{#method-git_branch}{\code{ProjectApi$git_branch()}}
\item \href{#method-git_deploy_key}{\code{ProjectApi$git_deploy_key()}}
\item \href{#method-lock_all}{\code{ProjectApi$lock_all()}}
\item \href{#method-manifest}{\code{ProjectApi$manifest()}}
\item \href{#method-project}{\code{ProjectApi$project()}}
\item \href{#method-project_file}{\code{ProjectApi$project_file()}}
\item \href{#method-project_validation_results}{\code{ProjectApi$project_validation_results()}}
\item \href{#method-project_workspace}{\code{ProjectApi$project_workspace()}}
\item \href{#method-reset_project_to_production}{\code{ProjectApi$reset_project_to_production()}}
\item \href{#method-reset_project_to_remote}{\code{ProjectApi$reset_project_to_remote()}}
\item \href{#method-run_git_connection_test}{\code{ProjectApi$run_git_connection_test()}}
\item \href{#method-run_lookml_test}{\code{ProjectApi$run_lookml_test()}}
\item \href{#method-tag_ref}{\code{ProjectApi$tag_ref()}}
\item \href{#method-update_git_branch}{\code{ProjectApi$update_git_branch()}}
\item \href{#method-update_project}{\code{ProjectApi$update_project()}}
\item \href{#method-update_repository_credential}{\code{ProjectApi$update_repository_credential()}}
\item \href{#method-validate_project}{\code{ProjectApi$validate_project()}}
\item \href{#method-clone}{\code{ProjectApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_git_branches"></a>}}
\if{latex}{\out{\hypertarget{method-all_git_branches}{}}}
\subsection{Method \code{all_git_branches()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectApi$all_git_branches(project_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_git_connection_tests"></a>}}
\if{latex}{\out{\hypertarget{method-all_git_connection_tests}{}}}
\subsection{Method \code{all_git_connection_tests()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectApi$all_git_connection_tests(project_id, remote_url, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_lookml_tests"></a>}}
\if{latex}{\out{\hypertarget{method-all_lookml_tests}{}}}
\subsection{Method \code{all_lookml_tests()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectApi$all_lookml_tests(project_id, file_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_project_files"></a>}}
\if{latex}{\out{\hypertarget{method-all_project_files}{}}}
\subsection{Method \code{all_project_files()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectApi$all_project_files(project_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_projects"></a>}}
\if{latex}{\out{\hypertarget{method-all_projects}{}}}
\subsection{Method \code{all_projects()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectApi$all_projects(fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_git_branch"></a>}}
\if{latex}{\out{\hypertarget{method-create_git_branch}{}}}
\subsection{Method \code{create_git_branch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectApi$create_git_branch(project_id, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_git_deploy_key"></a>}}
\if{latex}{\out{\hypertarget{method-create_git_deploy_key}{}}}
\subsection{Method \code{create_git_deploy_key()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectApi$create_git_deploy_key(project_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_project"></a>}}
\if{latex}{\out{\hypertarget{method-create_project}{}}}
\subsection{Method \code{create_project()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectApi$create_project(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_git_branch"></a>}}
\if{latex}{\out{\hypertarget{method-delete_git_branch}{}}}
\subsection{Method \code{delete_git_branch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectApi$delete_git_branch(project_id, branch_name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_repository_credential"></a>}}
\if{latex}{\out{\hypertarget{method-delete_repository_credential}{}}}
\subsection{Method \code{delete_repository_credential()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectApi$delete_repository_credential(root_project_id, credential_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deploy_ref_to_production"></a>}}
\if{latex}{\out{\hypertarget{method-deploy_ref_to_production}{}}}
\subsection{Method \code{deploy_ref_to_production()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectApi$deploy_ref_to_production(project_id, branch, ref, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deploy_to_production"></a>}}
\if{latex}{\out{\hypertarget{method-deploy_to_production}{}}}
\subsection{Method \code{deploy_to_production()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectApi$deploy_to_production(project_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-find_git_branch"></a>}}
\if{latex}{\out{\hypertarget{method-find_git_branch}{}}}
\subsection{Method \code{find_git_branch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectApi$find_git_branch(project_id, branch_name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_all_repository_credentials"></a>}}
\if{latex}{\out{\hypertarget{method-get_all_repository_credentials}{}}}
\subsection{Method \code{get_all_repository_credentials()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectApi$get_all_repository_credentials(root_project_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-git_branch"></a>}}
\if{latex}{\out{\hypertarget{method-git_branch}{}}}
\subsection{Method \code{git_branch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectApi$git_branch(project_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-git_deploy_key"></a>}}
\if{latex}{\out{\hypertarget{method-git_deploy_key}{}}}
\subsection{Method \code{git_deploy_key()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectApi$git_deploy_key(project_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lock_all"></a>}}
\if{latex}{\out{\hypertarget{method-lock_all}{}}}
\subsection{Method \code{lock_all()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectApi$lock_all(project_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-manifest"></a>}}
\if{latex}{\out{\hypertarget{method-manifest}{}}}
\subsection{Method \code{manifest()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectApi$manifest(project_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-project"></a>}}
\if{latex}{\out{\hypertarget{method-project}{}}}
\subsection{Method \code{project()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectApi$project(project_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-project_file"></a>}}
\if{latex}{\out{\hypertarget{method-project_file}{}}}
\subsection{Method \code{project_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectApi$project_file(project_id, file_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-project_validation_results"></a>}}
\if{latex}{\out{\hypertarget{method-project_validation_results}{}}}
\subsection{Method \code{project_validation_results()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectApi$project_validation_results(project_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-project_workspace"></a>}}
\if{latex}{\out{\hypertarget{method-project_workspace}{}}}
\subsection{Method \code{project_workspace()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectApi$project_workspace(project_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset_project_to_production"></a>}}
\if{latex}{\out{\hypertarget{method-reset_project_to_production}{}}}
\subsection{Method \code{reset_project_to_production()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectApi$reset_project_to_production(project_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset_project_to_remote"></a>}}
\if{latex}{\out{\hypertarget{method-reset_project_to_remote}{}}}
\subsection{Method \code{reset_project_to_remote()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectApi$reset_project_to_remote(project_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_git_connection_test"></a>}}
\if{latex}{\out{\hypertarget{method-run_git_connection_test}{}}}
\subsection{Method \code{run_git_connection_test()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectApi$run_git_connection_test(
  project_id,
  test_id,
  remote_url,
  use_production,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_lookml_test"></a>}}
\if{latex}{\out{\hypertarget{method-run_lookml_test}{}}}
\subsection{Method \code{run_lookml_test()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectApi$run_lookml_test(project_id, file_id, test, model, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tag_ref"></a>}}
\if{latex}{\out{\hypertarget{method-tag_ref}{}}}
\subsection{Method \code{tag_ref()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectApi$tag_ref(project_id, body, commit_sha, tag_name, tag_message, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_git_branch"></a>}}
\if{latex}{\out{\hypertarget{method-update_git_branch}{}}}
\subsection{Method \code{update_git_branch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectApi$update_git_branch(project_id, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_project"></a>}}
\if{latex}{\out{\hypertarget{method-update_project}{}}}
\subsection{Method \code{update_project()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectApi$update_project(project_id, body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_repository_credential"></a>}}
\if{latex}{\out{\hypertarget{method-update_repository_credential}{}}}
\subsection{Method \code{update_repository_credential()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectApi$update_repository_credential(
  root_project_id,
  credential_id,
  body,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate_project"></a>}}
\if{latex}{\out{\hypertarget{method-validate_project}{}}}
\subsection{Method \code{validate_project()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectApi$validate_project(project_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
