% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectError.r
\name{ProjectError}
\alias{ProjectError}
\title{ProjectError Class}
\description{
ProjectError Class

ProjectError Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{code}}{}

\item{\code{severity}}{}

\item{\code{kind}}{}

\item{\code{message}}{}

\item{\code{field_name}}{}

\item{\code{file_path}}{}

\item{\code{line_number}}{}

\item{\code{model_id}}{}

\item{\code{explore}}{}

\item{\code{help_url}}{}

\item{\code{params}}{}

\item{\code{sanitized_message}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ProjectError$new()}}
\item \href{#method-toJSON}{\code{ProjectError$toJSON()}}
\item \href{#method-fromJSON}{\code{ProjectError$fromJSON()}}
\item \href{#method-toJSONString}{\code{ProjectError$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ProjectError$fromJSONString()}}
\item \href{#method-clone}{\code{ProjectError$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectError$new(
  code,
  severity,
  kind,
  message,
  field_name,
  file_path,
  line_number,
  model_id,
  explore,
  help_url,
  params,
  sanitized_message
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectError$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectError$fromJSON(ProjectErrorJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectError$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectError$fromJSONString(ProjectErrorJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectError$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
