% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectFile.r
\docType{data}
\name{ProjectFile}
\alias{ProjectFile}
\title{ProjectFile Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ProjectFile
}
\description{
ProjectFile Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{path}}{}

\item{\code{title}}{}

\item{\code{type}}{}

\item{\code{mime_type}}{}

\item{\code{git_status}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
