% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectWorkspace.r
\name{ProjectWorkspace}
\alias{ProjectWorkspace}
\title{ProjectWorkspace Class}
\description{
ProjectWorkspace Class

ProjectWorkspace Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{can}}{}

\item{\code{project_id}}{}

\item{\code{workspace_id}}{}

\item{\code{git_status}}{}

\item{\code{git_head}}{}

\item{\code{dependency_status}}{}

\item{\code{git_branch}}{}

\item{\code{lookml_type}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ProjectWorkspace$new()}}
\item \href{#method-toJSON}{\code{ProjectWorkspace$toJSON()}}
\item \href{#method-fromJSONObject}{\code{ProjectWorkspace$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{ProjectWorkspace$fromJSON()}}
\item \href{#method-toJSONString}{\code{ProjectWorkspace$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ProjectWorkspace$fromJSONString()}}
\item \href{#method-clone}{\code{ProjectWorkspace$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectWorkspace$new(
  can,
  project_id,
  workspace_id,
  git_status,
  git_head,
  dependency_status,
  git_branch,
  lookml_type
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectWorkspace$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectWorkspace$fromJSONObject(ProjectWorkspaceJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectWorkspace$fromJSON(ProjectWorkspaceJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectWorkspace$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectWorkspace$fromJSONString(ProjectWorkspaceJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectWorkspace$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
