% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.r
\docType{data}
\name{Query}
\alias{Query}
\title{Query Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Query
}
\description{
Query Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{model}}{}

\item{\code{view}}{}

\item{\code{fields}}{}

\item{\code{pivots}}{}

\item{\code{fill_fields}}{}

\item{\code{filters}}{}

\item{\code{filter_expression}}{}

\item{\code{sorts}}{}

\item{\code{limit}}{}

\item{\code{column_limit}}{}

\item{\code{total}}{}

\item{\code{row_total}}{}

\item{\code{runtime}}{}

\item{\code{vis_config}}{}

\item{\code{filter_config}}{}

\item{\code{visible_ui_sections}}{}

\item{\code{slug}}{}

\item{\code{dynamic_fields}}{}

\item{\code{client_id}}{}

\item{\code{share_url}}{}

\item{\code{expanded_share_url}}{}

\item{\code{url}}{}

\item{\code{query_timezone}}{}

\item{\code{has_table_calculations}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
