% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryApi.r
\name{QueryApi}
\alias{QueryApi}
\title{Query operations}
\description{
looker.Query
}
\section{Methods}{

\describe{

create_query Create Query 


create_query_task Run Query Async 


query Get Query 


query_for_slug Get Query for Slug 


query_task Get Async Query Info 


query_task_multi_results Get Multiple Async Query Results 


query_task_results Get Async Query Results 


run_inline_query Run Inline Query 


run_query Run Query 


run_url_encoded_query Run Url Encoded Query 

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{QueryApi$new()}}
\item \href{#method-create_query}{\code{QueryApi$create_query()}}
\item \href{#method-create_query_task}{\code{QueryApi$create_query_task()}}
\item \href{#method-query}{\code{QueryApi$query()}}
\item \href{#method-query_for_slug}{\code{QueryApi$query_for_slug()}}
\item \href{#method-query_task}{\code{QueryApi$query_task()}}
\item \href{#method-query_task_multi_results}{\code{QueryApi$query_task_multi_results()}}
\item \href{#method-query_task_results}{\code{QueryApi$query_task_results()}}
\item \href{#method-run_inline_query}{\code{QueryApi$run_inline_query()}}
\item \href{#method-run_query}{\code{QueryApi$run_query()}}
\item \href{#method-run_url_encoded_query}{\code{QueryApi$run_url_encoded_query()}}
\item \href{#method-clone}{\code{QueryApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_query"></a>}}
\if{latex}{\out{\hypertarget{method-create_query}{}}}
\subsection{Method \code{create_query()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryApi$create_query(body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_query_task"></a>}}
\if{latex}{\out{\hypertarget{method-create_query_task}{}}}
\subsection{Method \code{create_query_task()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryApi$create_query_task(
  body,
  limit,
  apply_formatting,
  apply_vis,
  cache,
  image_width,
  image_height,
  generate_drill_links,
  force_production,
  cache_only,
  path_prefix,
  rebuild_pdts,
  server_table_calcs,
  fields,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-query"></a>}}
\if{latex}{\out{\hypertarget{method-query}{}}}
\subsection{Method \code{query()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryApi$query(query_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-query_for_slug"></a>}}
\if{latex}{\out{\hypertarget{method-query_for_slug}{}}}
\subsection{Method \code{query_for_slug()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryApi$query_for_slug(slug, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-query_task"></a>}}
\if{latex}{\out{\hypertarget{method-query_task}{}}}
\subsection{Method \code{query_task()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryApi$query_task(query_task_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-query_task_multi_results"></a>}}
\if{latex}{\out{\hypertarget{method-query_task_multi_results}{}}}
\subsection{Method \code{query_task_multi_results()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryApi$query_task_multi_results(query_task_ids, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-query_task_results"></a>}}
\if{latex}{\out{\hypertarget{method-query_task_results}{}}}
\subsection{Method \code{query_task_results()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryApi$query_task_results(query_task_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_inline_query"></a>}}
\if{latex}{\out{\hypertarget{method-run_inline_query}{}}}
\subsection{Method \code{run_inline_query()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryApi$run_inline_query(
  result_format,
  body,
  limit,
  apply_formatting,
  apply_vis,
  cache,
  image_width,
  image_height,
  generate_drill_links,
  force_production,
  cache_only,
  path_prefix,
  rebuild_pdts,
  server_table_calcs,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_query"></a>}}
\if{latex}{\out{\hypertarget{method-run_query}{}}}
\subsection{Method \code{run_query()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryApi$run_query(
  query_id,
  result_format,
  limit,
  apply_formatting,
  apply_vis,
  cache,
  image_width,
  image_height,
  generate_drill_links,
  force_production,
  cache_only,
  path_prefix,
  rebuild_pdts,
  server_table_calcs,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_url_encoded_query"></a>}}
\if{latex}{\out{\hypertarget{method-run_url_encoded_query}{}}}
\subsection{Method \code{run_url_encoded_query()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryApi$run_url_encoded_query(model_name, view_name, result_format, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
