% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryTask.r
\name{QueryTask}
\alias{QueryTask}
\title{QueryTask Class}
\description{
QueryTask Class

QueryTask Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{can}}{}

\item{\code{id}}{}

\item{\code{query_id}}{}

\item{\code{query}}{}

\item{\code{generate_links}}{}

\item{\code{force_production}}{}

\item{\code{path_prefix}}{}

\item{\code{cache}}{}

\item{\code{server_table_calcs}}{}

\item{\code{cache_only}}{}

\item{\code{cache_key}}{}

\item{\code{status}}{}

\item{\code{source}}{}

\item{\code{runtime}}{}

\item{\code{rebuild_pdts}}{}

\item{\code{result_source}}{}

\item{\code{look_id}}{}

\item{\code{dashboard_id}}{}

\item{\code{result_format}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{QueryTask$new()}}
\item \href{#method-toJSON}{\code{QueryTask$toJSON()}}
\item \href{#method-fromJSONObject}{\code{QueryTask$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{QueryTask$fromJSON()}}
\item \href{#method-toJSONString}{\code{QueryTask$toJSONString()}}
\item \href{#method-fromJSONString}{\code{QueryTask$fromJSONString()}}
\item \href{#method-clone}{\code{QueryTask$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryTask$new(
  can,
  id,
  query_id,
  query,
  generate_links,
  force_production,
  path_prefix,
  cache,
  server_table_calcs,
  cache_only,
  cache_key,
  status,
  source,
  runtime,
  rebuild_pdts,
  result_source,
  look_id,
  dashboard_id,
  result_format
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryTask$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryTask$fromJSONObject(QueryTaskJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryTask$fromJSON(QueryTaskJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryTask$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryTask$fromJSONString(QueryTaskJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryTask$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
