% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenderTask.r
\name{RenderTask}
\alias{RenderTask}
\title{RenderTask Class}
\description{
RenderTask Class

RenderTask Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{can}}{}

\item{\code{created_at}}{}

\item{\code{dashboard_filters}}{}

\item{\code{dashboard_id}}{}

\item{\code{dashboard_style}}{}

\item{\code{finalized_at}}{}

\item{\code{height}}{}

\item{\code{id}}{}

\item{\code{look_id}}{}

\item{\code{lookml_dashboard_id}}{}

\item{\code{query_id}}{}

\item{\code{dashboard_element_id}}{}

\item{\code{query_runtime}}{}

\item{\code{render_runtime}}{}

\item{\code{result_format}}{}

\item{\code{runtime}}{}

\item{\code{status}}{}

\item{\code{status_detail}}{}

\item{\code{user_id}}{}

\item{\code{width}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RenderTask$new()}}
\item \href{#method-toJSON}{\code{RenderTask$toJSON()}}
\item \href{#method-fromJSONObject}{\code{RenderTask$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{RenderTask$fromJSON()}}
\item \href{#method-toJSONString}{\code{RenderTask$toJSONString()}}
\item \href{#method-fromJSONString}{\code{RenderTask$fromJSONString()}}
\item \href{#method-clone}{\code{RenderTask$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RenderTask$new(
  can,
  created_at,
  dashboard_filters,
  dashboard_id,
  dashboard_style,
  finalized_at,
  height,
  id,
  look_id,
  lookml_dashboard_id,
  query_id,
  dashboard_element_id,
  query_runtime,
  render_runtime,
  result_format,
  runtime,
  status,
  status_detail,
  user_id,
  width
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RenderTask$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RenderTask$fromJSONObject(RenderTaskJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RenderTask$fromJSON(RenderTaskJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RenderTask$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RenderTask$fromJSONString(RenderTaskJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RenderTask$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
