% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenderTask.r
\docType{data}
\name{RenderTask}
\alias{RenderTask}
\title{RenderTask Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RenderTask
}
\description{
RenderTask Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{created_at}}{}

\item{\code{finalized_at}}{}

\item{\code{status}}{}

\item{\code{status_detail}}{}

\item{\code{user_id}}{}

\item{\code{runtime}}{}

\item{\code{query_runtime}}{}

\item{\code{render_runtime}}{}

\item{\code{result_format}}{}

\item{\code{look_id}}{}

\item{\code{dashboard_id}}{}

\item{\code{lookml_dashboard_id}}{}

\item{\code{query_id}}{}

\item{\code{width}}{}

\item{\code{height}}{}

\item{\code{dashboard_style}}{}

\item{\code{dashboard_filters}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
