% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenderTaskApi.r
\docType{data}
\name{RenderTaskApi}
\alias{RenderTaskApi}
\title{RenderTask operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RenderTaskApi
}
\description{
looker.RenderTask
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

create_dashboard_render_task Create Dashboard Render Task 


create_look_render_task Create Look Render Task 


create_lookml_dashboard_render_task Create Lookml Dashboard Render Task 


create_query_render_task Create Query Render Task 


render_task Get Render Task 


render_task_results Render Task Results 

}
}

\keyword{datasets}
