% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultMakerWithIdVisConfigAndDynamicFields.r
\docType{data}
\name{ResultMakerWithIdVisConfigAndDynamicFields}
\alias{ResultMakerWithIdVisConfigAndDynamicFields}
\title{ResultMakerWithIdVisConfigAndDynamicFields Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ResultMakerWithIdVisConfigAndDynamicFields
}
\description{
ResultMakerWithIdVisConfigAndDynamicFields Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{dynamic_fields}}{}

\item{\code{sorts}}{}

\item{\code{total}}{}

\item{\code{vis_config}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
