% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoleApi.r
\name{RoleApi}
\alias{RoleApi}
\title{Role operations}
\description{
looker.Role
}
\section{Methods}{

\describe{

all_model_sets Get All Model Sets 


all_permission_sets Get All Permission Sets 


all_permissions Get All Permissions 


all_roles Get All Roles 


create_model_set Create Model Set 


create_permission_set Create Permission Set 


create_role Create Role 


delete_model_set Delete Model Set 


delete_permission_set Delete Permission Set 


delete_role Delete Role 


model_set Get Model Set 


permission_set Get Permission Set 


role Get Role 


role_groups Get Role Groups 


role_users Get Role Users 


set_role_groups Update Role Groups 


set_role_users Update Role Users 


update_model_set Update Model Set 


update_permission_set Update Permission Set 


update_role Update Role 

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RoleApi$new()}}
\item \href{#method-all_model_sets}{\code{RoleApi$all_model_sets()}}
\item \href{#method-all_permission_sets}{\code{RoleApi$all_permission_sets()}}
\item \href{#method-all_permissions}{\code{RoleApi$all_permissions()}}
\item \href{#method-all_roles}{\code{RoleApi$all_roles()}}
\item \href{#method-create_model_set}{\code{RoleApi$create_model_set()}}
\item \href{#method-create_permission_set}{\code{RoleApi$create_permission_set()}}
\item \href{#method-create_role}{\code{RoleApi$create_role()}}
\item \href{#method-delete_model_set}{\code{RoleApi$delete_model_set()}}
\item \href{#method-delete_permission_set}{\code{RoleApi$delete_permission_set()}}
\item \href{#method-delete_role}{\code{RoleApi$delete_role()}}
\item \href{#method-model_set}{\code{RoleApi$model_set()}}
\item \href{#method-permission_set}{\code{RoleApi$permission_set()}}
\item \href{#method-role}{\code{RoleApi$role()}}
\item \href{#method-role_groups}{\code{RoleApi$role_groups()}}
\item \href{#method-role_users}{\code{RoleApi$role_users()}}
\item \href{#method-set_role_groups}{\code{RoleApi$set_role_groups()}}
\item \href{#method-set_role_users}{\code{RoleApi$set_role_users()}}
\item \href{#method-update_model_set}{\code{RoleApi$update_model_set()}}
\item \href{#method-update_permission_set}{\code{RoleApi$update_permission_set()}}
\item \href{#method-update_role}{\code{RoleApi$update_role()}}
\item \href{#method-clone}{\code{RoleApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RoleApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_model_sets"></a>}}
\if{latex}{\out{\hypertarget{method-all_model_sets}{}}}
\subsection{Method \code{all_model_sets()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RoleApi$all_model_sets(fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_permission_sets"></a>}}
\if{latex}{\out{\hypertarget{method-all_permission_sets}{}}}
\subsection{Method \code{all_permission_sets()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RoleApi$all_permission_sets(fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_permissions"></a>}}
\if{latex}{\out{\hypertarget{method-all_permissions}{}}}
\subsection{Method \code{all_permissions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RoleApi$all_permissions(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_roles"></a>}}
\if{latex}{\out{\hypertarget{method-all_roles}{}}}
\subsection{Method \code{all_roles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RoleApi$all_roles(fields, ids, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_model_set"></a>}}
\if{latex}{\out{\hypertarget{method-create_model_set}{}}}
\subsection{Method \code{create_model_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RoleApi$create_model_set(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_permission_set"></a>}}
\if{latex}{\out{\hypertarget{method-create_permission_set}{}}}
\subsection{Method \code{create_permission_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RoleApi$create_permission_set(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_role"></a>}}
\if{latex}{\out{\hypertarget{method-create_role}{}}}
\subsection{Method \code{create_role()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RoleApi$create_role(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_model_set"></a>}}
\if{latex}{\out{\hypertarget{method-delete_model_set}{}}}
\subsection{Method \code{delete_model_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RoleApi$delete_model_set(model_set_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_permission_set"></a>}}
\if{latex}{\out{\hypertarget{method-delete_permission_set}{}}}
\subsection{Method \code{delete_permission_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RoleApi$delete_permission_set(permission_set_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_role"></a>}}
\if{latex}{\out{\hypertarget{method-delete_role}{}}}
\subsection{Method \code{delete_role()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RoleApi$delete_role(role_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model_set"></a>}}
\if{latex}{\out{\hypertarget{method-model_set}{}}}
\subsection{Method \code{model_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RoleApi$model_set(model_set_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-permission_set"></a>}}
\if{latex}{\out{\hypertarget{method-permission_set}{}}}
\subsection{Method \code{permission_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RoleApi$permission_set(permission_set_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-role"></a>}}
\if{latex}{\out{\hypertarget{method-role}{}}}
\subsection{Method \code{role()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RoleApi$role(role_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-role_groups"></a>}}
\if{latex}{\out{\hypertarget{method-role_groups}{}}}
\subsection{Method \code{role_groups()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RoleApi$role_groups(role_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-role_users"></a>}}
\if{latex}{\out{\hypertarget{method-role_users}{}}}
\subsection{Method \code{role_users()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RoleApi$role_users(role_id, fields, direct_association_only, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_role_groups"></a>}}
\if{latex}{\out{\hypertarget{method-set_role_groups}{}}}
\subsection{Method \code{set_role_groups()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RoleApi$set_role_groups(role_id, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_role_users"></a>}}
\if{latex}{\out{\hypertarget{method-set_role_users}{}}}
\subsection{Method \code{set_role_users()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RoleApi$set_role_users(role_id, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_model_set"></a>}}
\if{latex}{\out{\hypertarget{method-update_model_set}{}}}
\subsection{Method \code{update_model_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RoleApi$update_model_set(model_set_id, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_permission_set"></a>}}
\if{latex}{\out{\hypertarget{method-update_permission_set}{}}}
\subsection{Method \code{update_permission_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RoleApi$update_permission_set(permission_set_id, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_role"></a>}}
\if{latex}{\out{\hypertarget{method-update_role}{}}}
\subsection{Method \code{update_role()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RoleApi$update_role(role_id, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RoleApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
