% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunningQueries.r
\docType{data}
\name{RunningQueries}
\alias{RunningQueries}
\title{RunningQueries Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RunningQueries
}
\description{
RunningQueries Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{user}}{}

\item{\code{query}}{}

\item{\code{sql_query}}{}

\item{\code{look}}{}

\item{\code{created_at}}{}

\item{\code{completed_at}}{}

\item{\code{query_id}}{}

\item{\code{source}}{}

\item{\code{node_id}}{}

\item{\code{slug}}{}

\item{\code{query_task_id}}{}

\item{\code{cache_key}}{}

\item{\code{connection_name}}{}

\item{\code{dialect}}{}

\item{\code{connection_id}}{}

\item{\code{message}}{}

\item{\code{status}}{}

\item{\code{runtime}}{}

\item{\code{sql}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
