% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunningQueriesApi.r
\docType{data}
\name{RunningQueriesApi}
\alias{RunningQueriesApi}
\title{RunningQueries operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RunningQueriesApi
}
\description{
looker.RunningQueries
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

all_running_queries Get All Running Queries 


kill_query Kill Running Query 

}
}

\keyword{datasets}
