% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamlUserAttributeRead.r
\docType{data}
\name{SamlUserAttributeRead}
\alias{SamlUserAttributeRead}
\title{SamlUserAttributeRead Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SamlUserAttributeRead
}
\description{
SamlUserAttributeRead Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{required}}{}

\item{\code{user_attributes}}{}

\item{\code{url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
