% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamlUserAttributeWrite.r
\name{SamlUserAttributeWrite}
\alias{SamlUserAttributeWrite}
\title{SamlUserAttributeWrite Class}
\description{
SamlUserAttributeWrite Class

SamlUserAttributeWrite Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{}

\item{\code{required}}{}

\item{\code{user_attribute_ids}}{}

\item{\code{url}}{}

\item{\code{can}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SamlUserAttributeWrite$new()}}
\item \href{#method-toJSON}{\code{SamlUserAttributeWrite$toJSON()}}
\item \href{#method-fromJSON}{\code{SamlUserAttributeWrite$fromJSON()}}
\item \href{#method-toJSONString}{\code{SamlUserAttributeWrite$toJSONString()}}
\item \href{#method-fromJSONString}{\code{SamlUserAttributeWrite$fromJSONString()}}
\item \href{#method-clone}{\code{SamlUserAttributeWrite$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SamlUserAttributeWrite$new(name, required, user_attribute_ids, url, can)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SamlUserAttributeWrite$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SamlUserAttributeWrite$fromJSON(SamlUserAttributeWriteJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SamlUserAttributeWrite$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SamlUserAttributeWrite$fromJSONString(SamlUserAttributeWriteJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SamlUserAttributeWrite$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
