% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScheduledPlan.r
\name{ScheduledPlan}
\alias{ScheduledPlan}
\title{ScheduledPlan Class}
\description{
ScheduledPlan Class

ScheduledPlan Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{}

\item{\code{name}}{}

\item{\code{created_at}}{}

\item{\code{updated_at}}{}

\item{\code{title}}{}

\item{\code{user_id}}{}

\item{\code{user}}{}

\item{\code{run_as_recipient}}{}

\item{\code{enabled}}{}

\item{\code{next_run_at}}{}

\item{\code{last_run_at}}{}

\item{\code{look_id}}{}

\item{\code{dashboard_id}}{}

\item{\code{lookml_dashboard_id}}{}

\item{\code{filters_string}}{}

\item{\code{dashboard_filters}}{}

\item{\code{require_results}}{}

\item{\code{require_no_results}}{}

\item{\code{require_change}}{}

\item{\code{send_all_results}}{}

\item{\code{crontab}}{}

\item{\code{datagroup}}{}

\item{\code{timezone}}{}

\item{\code{query_id}}{}

\item{\code{scheduled_plan_destination}}{}

\item{\code{run_once}}{}

\item{\code{include_links}}{}

\item{\code{can}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ScheduledPlan$new()}}
\item \href{#method-toJSON}{\code{ScheduledPlan$toJSON()}}
\item \href{#method-fromJSON}{\code{ScheduledPlan$fromJSON()}}
\item \href{#method-toJSONString}{\code{ScheduledPlan$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ScheduledPlan$fromJSONString()}}
\item \href{#method-clone}{\code{ScheduledPlan$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledPlan$new(
  id,
  name,
  created_at,
  updated_at,
  title,
  user_id,
  user,
  run_as_recipient,
  enabled,
  next_run_at,
  last_run_at,
  look_id,
  dashboard_id,
  lookml_dashboard_id,
  filters_string,
  dashboard_filters,
  require_results,
  require_no_results,
  require_change,
  send_all_results,
  crontab,
  datagroup,
  timezone,
  query_id,
  scheduled_plan_destination,
  run_once,
  include_links,
  can
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledPlan$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledPlan$fromJSON(ScheduledPlanJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledPlan$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledPlan$fromJSONString(ScheduledPlanJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledPlan$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
