% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScheduledPlanApi.r
\name{ScheduledPlanApi}
\alias{ScheduledPlanApi}
\title{ScheduledPlan operations}
\description{
looker.ScheduledPlan
}
\section{Methods}{

\describe{

all_scheduled_plans Get All Scheduled Plans 


create_scheduled_plan Create Scheduled Plan 


delete_scheduled_plan Delete Scheduled Plan 


scheduled_plan Get Scheduled Plan 


scheduled_plan_run_once Run Scheduled Plan Once 


scheduled_plans_for_dashboard Scheduled Plans for Dashboard 


scheduled_plans_for_look Scheduled Plans for Look 


scheduled_plans_for_lookml_dashboard Scheduled Plans for LookML Dashboard 


scheduled_plans_for_space Scheduled Plans for Space 


update_scheduled_plan Update Scheduled Plan 

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ScheduledPlanApi$new()}}
\item \href{#method-all_scheduled_plans}{\code{ScheduledPlanApi$all_scheduled_plans()}}
\item \href{#method-create_scheduled_plan}{\code{ScheduledPlanApi$create_scheduled_plan()}}
\item \href{#method-delete_scheduled_plan}{\code{ScheduledPlanApi$delete_scheduled_plan()}}
\item \href{#method-scheduled_plan}{\code{ScheduledPlanApi$scheduled_plan()}}
\item \href{#method-scheduled_plan_run_once}{\code{ScheduledPlanApi$scheduled_plan_run_once()}}
\item \href{#method-scheduled_plans_for_dashboard}{\code{ScheduledPlanApi$scheduled_plans_for_dashboard()}}
\item \href{#method-scheduled_plans_for_look}{\code{ScheduledPlanApi$scheduled_plans_for_look()}}
\item \href{#method-scheduled_plans_for_lookml_dashboard}{\code{ScheduledPlanApi$scheduled_plans_for_lookml_dashboard()}}
\item \href{#method-scheduled_plans_for_space}{\code{ScheduledPlanApi$scheduled_plans_for_space()}}
\item \href{#method-update_scheduled_plan}{\code{ScheduledPlanApi$update_scheduled_plan()}}
\item \href{#method-clone}{\code{ScheduledPlanApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledPlanApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_scheduled_plans"></a>}}
\if{latex}{\out{\hypertarget{method-all_scheduled_plans}{}}}
\subsection{Method \code{all_scheduled_plans()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledPlanApi$all_scheduled_plans(user_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_scheduled_plan"></a>}}
\if{latex}{\out{\hypertarget{method-create_scheduled_plan}{}}}
\subsection{Method \code{create_scheduled_plan()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledPlanApi$create_scheduled_plan(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_scheduled_plan"></a>}}
\if{latex}{\out{\hypertarget{method-delete_scheduled_plan}{}}}
\subsection{Method \code{delete_scheduled_plan()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledPlanApi$delete_scheduled_plan(scheduled_plan_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scheduled_plan"></a>}}
\if{latex}{\out{\hypertarget{method-scheduled_plan}{}}}
\subsection{Method \code{scheduled_plan()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledPlanApi$scheduled_plan(scheduled_plan_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scheduled_plan_run_once"></a>}}
\if{latex}{\out{\hypertarget{method-scheduled_plan_run_once}{}}}
\subsection{Method \code{scheduled_plan_run_once()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledPlanApi$scheduled_plan_run_once(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scheduled_plans_for_dashboard"></a>}}
\if{latex}{\out{\hypertarget{method-scheduled_plans_for_dashboard}{}}}
\subsection{Method \code{scheduled_plans_for_dashboard()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledPlanApi$scheduled_plans_for_dashboard(
  dashboard_id,
  user_id,
  fields,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scheduled_plans_for_look"></a>}}
\if{latex}{\out{\hypertarget{method-scheduled_plans_for_look}{}}}
\subsection{Method \code{scheduled_plans_for_look()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledPlanApi$scheduled_plans_for_look(look_id, user_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scheduled_plans_for_lookml_dashboard"></a>}}
\if{latex}{\out{\hypertarget{method-scheduled_plans_for_lookml_dashboard}{}}}
\subsection{Method \code{scheduled_plans_for_lookml_dashboard()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledPlanApi$scheduled_plans_for_lookml_dashboard(
  lookml_dashboard_id,
  user_id,
  fields,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scheduled_plans_for_space"></a>}}
\if{latex}{\out{\hypertarget{method-scheduled_plans_for_space}{}}}
\subsection{Method \code{scheduled_plans_for_space()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledPlanApi$scheduled_plans_for_space(space_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_scheduled_plan"></a>}}
\if{latex}{\out{\hypertarget{method-update_scheduled_plan}{}}}
\subsection{Method \code{update_scheduled_plan()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledPlanApi$update_scheduled_plan(scheduled_plan_id, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledPlanApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
