% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScheduledPlanDestination.r
\name{ScheduledPlanDestination}
\alias{ScheduledPlanDestination}
\title{ScheduledPlanDestination Class}
\description{
ScheduledPlanDestination Class

ScheduledPlanDestination Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{}

\item{\code{scheduled_plan_id}}{}

\item{\code{format}}{}

\item{\code{apply_formatting}}{}

\item{\code{apply_vis}}{}

\item{\code{address}}{}

\item{\code{looker_recipient}}{}

\item{\code{type}}{}

\item{\code{parameters}}{}

\item{\code{secret_parameters}}{}

\item{\code{message}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ScheduledPlanDestination$new()}}
\item \href{#method-toJSON}{\code{ScheduledPlanDestination$toJSON()}}
\item \href{#method-fromJSONObject}{\code{ScheduledPlanDestination$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{ScheduledPlanDestination$fromJSON()}}
\item \href{#method-toJSONString}{\code{ScheduledPlanDestination$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ScheduledPlanDestination$fromJSONString()}}
\item \href{#method-clone}{\code{ScheduledPlanDestination$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledPlanDestination$new(
  id,
  scheduled_plan_id,
  format,
  apply_formatting,
  apply_vis,
  address,
  looker_recipient,
  type,
  parameters,
  secret_parameters,
  message
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledPlanDestination$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledPlanDestination$fromJSONObject(ScheduledPlanDestinationJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledPlanDestination$fromJSON(ScheduledPlanDestinationJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledPlanDestination$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledPlanDestination$fromJSONString(ScheduledPlanDestinationJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledPlanDestination$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
