% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Session.r
\docType{data}
\name{Session}
\alias{Session}
\title{Session Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Session
}
\description{
Session Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{ip_address}}{}

\item{\code{browser}}{}

\item{\code{operating_system}}{}

\item{\code{city}}{}

\item{\code{state}}{}

\item{\code{country}}{}

\item{\code{credentials_type}}{}

\item{\code{extended_at}}{}

\item{\code{extended_count}}{}

\item{\code{sudo_user_id}}{}

\item{\code{created_at}}{}

\item{\code{expires_at}}{}

\item{\code{url}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
