% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SessionApi.r
\docType{data}
\name{SessionApi}
\alias{SessionApi}
\title{Session operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SessionApi
}
\description{
looker.Session
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

session Get Session 


update_session Update Session 

}
}

\keyword{datasets}
