% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Space.r
\name{Space}
\alias{Space}
\title{Space Class}
\description{
Space Class

Space Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{}

\item{\code{content_metadata_id}}{}

\item{\code{creator_id}}{}

\item{\code{name}}{}

\item{\code{is_personal}}{}

\item{\code{is_personal_descendant}}{}

\item{\code{is_shared_root}}{}

\item{\code{is_root}}{}

\item{\code{is_users_root}}{}

\item{\code{is_user_root}}{}

\item{\code{is_embed}}{}

\item{\code{is_embed_shared_root}}{}

\item{\code{is_embed_users_root}}{}

\item{\code{external_id}}{}

\item{\code{parent_id}}{}

\item{\code{looks}}{}

\item{\code{dashboards}}{}

\item{\code{can}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Space$new()}}
\item \href{#method-toJSON}{\code{Space$toJSON()}}
\item \href{#method-fromJSON}{\code{Space$fromJSON()}}
\item \href{#method-toJSONString}{\code{Space$toJSONString()}}
\item \href{#method-fromJSONString}{\code{Space$fromJSONString()}}
\item \href{#method-clone}{\code{Space$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Space$new(
  id,
  content_metadata_id,
  creator_id,
  name,
  is_personal,
  is_personal_descendant,
  is_shared_root,
  is_root,
  is_users_root,
  is_user_root,
  is_embed,
  is_embed_shared_root,
  is_embed_users_root,
  external_id,
  parent_id,
  looks,
  dashboards,
  can
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Space$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Space$fromJSON(SpaceJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Space$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Space$fromJSONString(SpaceJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Space$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
