% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceApi.r
\name{SpaceApi}
\alias{SpaceApi}
\title{Space operations}
\description{
looker.Space
}
\section{Methods}{

\describe{

all_spaces Get All Spaces 


create_space Create Space 


delete_space Delete Space 


search_spaces Search Spaces 


space Get Space 


space_ancestors Get Space Ancestors 


space_children Get Space Children 


space_children_search Search Space Children 


space_dashboards Get Space Dashboards 


space_looks Get Space Looks 


space_parent Get Space Parent 


update_space Update Space 

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SpaceApi$new()}}
\item \href{#method-all_spaces}{\code{SpaceApi$all_spaces()}}
\item \href{#method-create_space}{\code{SpaceApi$create_space()}}
\item \href{#method-delete_space}{\code{SpaceApi$delete_space()}}
\item \href{#method-search_spaces}{\code{SpaceApi$search_spaces()}}
\item \href{#method-space}{\code{SpaceApi$space()}}
\item \href{#method-space_ancestors}{\code{SpaceApi$space_ancestors()}}
\item \href{#method-space_children}{\code{SpaceApi$space_children()}}
\item \href{#method-space_children_search}{\code{SpaceApi$space_children_search()}}
\item \href{#method-space_dashboards}{\code{SpaceApi$space_dashboards()}}
\item \href{#method-space_looks}{\code{SpaceApi$space_looks()}}
\item \href{#method-space_parent}{\code{SpaceApi$space_parent()}}
\item \href{#method-update_space}{\code{SpaceApi$update_space()}}
\item \href{#method-clone}{\code{SpaceApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpaceApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_spaces"></a>}}
\if{latex}{\out{\hypertarget{method-all_spaces}{}}}
\subsection{Method \code{all_spaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpaceApi$all_spaces(fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_space"></a>}}
\if{latex}{\out{\hypertarget{method-create_space}{}}}
\subsection{Method \code{create_space()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpaceApi$create_space(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_space"></a>}}
\if{latex}{\out{\hypertarget{method-delete_space}{}}}
\subsection{Method \code{delete_space()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpaceApi$delete_space(space_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_spaces"></a>}}
\if{latex}{\out{\hypertarget{method-search_spaces}{}}}
\subsection{Method \code{search_spaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpaceApi$search_spaces(
  fields,
  page,
  per_page,
  limit,
  offset,
  sorts,
  name,
  id,
  parent_id,
  creator_id,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-space"></a>}}
\if{latex}{\out{\hypertarget{method-space}{}}}
\subsection{Method \code{space()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpaceApi$space(space_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-space_ancestors"></a>}}
\if{latex}{\out{\hypertarget{method-space_ancestors}{}}}
\subsection{Method \code{space_ancestors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpaceApi$space_ancestors(space_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-space_children"></a>}}
\if{latex}{\out{\hypertarget{method-space_children}{}}}
\subsection{Method \code{space_children()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpaceApi$space_children(space_id, fields, page, per_page, sorts, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-space_children_search"></a>}}
\if{latex}{\out{\hypertarget{method-space_children_search}{}}}
\subsection{Method \code{space_children_search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpaceApi$space_children_search(space_id, fields, sorts, name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-space_dashboards"></a>}}
\if{latex}{\out{\hypertarget{method-space_dashboards}{}}}
\subsection{Method \code{space_dashboards()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpaceApi$space_dashboards(space_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-space_looks"></a>}}
\if{latex}{\out{\hypertarget{method-space_looks}{}}}
\subsection{Method \code{space_looks()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpaceApi$space_looks(space_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-space_parent"></a>}}
\if{latex}{\out{\hypertarget{method-space_parent}{}}}
\subsection{Method \code{space_parent()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpaceApi$space_parent(space_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_space"></a>}}
\if{latex}{\out{\hypertarget{method-update_space}{}}}
\subsection{Method \code{update_space()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpaceApi$update_space(space_id, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpaceApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
