% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SqlQuery.r
\docType{data}
\name{SqlQuery}
\alias{SqlQuery}
\title{SqlQuery Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SqlQuery
}
\description{
SqlQuery Class
}
\section{Fields}{

\describe{
\item{\code{slug}}{}

\item{\code{last_runtime}}{}

\item{\code{run_count}}{}

\item{\code{browser_limit}}{}

\item{\code{sql}}{}

\item{\code{last_run_at}}{}

\item{\code{connection}}{}

\item{\code{creator}}{}

\item{\code{explore_url}}{}

\item{\code{plaintext}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
