% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SqlQueryApi.r
\name{SqlQueryApi}
\alias{SqlQueryApi}
\title{SqlQuery operations}
\description{
looker.SqlQuery
}
\section{Methods}{

\describe{

create_sql_query Create SQL Runner Query 


sql_query Get SQL Runner Query 

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SqlQueryApi$new()}}
\item \href{#method-create_sql_query}{\code{SqlQueryApi$create_sql_query()}}
\item \href{#method-sql_query}{\code{SqlQueryApi$sql_query()}}
\item \href{#method-clone}{\code{SqlQueryApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SqlQueryApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_sql_query"></a>}}
\if{latex}{\out{\hypertarget{method-create_sql_query}{}}}
\subsection{Method \code{create_sql_query()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SqlQueryApi$create_sql_query(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sql_query"></a>}}
\if{latex}{\out{\hypertarget{method-sql_query}{}}}
\subsection{Method \code{sql_query()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SqlQueryApi$sql_query(slug, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SqlQueryApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
