% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{TimeSlice}
\alias{TimeSlice}
\title{Extract an interval of time in a Trial}
\usage{
TimeSlice(x, from = lwl_opts$get("timeslice_start"),
  to = lwl_opts$get("timeslice_end"))
}
\arguments{
\item{x}{a Trial or a TrialList object}

\item{from}{the time at which to start slicing. This value can be a
\code{character} string naming a valid timing attribute of the Trial (e.g.,
\code{"TargetOnset"}), a numeric value specifying a particular time, or
\code{NULL} in which case the function slices from the first frame of time.}

\item{to}{time to which to finish slicing. The parameter may be of the same
classes as described above for \code{from}, but when \code{NULL} is passed
a value, the final time frame is used for slicing.}
}
\value{
the Time column of the Trial(s) is updated so that minimum time is
  within one frame of \code{from} and the maximum time is within one frame of
  {to}. The updated Trial(s) also has the added attribute
  \code{"NumberOfFrames"}.
}
\description{
Extract an interval of time in a Trial
}

