% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Timezone.r
\docType{data}
\name{Timezone}
\alias{Timezone}
\title{Timezone Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Timezone
}
\description{
Timezone Class
}
\section{Fields}{

\describe{
\item{\code{value}}{}

\item{\code{label}}{}

\item{\code{group}}{}
}}

\keyword{datasets}
