% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/task.R
\name{TryLoad}
\alias{TryLoad}
\title{Try to load an experimental block, failing gracefully if unsuccessful}
\usage{
TryLoad(subject_path, handler = warning)
}
\arguments{
\item{subject_path}{a path to a subject's directory which may contain blocks
from LWL tasks.}

\item{handler}{a function to apply to the file-path of an experimental block
hat raised an error. The default is to \code{print} the file-path.}
}
\value{
\code{NULL} if there are no loadable experimental blocks in
  \code{subject_path}, otherwise a \code{Session} containing the blocks that
  could be loaded from \code{subject_path}.
}
\description{
\code{TryLoad(...)} tries to load the blocks in a subject directory. If it
encounters an error, it catches the error and moves on to the next in the
subject directory.
}

