% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User.r
\name{User}
\alias{User}
\title{User Class}
\description{
User Class

User Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{can}}{}

\item{\code{avatar_url}}{}

\item{\code{avatar_url_without_sizing}}{}

\item{\code{credentials_api3}}{}

\item{\code{credentials_email}}{}

\item{\code{credentials_embed}}{}

\item{\code{credentials_google}}{}

\item{\code{credentials_ldap}}{}

\item{\code{credentials_looker_openid}}{}

\item{\code{credentials_oidc}}{}

\item{\code{credentials_saml}}{}

\item{\code{credentials_totp}}{}

\item{\code{display_name}}{}

\item{\code{email}}{}

\item{\code{embed_group_space_id}}{}

\item{\code{first_name}}{}

\item{\code{group_ids}}{}

\item{\code{home_folder_id}}{}

\item{\code{id}}{}

\item{\code{is_disabled}}{}

\item{\code{last_name}}{}

\item{\code{locale}}{}

\item{\code{looker_versions}}{}

\item{\code{models_dir_validated}}{}

\item{\code{personal_folder_id}}{}

\item{\code{presumed_looker_employee}}{}

\item{\code{role_ids}}{}

\item{\code{sessions}}{}

\item{\code{ui_state}}{}

\item{\code{verified_looker_employee}}{}

\item{\code{roles_externally_managed}}{}

\item{\code{allow_direct_roles}}{}

\item{\code{allow_normal_group_membership}}{}

\item{\code{allow_roles_from_normal_groups}}{}

\item{\code{url}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{User$new()}}
\item \href{#method-toJSON}{\code{User$toJSON()}}
\item \href{#method-fromJSONObject}{\code{User$fromJSONObject()}}
\item \href{#method-fromJSON}{\code{User$fromJSON()}}
\item \href{#method-toJSONString}{\code{User$toJSONString()}}
\item \href{#method-fromJSONString}{\code{User$fromJSONString()}}
\item \href{#method-clone}{\code{User$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$new(
  can,
  avatar_url,
  avatar_url_without_sizing,
  credentials_api3,
  credentials_email,
  credentials_embed,
  credentials_google,
  credentials_ldap,
  credentials_looker_openid,
  credentials_oidc,
  credentials_saml,
  credentials_totp,
  display_name,
  email,
  embed_group_space_id,
  first_name,
  group_ids,
  home_folder_id,
  id,
  is_disabled,
  last_name,
  locale,
  looker_versions,
  models_dir_validated,
  personal_folder_id,
  presumed_looker_employee,
  role_ids,
  sessions,
  ui_state,
  verified_looker_employee,
  roles_externally_managed,
  allow_direct_roles,
  allow_normal_group_membership,
  allow_roles_from_normal_groups,
  url
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONObject}{}}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$fromJSONObject(UserJsonObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$fromJSON(UserJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$fromJSONString(UserJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
