% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User.r
\name{User}
\alias{User}
\title{User Class}
\description{
User Class

User Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{}

\item{\code{first_name}}{}

\item{\code{last_name}}{}

\item{\code{display_name}}{}

\item{\code{email}}{}

\item{\code{is_disabled}}{}

\item{\code{avatar_url}}{}

\item{\code{home_space_id}}{}

\item{\code{personal_space_id}}{}

\item{\code{embed_group_space_id}}{}

\item{\code{access_filters}}{}

\item{\code{credentials_email}}{}

\item{\code{credentials_totp}}{}

\item{\code{credentials_ldap}}{}

\item{\code{credentials_google}}{}

\item{\code{credentials_saml}}{}

\item{\code{credentials_oidc}}{}

\item{\code{credentials_api}}{}

\item{\code{credentials_api3}}{}

\item{\code{credentials_embed}}{}

\item{\code{credentials_looker_openid}}{}

\item{\code{sessions}}{}

\item{\code{role_ids}}{}

\item{\code{group_ids}}{}

\item{\code{presumed_looker_employee}}{}

\item{\code{verified_looker_employee}}{}

\item{\code{looker_versions}}{}

\item{\code{ui_state}}{}

\item{\code{url}}{}

\item{\code{can}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{User$new()}}
\item \href{#method-toJSON}{\code{User$toJSON()}}
\item \href{#method-fromJSON}{\code{User$fromJSON()}}
\item \href{#method-toJSONString}{\code{User$toJSONString()}}
\item \href{#method-fromJSONString}{\code{User$fromJSONString()}}
\item \href{#method-clone}{\code{User$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$new(
  id,
  first_name,
  last_name,
  display_name,
  email,
  is_disabled,
  avatar_url,
  home_space_id,
  personal_space_id,
  embed_group_space_id,
  access_filters,
  credentials_email,
  credentials_totp,
  credentials_ldap,
  credentials_google,
  credentials_saml,
  credentials_oidc,
  credentials_api,
  credentials_api3,
  credentials_embed,
  credentials_looker_openid,
  sessions,
  role_ids,
  group_ids,
  presumed_looker_employee,
  verified_looker_employee,
  looker_versions,
  ui_state,
  url,
  can
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$fromJSON(UserJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$fromJSONString(UserJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
