% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserApi.r
\name{UserApi}
\alias{UserApi}
\title{User operations}
\description{
looker.User
}
\section{Methods}{

\describe{

all_user_credentials_api3s Get All API 3 Credentials


all_user_credentials_embeds Get All Embedding Credentials


all_user_sessions Get All Web Login Sessions


all_users Get All Users


create_embed_user Create an embed user from an external user ID


create_user Create User


create_user_credentials_api3 Create API 3 Credential


create_user_credentials_email Create Email/Password Credential


create_user_credentials_email_password_reset Create Password Reset Token


create_user_credentials_totp Create Two-Factor Credential


delete_user Delete User


delete_user_attribute_user_value Delete User Attribute User Value


delete_user_credentials_api3 Delete API 3 Credential


delete_user_credentials_email Delete Email/Password Credential


delete_user_credentials_embed Delete Embedding Credential


delete_user_credentials_google Delete Google Auth Credential


delete_user_credentials_ldap Delete LDAP Credential


delete_user_credentials_looker_openid Delete Looker OpenId Credential


delete_user_credentials_oidc Delete OIDC Auth Credential


delete_user_credentials_saml Delete Saml Auth Credential


delete_user_credentials_totp Delete Two-Factor Credential


delete_user_session Delete Web Login Session


me Get Current User


search_credentials_email Search CredentialsEmail


search_users Search Users


search_users_names Search User Names


send_user_credentials_email_password_reset Send Password Reset Token


set_user_attribute_user_value Set User Attribute User Value


set_user_roles Set User Roles


update_user Update User


update_user_credentials_email Update Email/Password Credential


user Get User by Id


user_attribute_user_values Get User Attribute Values


user_credentials_api3 Get API 3 Credential


user_credentials_email Get Email/Password Credential


user_credentials_embed Get Embedding Credential


user_credentials_google Get Google Auth Credential


user_credentials_ldap Get LDAP Credential


user_credentials_looker_openid Get Looker OpenId Credential


user_credentials_oidc Get OIDC Auth Credential


user_credentials_saml Get Saml Auth Credential


user_credentials_totp Get Two-Factor Credential


user_for_credential Get User by Credential Id


user_roles Get User Roles


user_session Get Web Login Session


wipeout_user_emails Wipeout User Emails

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UserApi$new()}}
\item \href{#method-all_user_credentials_api3s}{\code{UserApi$all_user_credentials_api3s()}}
\item \href{#method-all_user_credentials_embeds}{\code{UserApi$all_user_credentials_embeds()}}
\item \href{#method-all_user_sessions}{\code{UserApi$all_user_sessions()}}
\item \href{#method-all_users}{\code{UserApi$all_users()}}
\item \href{#method-create_embed_user}{\code{UserApi$create_embed_user()}}
\item \href{#method-create_user}{\code{UserApi$create_user()}}
\item \href{#method-create_user_credentials_api3}{\code{UserApi$create_user_credentials_api3()}}
\item \href{#method-create_user_credentials_email}{\code{UserApi$create_user_credentials_email()}}
\item \href{#method-create_user_credentials_email_password_reset}{\code{UserApi$create_user_credentials_email_password_reset()}}
\item \href{#method-create_user_credentials_totp}{\code{UserApi$create_user_credentials_totp()}}
\item \href{#method-delete_user}{\code{UserApi$delete_user()}}
\item \href{#method-delete_user_attribute_user_value}{\code{UserApi$delete_user_attribute_user_value()}}
\item \href{#method-delete_user_credentials_api3}{\code{UserApi$delete_user_credentials_api3()}}
\item \href{#method-delete_user_credentials_email}{\code{UserApi$delete_user_credentials_email()}}
\item \href{#method-delete_user_credentials_embed}{\code{UserApi$delete_user_credentials_embed()}}
\item \href{#method-delete_user_credentials_google}{\code{UserApi$delete_user_credentials_google()}}
\item \href{#method-delete_user_credentials_ldap}{\code{UserApi$delete_user_credentials_ldap()}}
\item \href{#method-delete_user_credentials_looker_openid}{\code{UserApi$delete_user_credentials_looker_openid()}}
\item \href{#method-delete_user_credentials_oidc}{\code{UserApi$delete_user_credentials_oidc()}}
\item \href{#method-delete_user_credentials_saml}{\code{UserApi$delete_user_credentials_saml()}}
\item \href{#method-delete_user_credentials_totp}{\code{UserApi$delete_user_credentials_totp()}}
\item \href{#method-delete_user_session}{\code{UserApi$delete_user_session()}}
\item \href{#method-me}{\code{UserApi$me()}}
\item \href{#method-search_credentials_email}{\code{UserApi$search_credentials_email()}}
\item \href{#method-search_users}{\code{UserApi$search_users()}}
\item \href{#method-search_users_names}{\code{UserApi$search_users_names()}}
\item \href{#method-send_user_credentials_email_password_reset}{\code{UserApi$send_user_credentials_email_password_reset()}}
\item \href{#method-set_user_attribute_user_value}{\code{UserApi$set_user_attribute_user_value()}}
\item \href{#method-set_user_roles}{\code{UserApi$set_user_roles()}}
\item \href{#method-update_user}{\code{UserApi$update_user()}}
\item \href{#method-update_user_credentials_email}{\code{UserApi$update_user_credentials_email()}}
\item \href{#method-user}{\code{UserApi$user()}}
\item \href{#method-user_attribute_user_values}{\code{UserApi$user_attribute_user_values()}}
\item \href{#method-user_credentials_api3}{\code{UserApi$user_credentials_api3()}}
\item \href{#method-user_credentials_email}{\code{UserApi$user_credentials_email()}}
\item \href{#method-user_credentials_embed}{\code{UserApi$user_credentials_embed()}}
\item \href{#method-user_credentials_google}{\code{UserApi$user_credentials_google()}}
\item \href{#method-user_credentials_ldap}{\code{UserApi$user_credentials_ldap()}}
\item \href{#method-user_credentials_looker_openid}{\code{UserApi$user_credentials_looker_openid()}}
\item \href{#method-user_credentials_oidc}{\code{UserApi$user_credentials_oidc()}}
\item \href{#method-user_credentials_saml}{\code{UserApi$user_credentials_saml()}}
\item \href{#method-user_credentials_totp}{\code{UserApi$user_credentials_totp()}}
\item \href{#method-user_for_credential}{\code{UserApi$user_for_credential()}}
\item \href{#method-user_roles}{\code{UserApi$user_roles()}}
\item \href{#method-user_session}{\code{UserApi$user_session()}}
\item \href{#method-wipeout_user_emails}{\code{UserApi$wipeout_user_emails()}}
\item \href{#method-clone}{\code{UserApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_user_credentials_api3s"></a>}}
\if{latex}{\out{\hypertarget{method-all_user_credentials_api3s}{}}}
\subsection{Method \code{all_user_credentials_api3s()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$all_user_credentials_api3s(user_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_user_credentials_embeds"></a>}}
\if{latex}{\out{\hypertarget{method-all_user_credentials_embeds}{}}}
\subsection{Method \code{all_user_credentials_embeds()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$all_user_credentials_embeds(user_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_user_sessions"></a>}}
\if{latex}{\out{\hypertarget{method-all_user_sessions}{}}}
\subsection{Method \code{all_user_sessions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$all_user_sessions(user_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_users"></a>}}
\if{latex}{\out{\hypertarget{method-all_users}{}}}
\subsection{Method \code{all_users()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$all_users(fields, page, per_page, limit, offset, sorts, ids, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_embed_user"></a>}}
\if{latex}{\out{\hypertarget{method-create_embed_user}{}}}
\subsection{Method \code{create_embed_user()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$create_embed_user(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_user"></a>}}
\if{latex}{\out{\hypertarget{method-create_user}{}}}
\subsection{Method \code{create_user()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$create_user(body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_user_credentials_api3"></a>}}
\if{latex}{\out{\hypertarget{method-create_user_credentials_api3}{}}}
\subsection{Method \code{create_user_credentials_api3()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$create_user_credentials_api3(user_id, body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_user_credentials_email"></a>}}
\if{latex}{\out{\hypertarget{method-create_user_credentials_email}{}}}
\subsection{Method \code{create_user_credentials_email()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$create_user_credentials_email(user_id, body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_user_credentials_email_password_reset"></a>}}
\if{latex}{\out{\hypertarget{method-create_user_credentials_email_password_reset}{}}}
\subsection{Method \code{create_user_credentials_email_password_reset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$create_user_credentials_email_password_reset(
  user_id,
  expires,
  fields,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_user_credentials_totp"></a>}}
\if{latex}{\out{\hypertarget{method-create_user_credentials_totp}{}}}
\subsection{Method \code{create_user_credentials_totp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$create_user_credentials_totp(user_id, body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_user"></a>}}
\if{latex}{\out{\hypertarget{method-delete_user}{}}}
\subsection{Method \code{delete_user()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$delete_user(user_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_user_attribute_user_value"></a>}}
\if{latex}{\out{\hypertarget{method-delete_user_attribute_user_value}{}}}
\subsection{Method \code{delete_user_attribute_user_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$delete_user_attribute_user_value(user_id, user_attribute_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_user_credentials_api3"></a>}}
\if{latex}{\out{\hypertarget{method-delete_user_credentials_api3}{}}}
\subsection{Method \code{delete_user_credentials_api3()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$delete_user_credentials_api3(user_id, credentials_api3_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_user_credentials_email"></a>}}
\if{latex}{\out{\hypertarget{method-delete_user_credentials_email}{}}}
\subsection{Method \code{delete_user_credentials_email()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$delete_user_credentials_email(user_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_user_credentials_embed"></a>}}
\if{latex}{\out{\hypertarget{method-delete_user_credentials_embed}{}}}
\subsection{Method \code{delete_user_credentials_embed()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$delete_user_credentials_embed(user_id, credentials_embed_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_user_credentials_google"></a>}}
\if{latex}{\out{\hypertarget{method-delete_user_credentials_google}{}}}
\subsection{Method \code{delete_user_credentials_google()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$delete_user_credentials_google(user_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_user_credentials_ldap"></a>}}
\if{latex}{\out{\hypertarget{method-delete_user_credentials_ldap}{}}}
\subsection{Method \code{delete_user_credentials_ldap()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$delete_user_credentials_ldap(user_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_user_credentials_looker_openid"></a>}}
\if{latex}{\out{\hypertarget{method-delete_user_credentials_looker_openid}{}}}
\subsection{Method \code{delete_user_credentials_looker_openid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$delete_user_credentials_looker_openid(user_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_user_credentials_oidc"></a>}}
\if{latex}{\out{\hypertarget{method-delete_user_credentials_oidc}{}}}
\subsection{Method \code{delete_user_credentials_oidc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$delete_user_credentials_oidc(user_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_user_credentials_saml"></a>}}
\if{latex}{\out{\hypertarget{method-delete_user_credentials_saml}{}}}
\subsection{Method \code{delete_user_credentials_saml()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$delete_user_credentials_saml(user_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_user_credentials_totp"></a>}}
\if{latex}{\out{\hypertarget{method-delete_user_credentials_totp}{}}}
\subsection{Method \code{delete_user_credentials_totp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$delete_user_credentials_totp(user_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_user_session"></a>}}
\if{latex}{\out{\hypertarget{method-delete_user_session}{}}}
\subsection{Method \code{delete_user_session()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$delete_user_session(user_id, session_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-me"></a>}}
\if{latex}{\out{\hypertarget{method-me}{}}}
\subsection{Method \code{me()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$me(fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_credentials_email"></a>}}
\if{latex}{\out{\hypertarget{method-search_credentials_email}{}}}
\subsection{Method \code{search_credentials_email()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$search_credentials_email(
  fields,
  limit,
  offset,
  sorts,
  id,
  email,
  emails,
  filter_or,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_users"></a>}}
\if{latex}{\out{\hypertarget{method-search_users}{}}}
\subsection{Method \code{search_users()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$search_users(
  fields,
  page,
  per_page,
  limit,
  offset,
  sorts,
  id,
  first_name,
  last_name,
  verified_looker_employee,
  embed_user,
  email,
  is_disabled,
  filter_or,
  content_metadata_id,
  group_id,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_users_names"></a>}}
\if{latex}{\out{\hypertarget{method-search_users_names}{}}}
\subsection{Method \code{search_users_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$search_users_names(
  pattern,
  fields,
  page,
  per_page,
  limit,
  offset,
  sorts,
  id,
  first_name,
  last_name,
  verified_looker_employee,
  email,
  is_disabled,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-send_user_credentials_email_password_reset"></a>}}
\if{latex}{\out{\hypertarget{method-send_user_credentials_email_password_reset}{}}}
\subsection{Method \code{send_user_credentials_email_password_reset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$send_user_credentials_email_password_reset(user_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_user_attribute_user_value"></a>}}
\if{latex}{\out{\hypertarget{method-set_user_attribute_user_value}{}}}
\subsection{Method \code{set_user_attribute_user_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$set_user_attribute_user_value(user_id, user_attribute_id, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_user_roles"></a>}}
\if{latex}{\out{\hypertarget{method-set_user_roles}{}}}
\subsection{Method \code{set_user_roles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$set_user_roles(user_id, body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_user"></a>}}
\if{latex}{\out{\hypertarget{method-update_user}{}}}
\subsection{Method \code{update_user()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$update_user(user_id, body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_user_credentials_email"></a>}}
\if{latex}{\out{\hypertarget{method-update_user_credentials_email}{}}}
\subsection{Method \code{update_user_credentials_email()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$update_user_credentials_email(user_id, body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-user"></a>}}
\if{latex}{\out{\hypertarget{method-user}{}}}
\subsection{Method \code{user()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$user(user_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-user_attribute_user_values"></a>}}
\if{latex}{\out{\hypertarget{method-user_attribute_user_values}{}}}
\subsection{Method \code{user_attribute_user_values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$user_attribute_user_values(
  user_id,
  fields,
  user_attribute_ids,
  all_values,
  include_unset,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-user_credentials_api3"></a>}}
\if{latex}{\out{\hypertarget{method-user_credentials_api3}{}}}
\subsection{Method \code{user_credentials_api3()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$user_credentials_api3(user_id, credentials_api3_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-user_credentials_email"></a>}}
\if{latex}{\out{\hypertarget{method-user_credentials_email}{}}}
\subsection{Method \code{user_credentials_email()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$user_credentials_email(user_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-user_credentials_embed"></a>}}
\if{latex}{\out{\hypertarget{method-user_credentials_embed}{}}}
\subsection{Method \code{user_credentials_embed()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$user_credentials_embed(user_id, credentials_embed_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-user_credentials_google"></a>}}
\if{latex}{\out{\hypertarget{method-user_credentials_google}{}}}
\subsection{Method \code{user_credentials_google()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$user_credentials_google(user_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-user_credentials_ldap"></a>}}
\if{latex}{\out{\hypertarget{method-user_credentials_ldap}{}}}
\subsection{Method \code{user_credentials_ldap()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$user_credentials_ldap(user_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-user_credentials_looker_openid"></a>}}
\if{latex}{\out{\hypertarget{method-user_credentials_looker_openid}{}}}
\subsection{Method \code{user_credentials_looker_openid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$user_credentials_looker_openid(user_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-user_credentials_oidc"></a>}}
\if{latex}{\out{\hypertarget{method-user_credentials_oidc}{}}}
\subsection{Method \code{user_credentials_oidc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$user_credentials_oidc(user_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-user_credentials_saml"></a>}}
\if{latex}{\out{\hypertarget{method-user_credentials_saml}{}}}
\subsection{Method \code{user_credentials_saml()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$user_credentials_saml(user_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-user_credentials_totp"></a>}}
\if{latex}{\out{\hypertarget{method-user_credentials_totp}{}}}
\subsection{Method \code{user_credentials_totp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$user_credentials_totp(user_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-user_for_credential"></a>}}
\if{latex}{\out{\hypertarget{method-user_for_credential}{}}}
\subsection{Method \code{user_for_credential()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$user_for_credential(credential_type, credential_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-user_roles"></a>}}
\if{latex}{\out{\hypertarget{method-user_roles}{}}}
\subsection{Method \code{user_roles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$user_roles(user_id, fields, direct_association_only, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-user_session"></a>}}
\if{latex}{\out{\hypertarget{method-user_session}{}}}
\subsection{Method \code{user_session()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$user_session(user_id, session_id, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wipeout_user_emails"></a>}}
\if{latex}{\out{\hypertarget{method-wipeout_user_emails}{}}}
\subsection{Method \code{wipeout_user_emails()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$wipeout_user_emails(user_id, body, fields, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
