% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserAttribute.r
\docType{data}
\name{UserAttribute}
\alias{UserAttribute}
\title{UserAttribute Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
UserAttribute
}
\description{
UserAttribute Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{name}}{}

\item{\code{label}}{}

\item{\code{type}}{}

\item{\code{default_value}}{}

\item{\code{is_system}}{}

\item{\code{value_is_hidden}}{}

\item{\code{user_can_view}}{}

\item{\code{user_can_edit}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
