% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserAttributeApi.r
\docType{data}
\name{UserAttributeApi}
\alias{UserAttributeApi}
\title{UserAttribute operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
UserAttributeApi
}
\description{
looker.UserAttribute
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

all_user_attribute_group_values Get User Attribute Group Values 


all_user_attributes Get All User Attributes 


create_user_attribute Create User Attribute 


delete_user_attribute Delete User Attribute 


set_user_attribute_group_values Set User Attribute Group Values 


update_user_attribute Update User Attribute 


user_attribute Get User Attribute 

}
}

\keyword{datasets}
