% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserAttributeWithValue.r
\docType{data}
\name{UserAttributeWithValue}
\alias{UserAttributeWithValue}
\title{UserAttributeWithValue Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
UserAttributeWithValue
}
\description{
UserAttributeWithValue Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{label}}{}

\item{\code{rank}}{}

\item{\code{value}}{}

\item{\code{user_id}}{}

\item{\code{user_can_edit}}{}

\item{\code{value_is_hidden}}{}

\item{\code{user_attribute_id}}{}

\item{\code{source}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
