% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserIdOnly.r
\docType{data}
\name{UserIdOnly}
\alias{UserIdOnly}
\title{UserIdOnly Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
UserIdOnly
}
\description{
UserIdOnly Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}
}}

\keyword{datasets}
