% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Whitelabel.r
\name{Whitelabel}
\alias{Whitelabel}
\title{Whitelabel Class}
\description{
Whitelabel Class

Whitelabel Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{}

\item{\code{logo_file}}{}

\item{\code{logo_url}}{}

\item{\code{favicon_file}}{}

\item{\code{favicon_url}}{}

\item{\code{default_title}}{}

\item{\code{show_help_menu}}{}

\item{\code{show_docs}}{}

\item{\code{show_email_sub_options}}{}

\item{\code{allow_looker_mentions}}{}

\item{\code{allow_looker_links}}{}

\item{\code{can}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Whitelabel$new()}}
\item \href{#method-toJSON}{\code{Whitelabel$toJSON()}}
\item \href{#method-fromJSON}{\code{Whitelabel$fromJSON()}}
\item \href{#method-toJSONString}{\code{Whitelabel$toJSONString()}}
\item \href{#method-fromJSONString}{\code{Whitelabel$fromJSONString()}}
\item \href{#method-clone}{\code{Whitelabel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Whitelabel$new(
  id,
  logo_file,
  logo_url,
  favicon_file,
  favicon_url,
  default_title,
  show_help_menu,
  show_docs,
  show_email_sub_options,
  allow_looker_mentions,
  allow_looker_links,
  can
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Whitelabel$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Whitelabel$fromJSON(WhitelabelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Whitelabel$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Whitelabel$fromJSONString(WhitelabelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Whitelabel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
