% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workspace.r
\docType{data}
\name{Workspace}
\alias{Workspace}
\title{Workspace Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Workspace
}
\description{
Workspace Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{projects}}{}

\item{\code{can}}{}
}}

\keyword{datasets}
