% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkspaceApi.r
\name{WorkspaceApi}
\alias{WorkspaceApi}
\title{Workspace operations}
\description{
looker.Workspace
}
\section{Methods}{

\describe{

all_workspaces Get All Workspaces


workspace Get Workspace

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{WorkspaceApi$new()}}
\item \href{#method-all_workspaces}{\code{WorkspaceApi$all_workspaces()}}
\item \href{#method-workspace}{\code{WorkspaceApi$workspace()}}
\item \href{#method-clone}{\code{WorkspaceApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkspaceApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_workspaces"></a>}}
\if{latex}{\out{\hypertarget{method-all_workspaces}{}}}
\subsection{Method \code{all_workspaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkspaceApi$all_workspaces(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-workspace"></a>}}
\if{latex}{\out{\hypertarget{method-workspace}{}}}
\subsection{Method \code{workspace()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkspaceApi$workspace(workspace_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkspaceApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
